% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Site.R
\name{add_aqua_Site}
\alias{add_aqua_Site}
\title{Site: Add a Aquatic Habitat Resource}
\usage{
add_aqua_Site(Aqua_Resource)
}
\arguments{
\item{Aqua_Resource}{a reference to an object deriving from \code{\link{Aqua_Resource}}}
}
\description{
Add a aquatic habitat resource to this site.
\itemize{
\item binding: \code{Site$add_aqua}
}
}
