% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Emerge.R
\name{add_egg_Emerge}
\alias{add_egg_Emerge}
\title{Aquatic Ecology: Emerge - Add Eggs}
\usage{
add_egg_Emerge(EggQ)
}
\arguments{
\item{EggQ}{egg batch}
}
\description{
Add an egg batch from a successful oviposition attempt to this \code{\link{Aqua_Resource_Emerge}}.
}
\details{
\itemize{
\item This method is bound to \code{Aqua_Resource_Emerge$add_egg()}.
}
}
