% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Site.R
\name{add_feed_Site}
\alias{add_feed_Site}
\title{Site: Add a Blood Feeding Resource}
\usage{
add_feed_Site(Feeding_Resource)
}
\arguments{
\item{Feeding_Resource}{a reference to a \code{\link[MBITES]{Feeding_Resource}} object}
}
\description{
Add a blood feeding resource to this site.
\itemize{
\item binding: \code{Site$add_feed}
}
}
