% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Site.R
\name{add_mate_Site}
\alias{add_mate_Site}
\title{Site: Add a Mating Resource}
\usage{
add_mate_Site(Mating_Resource)
}
\arguments{
\item{Mating_Resource}{a reference to a \code{\link{Mating_Resource}} object}
}
\description{
Add a mating resource to this site.
\itemize{
\item binding: \code{Site$add_mate}
}
}
