% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Globals.R
\name{add_tile_MBITES_Globals}
\alias{add_tile_MBITES_Globals}
\title{MBITES Globals: Add a Tile Reference}
\usage{
add_tile_MBITES_Globals(tile)
}
\arguments{
\item{tile}{a reference to a \code{\link[MBITES]{Tile}} object}
}
\description{
Adds a reference to a tile to the list of all tiles in the simulation.
}
\details{
\itemize{
\item This method is bound to \code{MBITES_Globals$add_tile}.
}
}
