% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-AquaPop.R
\docType{class}
\name{Aqua_Pop}
\alias{Aqua_Pop}
\title{Landscape Aquatic Population Model Class}
\format{An \code{\link{R6Class}} generator object}
\usage{
Aqua_Pop
}
\description{
A \code{Aqua_Pop} object is an abstract base class for models of aquatic population dynamics following the pimpl idiom within a
\code{\link[MBITES]{Aqua_Resource}}. This class defines an interface called from \code{\link[MBITES]{Aqua_Resource}}, such that
any aquatic population model must provide definitions of the interface functions.
}
\section{**Constructor**}{

 * none
}

\section{**Methods**}{

 * add_egg: function that must take an EggQ (a \code{list} object) and add egg batches to the population; it should return a modified EggQ with those elements corresponding to 'processed' egg batches zeroed out
 * one_day: function with void return that runs one day of the specific aquatic population simulation implementation
 * get_imago: function that returns an ImagoQ (a \code{list} object) for imagos (adult mosquitoes) ready to emerge on that day
}

\section{**Fields**}{

 * none
}

\keyword{R6}
\keyword{class}
