% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{BFAB_B2X}
\alias{BFAB_B2X}
\title{MBDETES: Blood Feeding Attempt Bout Transition Probabilities}
\usage{
BFAB_B2X(PAR)
}
\arguments{
\item{PAR}{a named list of parameters}
}
\description{
After receiving parameter values, compute state transition probabilities
from the blood feeding attempt bout state to all other states.
This function is called from \code{\link{MBDETES_BstateTransitions}}.
}
