% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Parameters.R
\name{BFAB_PAR}
\alias{BFAB_PAR}
\title{MBDETES: Parameters for Transitions from B}
\usage{
BFAB_PAR(A = 1, B0 = 0, B1 = 0.9, B2 = 0.1, C1 = 0.01,
  C2 = 0.85, C4 = 0.01, C5 = 0.85, C7 = 0.5, D1 = 0.01,
  D2 = 0.9, E = 0.99, F1 = 0.99, F2 = 0.99, G = 0.2, H1 = 0.8,
  H2 = 0.8, H3 = 0.7)
}
\arguments{
\item{A}{probability of successful launch (maps to \code{B_succeed})}

\item{B0}{probability of empty host queue (not possible to directly set)}

\item{B1}{weight on human hosts (add to sites on a tile through \code{\link{Human_NULL_Initialize}})}

\item{B2}{weight on zoo hosts (add to sites on a tile through \code{\link{Tile_Initialize}})}

\item{C1}{probability of death during human host approach (maps to 1 - \code{surviveH})}

\item{C2}{probability to not be deterred during probing a human host (maps to \code{surviveH}*\code{probeH})}

\item{C4}{probability of death during zoo host approach (maps to 1 - \code{surviveZ})}

\item{C5}{probability to successfully begin blood feeding a zoo host (maps to \code{surviveZ}*\code{feedZ})}

\item{C7}{probability of death when encountering a blood trap (pending traps in MBITES ...)}

\item{D1}{probability of death during probing a human host (maps to 1 - \code{surviveprobeH})}

\item{D2}{probability to be successfully begin blood feeding on a human (maps to \code{surviveprobeH}*\code{feedH})}

\item{E}{probability to survive post-prandial resting flight after a blood meal (calculate with \code{\link{MBDETES_PrPPRFlight}}, uses parameters \code{PPR_a}, \code{PPR_b}, \code{bm_a}, \code{bm_b})}

\item{F1}{probability to survive resting (\code{updateState}) given a successful blood meal (calculate with \code{MBDETES_PrSurvive(site,"B")}, uses parameters \code{B_surv}, and the local \code{\link{Site}} object's hazard)}

\item{F2}{probability to survive resting (\code{updateState}) given an unsuccessful blood meal attempt (calculate with \code{MBDETES_PrSurvive(site,"B")}, uses parameters \code{B_surv}, and the local \code{\link{Site}} object's hazard)}

\item{G}{probability to refeed following resting (calculate with \code{\link{MBDETES_PrRefeed}}, uses parameters \code{rf_a}, \code{rf_b}, \code{bm_a}, \code{bm_b})}

\item{H1}{probability to leave the current \code{\link{Site}} and go on a blood feeding search given a successful blood meal attempt (calculate with \code{1 - MBDETES_PrLeave(site,"feed",FALSE)}, uses parameters \code{boutFail_p}, \code{disperse})}

\item{H2}{probability to leave the current \code{\link{Site}} and go on an oviposition search given a successful blood meal attempt (calculate with \code{1 - MBDETES_PrLeave(site,"aqua",FALSE)}, uses parameters \code{boutFail_p}, \code{disperse})}

\item{H3}{probability to leave the current \code{\link{Site}} and go on a blood feeding search given an unsuccessful blood meal attempt (calculate with \code{1 - MBDETES_PrLeave(site,"feed",TRUE)}, uses parameters \code{boutFail_p}, \code{disperse})}
}
\description{
Produce named list of parameters for \code{\link{BFAB_B2X}} and \code{\link{BFAB_R2X}}.
This function is useful for finding parameter sets that map between
MBITES and MBDETES parameter spaces.
}
