% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{BFAB_R2X}
\alias{BFAB_R2X}
\title{MBDETES: Resting Period Transition Probability Vector}
\usage{
BFAB_R2X(PAR)
}
\arguments{
\item{PAR}{a named list of parameters}
}
\description{
Calculate vector of probabilities to transition between: R-> B, F, L, O, D.
This function is called from \code{\link{MBDETES_RperiodTransitions}}.
}
