% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{BFSB_F2X}
\alias{BFSB_F2X}
\title{MBDETES: Blood Feeding Search Bout Transition Probabilities}
\usage{
BFSB_F2X(PAR)
}
\arguments{
\item{PAR}{a named list of parameters}
}
\description{
After receiving parameter values, compute state transition probabilities
from the blood feeding search bout state to all other states.
This function is called from \code{\link{MBDETES_FstateTransitions}}.
}
