% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bionomics.R
\name{Bionomics_absoluteDisperse}
\alias{Bionomics_absoluteDisperse}
\title{Bionomics: Compute Absolute Dispersal}
\usage{
Bionomics_absoluteDisperse(mosquitos)
}
\arguments{
\item{mosquitos}{a data.frame of parsed JSON mosquito output}
}
\description{
Takes in JSON output parsed into a data.frame object from
an MBITES simulation run and returns a data.frame of mosquito lifespans.
Mosquitoes that were still alive at the end of simulation are filtered out.
Returns a list where each element gives the sites a mosquito visited. If the mosquito
never left the natal site, just return 0 (0 cumulative distance traveled).
}
