% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bionomics.R
\name{Bionomics_bloodfeedingRate}
\alias{Bionomics_bloodfeedingRate}
\title{Bionomics: Compute Blood Feeding Events by Age}
\usage{
Bionomics_bloodfeedingRate(mosquitos)
}
\arguments{
\item{mosquitos}{a data.frame of parsed JSON mosquito output}
}
\description{
Takes in JSON output parsed into a data.frame object from
an MBITES simulation run.
Computes for all processed mosquitoes a vector of ages when successful blood feeding events
were recorded.
Blood feeding rate by age can be calculated from this output by calculating the density
of events as a function of age.
Mosquitoes that were still alive at the end of simulation or never oviposited are filtered out.
}
