% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bionomics.R
\name{Bionomics_humanBitingRate}
\alias{Bionomics_humanBitingRate}
\title{Bionomics: Compute Proportion of Blood Meals on Humans
Takes in JSON output parsed into a data.frame object from
an MBITES simulation run.
This function returns a single value, corresponding to \deqn{Q} in
Ross-Macdonald parameters.
Mosquitoes that were still alive at the end of simulation are filtered out.}
\usage{
Bionomics_humanBitingRate(mosquitos)
}
\arguments{
\item{mosquitos}{a data.frame of parsed JSON mosquito output}
}
\description{
Bionomics: Compute Proportion of Blood Meals on Humans
Takes in JSON output parsed into a data.frame object from
an MBITES simulation run.
This function returns a single value, corresponding to \deqn{Q} in
Ross-Macdonald parameters.
Mosquitoes that were still alive at the end of simulation are filtered out.
}
