% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bionomics.R
\name{Bionomics_humanBloodHost}
\alias{Bionomics_humanBloodHost}
\title{Bionomics: Compute Number of Blood Meals
Takes in JSON output parsed into a data.frame object from
an MBITES simulation run.
This function returns a data.frame of lifetime blood meals
for each mosquito for a chosen host type. It does not filter mosquitos who took no blood meals
(that is to say 0s are counted and included in the output).
Mosquitoes that were still alive at the end of simulation are filtered out.}
\usage{
Bionomics_humanBloodHost(mosquitos, who = "human")
}
\arguments{
\item{mosquitos}{a data.frame of parsed JSON mosquito output}

\item{who}{either 'human', 'zoo', or 'all'}
}
\description{
Bionomics: Compute Number of Blood Meals
Takes in JSON output parsed into a data.frame object from
an MBITES simulation run.
This function returns a data.frame of lifetime blood meals
for each mosquito for a chosen host type. It does not filter mosquitos who took no blood meals
(that is to say 0s are counted and included in the output).
Mosquitoes that were still alive at the end of simulation are filtered out.
}
