% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Bionomics.R
\name{Bionomics_MBDETES_Approx}
\alias{Bionomics_MBDETES_Approx}
\title{Bionomics: Calculate MBITES Bionomics and run Approximating MBDETES Simulation}
\usage{
Bionomics_MBDETES_Approx(mosquitos)
}
\arguments{
\item{mosquitos}{a processed data frame of mosquito output (use \code{\link[jsonlite]{fromJSON}}) to read in MBITES output}
}
\description{
Bionomics: Calculate MBITES Bionomics and run Approximating MBDETES Simulation
}
