% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bionomics.R
\name{Bionomics_ovipositionRate}
\alias{Bionomics_ovipositionRate}
\title{Bionomics: Compute Batch Sizes (Oviposition Events) by Age}
\usage{
Bionomics_ovipositionRate(mosquitos)
}
\arguments{
\item{mosquitos}{a data.frame of parsed JSON mosquito output}
}
\description{
Takes in JSON output parsed into a data.frame object from
an MBITES simulation run.
Computes for all processed mosquitoes a vector of ages when successful oviposition events
were recorded and a vector of egg batch sizes sorted by ages (vectors are the same length).
Oviposition rate by age can be calculated from this output by calculating the density
of events as a function of age.
Mosquitoes that were still alive at the end of simulation or never oviposited are filtered out.
}
