% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hash-Class.R
\name{eapply_HashMap}
\alias{eapply_HashMap}
\title{MASH-CPP: HashMap Call a Method Over All Values in HashMap}
\usage{
eapply_HashMap(tag, return = FALSE, ...)
}
\arguments{
\item{tag}{character tag of the method to call for each element in the HashMap}

\item{return}{do we want to capture the ouput of \code{\link[base]{eapply}} and return; if \code{FALSE} apply the function invisibly and return \code{NULL}}

\item{...}{additional named arguments passed to method}
}
\description{
Use \code{\link[base]{eapply}} to call a method found by argument \code{tag} for all values stored in the HashMap \code{private$storage}.
The HashMap class is intended to only store instantiations of R6 objects from a single class, it must not store mixed data types.
}
