% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{ELAB_O2X}
\alias{ELAB_O2X}
\title{MBDETES: Egg Laying Attempt Bout Transition Probabilities}
\usage{
ELAB_O2X(PAR)
}
\arguments{
\item{PAR}{a named list of parameters}
}
\description{
After receiving parameter values, compute state transition probabilities
from the egg laying attempt bout state to all other states.
This function is called from \code{\link{MBDETES_OstateTransitions}}.
}
