% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Parameters.R
\name{ELAB_PAR}
\alias{ELAB_PAR}
\title{MBDETES: Parameters for Transitions from O}
\usage{
ELAB_PAR(A = 0.83, B0 = 0, B1 = 1, C1 = 0, C2 = 1, C4 = 0,
  D1 = 0.95, D2 = 0.95, E = 0, F1 = 0.7, F2 = 0.8, F3 = 0.7)
}
\arguments{
\item{A}{probability of successful launch (maps to \code{O_succeed})}

\item{B0}{probability of empty habitats (can create \code{\link{Site}} objects with no aquatic habitats through \code{\link{Tile_Initialize}})}

\item{B1}{probability to approach an aquatic habitat (no traps currently in MBITES so must equal 1)}

\item{C1}{probability of death during aquatic habitat approach (no current MBITES mapping ...)}

\item{C2}{probability to not be deterred during aquatic habitat approach (no current MBITES mapping ...)}

\item{C4}{probability of death when encountering an ovitrap (pending traps in MBITES ...)}

\item{D1}{probability to survive resting (\code{\link{upstateState}}) given successful oviposition (calculate with \code{MBDETES_PrSurvive(site,"O")}, uses parameters \code{O_surv}, and the local \code{\link{Site}} object's hazard)}

\item{D2}{probability to survive resting (\code{\link{upstateState}}) given unsuccessful oviposition (calculate with \code{MBDETES_PrSurvive(site,"O")}, uses parameters \code{O_surv}, and the local \code{\link{Site}} object's hazard)}

\item{E}{probability of skip oviposition (no skip oviposition in MBITES yet so must equal 0)}

\item{F1}{only used if skip-oviposition is implementd (but E must be 0)}

\item{F2}{probability to leave the current \code{\link{Site}} and go on a blood feeding search given an successful oviposition (calculate with \code{1 - MBDETES_PrLeave(site,"feed",FALSE)}, uses parameters \code{boutFail_p}, \code{disperse})}

\item{F3}{probability to leave the current \code{\link{Site}} and go on a oviposition search given an unsuccessful oviposition attempt (calculate with \code{1 - MBDETES_PrLeave(site,"aqua",TRUE)}, uses parameters \code{boutFail_p}, \code{disperse})}
}
\description{
Produce named list of parameters for \code{\link{ELAB_O2X}}.
This function is useful for finding parameter sets that map between
MBITES and MBDETES parameter spaces.
}
