% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{ELSB_L2X}
\alias{ELSB_L2X}
\title{MBDETES: Egg Laying Search Bout Transition Probabilities}
\usage{
ELSB_L2X(PAR)
}
\arguments{
\item{PAR}{a named list of parameters}
}
\description{
After receiving parameter values, compute state transition probabilities
from the egg laying search bout state to all other states.
This function is called from \code{\link{MBDETES_LstateTransitions}}.
}
