% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hash-Class.R
\name{exists_HashMap}
\alias{exists_HashMap}
\title{HashMap: Check if Value Exists}
\usage{
exists_HashMap(key)
}
\arguments{
\item{key}{given a character key return a boolean if the value associated to that key exists in the storage hash table}
}
\description{
Does a value exist in \code{private$storage} or not.
}
