% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Null.R
\name{feedHost_NULL}
\alias{feedHost_NULL}
\title{Null Host Blood feeding}
\usage{
feedHost_NULL()
}
\description{
This method fills in for pathogen model specific host blood feeding (human to mosquito transmission) methods
if using the null human & pathogen model. It calls \code{\link{pushFeed_Human_NULL}} to log probing events.
}
