% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{fequal}
\alias{fequal}
\title{Approximate Equality for Floating Point Numbers}
\usage{
fequal(a, b, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{a}{\code{\link[base]{numeric}} value}

\item{b}{\code{\link[base]{numeric}} value}

\item{tol}{numeric tolerance}
}
\description{
Approximate Equality for Floating Point Numbers
}
