% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Parameters.R
\name{get_aqua_model_MBITES_Parameters}
\alias{get_aqua_model_MBITES_Parameters}
\title{get the string telling you what aquatic ecology model is being used}
\usage{
get_aqua_model_MBITES_Parameters()
}
\description{
get the string telling you what aquatic ecology model is being used
}
