% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Site.R
\name{get_aqua_Site}
\alias{get_aqua_Site}
\title{Site: Return Aquatic Habitat Resource}
\usage{
get_aqua_Site(i)
}
\description{
Returns a reference to a \code{\link{Aqua_Resource}} object at this site.
\itemize{
\item binding: \code{Site$get_aqua}
}
}
