% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Parameters.R
\name{get_boutFail_p_MBITES_Parameters}
\alias{get_boutFail_p_MBITES_Parameters}
\title{get 1/number of failed bouts until mosquito gives up and searches}
\usage{
get_boutFail_p_MBITES_Parameters()
}
\description{
get 1/number of failed bouts until mosquito gives up and searches
}
