% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Parameters.R
\name{get_disperse_MBITES_Parameters}
\alias{get_disperse_MBITES_Parameters}
\title{get dispersal (parameter to move even if resources are present)}
\usage{
get_disperse_MBITES_Parameters()
}
\description{
get dispersal (parameter to move even if resources are present)
}
