% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Feeding.R
\name{get_enter_p_Feeding_Resource}
\alias{get_enter_p_Feeding_Resource}
\title{Blood Feeding Resource: Get House Entry Probability}
\usage{
get_enter_p_Feeding_Resource()
}
\description{
Get the probability of a mosquito successfully enters this house.
\itemize{
\item This method is bound to \code{Feeding_Resource$get_enter_p()}.
}
}
