% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Site.R
\name{get_feed_Site}
\alias{get_feed_Site}
\title{Site: Return Blood Feeding Resource Reference}
\usage{
get_feed_Site(i)
}
\arguments{
\item{i}{integer index of resource to return}
}
\description{
Return a reference to the i'th \code{\link{Feeding_Resource}}
\itemize{
\item binding: \code{Site$get_feed}
}
}
