% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Site.R
\name{get_feed_w_Site}
\alias{get_feed_w_Site}
\title{Site: Return Blood Feeding Resource Weights}
\usage{
get_feed_w_Site()
}
\description{
Return a numeric vector of \code{\link[MBITES]{Feeding_Resource}} weights at this site.
\itemize{
\item binding: \code{Site$get_feed_w}
}
}
