% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hash-Class.R
\name{get_HashMap}
\alias{get_HashMap}
\title{HashMap: Get the Value Assigned to a Key}
\usage{
get_HashMap(key)
}
\arguments{
\item{key}{character key}

\item{value}{a value to be assigned}
}
\description{
Assign a key-value pair to \code{private$storage}.
}
