% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Globals.R
\name{get_human_id_MBITES_Globals}
\alias{get_human_id_MBITES_Globals}
\title{MBITES Globals: Get a new Human ID}
\usage{
get_human_id_MBITES_Globals()
}
\description{
Increments and gets a new mosquito ID; this function should only be called from the constructor of \code{\link{Human_NULL}} or Human
objects.
}
\details{
\itemize{
\item This method is bound to \code{MBITES_Globals$get_human_id}.
}
}
