% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Parameters.R
\name{get_InAndOut_row_MBITES_Parameters}
\alias{get_InAndOut_row_MBITES_Parameters}
\title{should be a named matrix because indexing is going to be on dimension names.}
\usage{
get_InAndOut_row_MBITES_Parameters(i)
}
\description{
should be a named matrix because indexing is going to be on dimension names.
}
