% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Site.R
\name{get_mate_Site}
\alias{get_mate_Site}
\title{Site: Return Mating Swarm Resource}
\usage{
get_mate_Site(i)
}
\description{
Returns a reference to the first \code{\link{Mating_Resource}} object at this site.
\itemize{
\item binding: \code{Site$get_mate}
}
}
