% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Globals.R
\name{get_mosquito_m_id_MBITES_Globals}
\alias{get_mosquito_m_id_MBITES_Globals}
\title{MBITES Globals: Get a new Male Mosquito ID}
\usage{
get_mosquito_m_id_MBITES_Globals()
}
\description{
Increments and gets a new mosquito ID; this function should only be called from the constructor of \code{\link[MBITES]{Mosquito}}
objects.
}
\details{
\itemize{
\item This method is bound to \code{MBITES_Globals$get_mosquito_m_id}.
}
}
