% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Globals.R
\name{get_n_tile_MBITES_Globals}
\alias{get_n_tile_MBITES_Globals}
\title{MBITES Globals: Get Number of Tiles}
\usage{
get_n_tile_MBITES_Globals()
}
\description{
Return the total number of tiles created.
}
\details{
\itemize{
\item This method is bound to \code{MBITES_Globals$get_n_tile}.
}
}
