% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Resource.R
\name{get_site_Resource}
\alias{get_site_Resource}
\title{Resource: Get Site}
\usage{
get_site_Resource()
}
\description{
Get a reference to the enclosing \code{\link[MBITES]{Site}} object.
\itemize{
\item binding: \code{Resource$get_site}
}
}
