% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Site.R
\name{get_sugar_Site}
\alias{get_sugar_Site}
\title{Site: Return Sugar Feeding Resource}
\usage{
get_sugar_Site(i)
}
\description{
Returns a reference to a first \code{\link{Sugar_Resource}} object at this site.
\itemize{
\item binding: \code{Site$get_sugar}
}
}
