% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Globals.R
\name{get_tile_MBITES_Globals}
\alias{get_tile_MBITES_Globals}
\title{MBITES Globals: Return a Tile Reference}
\usage{
get_tile_MBITES_Globals(id)
}
\arguments{
\item{id}{an integer tile id}
}
\description{
Returns the reference to the \code{\link[MBITES]{Tile}} object with the associated id.
}
\details{
\itemize{
\item This method is bound to \code{MBITES_Globals$get_tile}.
}
}
