% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Site.R
\name{get_tile_Site}
\alias{get_tile_Site}
\title{Site: Return Tile Reference}
\usage{
get_tile_Site()
}
\description{
Return a reference to the enclosing \code{\link[MBITES]{Tile}}
\itemize{
\item binding: \code{Site$get_tile}
}
}
