% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Globals.R
\name{get_tileID_MBITES_Globals}
\alias{get_tileID_MBITES_Globals}
\title{MBITES Globals: Get a new Tile ID}
\usage{
get_tileID_MBITES_Globals()
}
\description{
Increments and gets a new tile ID; this function should only be called from the constructor of \code{\link{Tile}}
objects.
}
\details{
\itemize{
\item This method is bound to \code{MBITES_Globals$get_tileID}.
}
}
