% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Site.R
\name{get_tileID_Site}
\alias{get_tileID_Site}
\title{Site: Return ID of Tile this Site is in}
\usage{
get_tileID_Site()
}
\description{
Return the id to the enclosing \code{\link{Tile}}
\itemize{
\item binding: \code{Site$get_tileID}
}
}
