% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Site.R
\name{get_type_Site}
\alias{get_type_Site}
\title{Site: Return Type of this Site}
\usage{
get_type_Site()
}
\description{
Return the integer type of this site.
\itemize{
\item binding: \code{Site$get_type}
}
}
