% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Aqua.R
\name{get_w_Aqua_Resource}
\alias{get_w_Aqua_Resource}
\title{Aqua_Resource: Get Resource Weight}
\usage{
get_w_Aqua_Resource()
}
\description{
Get the weight associated to this resource.
\itemize{
\item binding: \code{Aqua_Resource$get_w}
}
}
