% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Feeding.R
\name{get_w_Feeding_Resource}
\alias{get_w_Feeding_Resource}
\title{Feeding_Resource: Get Resource Weight}
\usage{
get_w_Feeding_Resource()
}
\description{
Get the weight associated to this resource.
\itemize{
\item binding: \code{Feeding_Resource$get_w}
}
}
