% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Mating.R
\name{get_w_Mating_Resource}
\alias{get_w_Mating_Resource}
\title{Mating_Resource: Get Resource Weight}
\usage{
get_w_Mating_Resource()
}
\description{
Get the weight associated to this resource.
\itemize{
\item binding: \code{Mating_Resource$get_w}
}
}
