% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Resource.R
\name{get_w_Resource}
\alias{get_w_Resource}
\title{Resource: Get Weight}
\usage{
get_w_Resource()
}
\description{
Get the numeric weight associated with this resource.
\itemize{
\item binding: \code{Resource$get_w}
}
}
