% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Parameters.R
\name{get_wts_MBITES_Parameters}
\alias{get_wts_MBITES_Parameters}
\title{MBITES Parameters: Return Resting Spot Weights for Behavioral States}
\usage{
get_wts_MBITES_Parameters(state)
}
\description{
Return the specific weights for different resting spots associated with mosquito behavioral states.
}
