% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Globals.R
\name{hardreset_MBITES_Globals}
\alias{hardreset_MBITES_Globals}
\title{MBITES Globals: Hard Reset}
\usage{
hardreset_MBITES_Globals()
}
\description{
If there is a severe error, preform a hard reset (erase all tiles, reset all world-state globals, and clear logging connections).
This can be accessed by users via \code{\link{hardreset}}
\itemize{
\item This method is bound to \code{MBITES_Globals$hardreset}
}
}
