% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Site.R
\name{has_feed_Site}
\alias{has_feed_Site}
\title{Site: Has Blood Feeding Resources?}
\usage{
has_feed_Site()
}
\description{
Query if this site has \code{\link[MBITES]{Feeding_Resource}} present.
\itemize{
\item binding: \code{Site$has_feed}
}
}
