% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Site.R
\name{has_mate_Site}
\alias{has_mate_Site}
\title{Site: Has Mating Swarms?}
\usage{
has_mate_Site()
}
\description{
Query if this site has \code{\link[MBITES]{Mating_Resource}} present.
\itemize{
\item binding: \code{Site$has_mate}
}
}
