% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Site.R
\name{has_sugar_Site}
\alias{has_sugar_Site}
\title{Site: Has Sugar Feeding Resources?}
\usage{
has_sugar_Site()
}
\description{
Query if this site has \code{\link[MBITES]{Sugar_Resource}} present.
\itemize{
\item binding: \code{Site$has_sugar}
}
}
