% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hash-Class.R
\name{hash_apply}
\alias{hash_apply}
\title{Map a function over a hash table}
\usage{
hash_apply(X, FUN, ...)
}
\arguments{
\item{X}{a hashed \code{\link[base]{environment}}}

\item{FUN}{a function}

\item{...}{optional named arguments}
}
\description{
This is an internal function for MBITES that rewrites eapply such that
no memory is allocated for return values, the function is mapped with optional
arguments across all objects in the hash table.
It is only for functions being called for their side effects.
}
