% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hash-Class.R
\name{hash_size}
\alias{hash_size}
\title{Get size of hash table}
\usage{
hash_size(rho)
}
\arguments{
\item{rho}{a hashed \code{\link[base]{environment}}}
}
\description{
Return the total number of objects stored in the hash table.
}
