% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-HostEncounter.R
\name{hostEncounter}
\alias{hostEncounter}
\title{MBITES: Host Encounter}
\description{
Due to the different design of the R6/C++ OOP version of MASH-MICRO, all
host probing and infection functions are moved to their respective PATHOGEN
method files PATHOGEN-XX-Methods.R
}
