% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Null.R
\name{HUMAN-Null}
\alias{HUMAN-Null}
\title{HUMAN: Null Human}
\description{
The null human class (see \code{\link{Human_NULL}}) is a simple logging class to record biting events
when human dynamics are not of interest. It can only be used with the \code{\link{PathogenNull}} model.
}
