% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Null.R
\docType{class}
\name{Human_NULL}
\alias{Human_NULL}
\title{NULL Human Class}
\format{
An \code{\link{R6Class}} generator object
}
\description{
NULL Human Class

NULL Human Class
}
\details{
A \code{Human_NULL} can be used as a drop-in replacement when human dynamics do not need
to be simulated explicitly.
}
\section{\strong{Constructor}}{

\itemize{
\item id: integer id
\item w: numeric biting weight
\item feedingID: id of the \code{\link{Feeding_Resource}} where my biting weight will be added
\item siteID: id of the \code{\link{Site}} where my feeding site resides
\item tileID: id of the \code{\link{Tile}} where I reside
}
}

\section{\strong{Methods}}{

\itemize{
\item method: i'm a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item id: integer id
\item w: numeric biting weight
\item UNBITTEN: logical flag
\item mosquito_id: integer vector of mosquitoes that have bitten me
\item mosquito_t: numeric vector of times i was bitten
\item bloodFeed: \code{FALSE} corresponds to probing-only events, \code{TRUE} corresponds to combined probe-bloodfeed events.
}
}

\keyword{R6}
\keyword{class}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Human_NULL$new()}}
\item \href{#method-add_pathogen}{\code{Human_NULL$add_pathogen()}}
\item \href{#method-feedHost}{\code{Human_NULL$feedHost()}}
\item \href{#method-remove_pathogen}{\code{Human_NULL$remove_pathogen()}}
\item \href{#method-finalize}{\code{Human_NULL$finalize()}}
\item \href{#method-exit}{\code{Human_NULL$exit()}}
\item \href{#method-oneDay_EventQ}{\code{Human_NULL$oneDay_EventQ()}}
\item \href{#method-oneDay_ActivitySpace}{\code{Human_NULL$oneDay_ActivitySpace()}}
\item \href{#method-pushProbe}{\code{Human_NULL$pushProbe()}}
\item \href{#method-pushFeed}{\code{Human_NULL$pushFeed()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a human
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human_NULL$new(id, w, feedingID, siteID, tileID)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Integer identifier of human.}

\item{\code{w}}{biting weight of human.}

\item{\code{feedingID}}{where it is a feeding site}

\item{\code{siteID}}{site for human}

\item{\code{tileID}}{The tile that has that site.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_pathogen"></a>}}
\if{latex}{\out{\hypertarget{method-add_pathogen}{}}}
\subsection{Method \code{add_pathogen()}}{
Add a pathogen to this human's pathogens.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human_NULL$add_pathogen(pathogen)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pathogen}}{The pathogen to add to the list.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feedHost"></a>}}
\if{latex}{\out{\hypertarget{method-feedHost}{}}}
\subsection{Method \code{feedHost()}}{
Decide which pathogens from the host will transmit to the mosquito.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human_NULL$feedHost(mPathogens)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mPathogens}}{The mosquito pathogens that it already has.
Returns a list of new pathogens to add to the mosquito.
This list may contain NULL values when no pathogen is added.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_pathogen"></a>}}
\if{latex}{\out{\hypertarget{method-remove_pathogen}{}}}
\subsection{Method \code{remove_pathogen()}}{
Take a pathogen from this human's list of pathogens.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human_NULL$remove_pathogen(pathogen)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pathogen}}{The pathogen to remove.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Destructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human_NULL$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exit"></a>}}
\if{latex}{\out{\hypertarget{method-exit}{}}}
\subsection{Method \code{exit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human_NULL$exit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-oneDay_EventQ"></a>}}
\if{latex}{\out{\hypertarget{method-oneDay_EventQ}{}}}
\subsection{Method \code{oneDay_EventQ()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human_NULL$oneDay_EventQ()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-oneDay_ActivitySpace"></a>}}
\if{latex}{\out{\hypertarget{method-oneDay_ActivitySpace}{}}}
\subsection{Method \code{oneDay_ActivitySpace()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human_NULL$oneDay_ActivitySpace()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pushProbe"></a>}}
\if{latex}{\out{\hypertarget{method-pushProbe}{}}}
\subsection{Method \code{pushProbe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human_NULL$pushProbe(m_id, t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pushFeed"></a>}}
\if{latex}{\out{\hypertarget{method-pushFeed}{}}}
\subsection{Method \code{pushFeed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Human_NULL$pushFeed()}\if{html}{\out{</div>}}
}

}
}
