% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Initialize.R
\name{Human_NULL_Initialize}
\alias{Human_NULL_Initialize}
\title{Initialize Null Human Populations}
\usage{
Human_NULL_Initialize(human_init)
}
\arguments{
\item{human_init}{a \code{\link{data.frame}} object where each row corresponds to a human}
}
\description{
Initialize \code{\link{Human_NULL}} populations
}
