% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TILE-Initialize.R
\name{Initialize Tile}
\alias{Initialize Tile}
\title{Add a Tile to MBITES Simulation}
\description{
To add a tile to MBITES, \code{\link{Tile_Initialize}} must be called to
initialize the landscape (collection of sites and associated resources) on a tile prior to
initializing human or mosquito populations. This function generates a \code{\link{Tile}} by calling its constructor
which automatically gets a new tile ID and registers the tile with \code{\link{MBITES_Globals}}.

say something about the rest of things on the tile.
}
\details{
\enumerate{
\item Initialize Landscape:
}
\itemize{
\item \code{\link{Tile_Initialize}}: see the help file for the expected parameters
}
\enumerate{
\item Initialize Humans:
\item Initialize Mosquitoes:
\item (Optional) Initialize Pathogens:
}
}
