% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Aqua.R
\name{make_EggQ}
\alias{make_EggQ}
\title{Make a Egg Queue}
\usage{
make_EggQ()
}
\description{
Generates a closure that contains the following fields:
\itemize{
\item batches: integer vector of egg batch sizes
\item times: numeric vector of egg hatching times
}
}
\details{
The closure also contains the following functions:
\itemize{
\item add2Q(batch_n,time_n): add a new egg batch to the queue
\item popQ(time_p): return a vector of egg batches that have hatched
\item printQ(): print the egg queue
}
}
