% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Aqua.R
\name{make_ImagoQ}
\alias{make_ImagoQ}
\title{Make a Imago Queue}
\usage{
make_ImagoQ()
}
\description{
Generates a closure that contains the following fields:
\itemize{
\item imagos: integer vector of emerging imago cohort sizes
\item female: boolean vector indicating female or male cohort
\item times: numeric vector of imago emergence times
}
}
\details{
The closure also contains the following functions:
\itemize{
\item add2Q(batch_n,time_e,imago_f): add a new imago cohort of size \code{imago_n} and boolean sex \code{imago_f} to the queue that will emerge at \code{time_e}
\item popQ(time_e,imago_f): return a vector of imago cohorts that are due to emerge by \code{time_e} and of boolean sex \code{imago_f}
\item printQ(): print the imago queue
}
}
