% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Mating.R
\name{make_MatingQ}
\alias{make_MatingQ}
\title{Make a Mating Queue}
\usage{
make_MatingQ()
}
\description{
Generates a closure that contains the following fields:
\itemize{
\item id: integer vector of males at this mating queue
\item weight: numeric vector of male mating weights
}
}
\details{
The closure also contains the following functions:
\itemize{
\item add2Q(id_n,weight_n): add a new mosquito to the queue
\item clearQ(): clear the mating queue
\item sampleQ(): sample a mosquito id from the queue
\item printQ(): print the mating queue
}
}
