% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Parameters.R
\name{make_ttEvent_Det}
\alias{make_ttEvent_Det}
\title{MBITES Parameters: Make Deterministic Time-to-Event Closure}
\usage{
make_ttEvent_Det(wait)
}
\description{
Generates a closure that contains the following fields:
\itemize{
\item wait: numeric value of deterministic waiting time
}
}
\details{
The closure also contains the following functions:
\itemize{
\item tteDet(t): samples a time to event following shifted exponential distribution
}
}
