% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Parameters.R
\name{make_ttEvent_Diurnal}
\alias{make_ttEvent_Diurnal}
\title{Make a Shifted Diurnal Time-to-Event Closure}
\usage{
make_ttEvent_Diurnal(peak, tmin)
}
\description{
Generates a closure that contains the following fields:
\itemize{
\item field: i'm a field!
}
}
\details{
The closure also contains the following functions:
\itemize{
\item function: i'm a function!
}
}
