% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Parameters.R
\name{make_ttEvent_Exp}
\alias{make_ttEvent_Exp}
\title{Make a Shifted Exponential Time-to-Event Closure}
\usage{
make_ttEvent_Exp(rate, tmin)
}
\description{
Generates a closure that contains the following fields:
\itemize{
\item rate: numeric value of rate parameter
\item tmin: numeric value of minimum waiting time
}
}
\details{
The closure also contains the following functions:
\itemize{
\item tteShiftExp(t): samples a time to event following shifted exponential distribution
}
}
