% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Model.R
\name{MBDETES-Model}
\alias{MBDETES-Model}
\title{MBDETES: Model Specification}
\description{
MBDETES: Model Specification
}
\section{Background}{


Under some restricted conditions on waiting times, MBITES is what we would get if we translated MBITES into a stochastic model, under the Gillespie algorithm. We can, therefore, use MBDETES to compute expectations for MBITES under these same contitions.

The utilties in MBDETES-Auxiliary.R and MBDETES-Model.R outputs state transitions as described in the following.
\itemize{
\item Let \eqn{0\leq P_{X,Y}\leq 1} denote the probability of making a transition from state \eqn{X} to state \eqn{Y} after one bout
\item The probability of dying, \eqn{0 \leq P_{X,D} \leq 1} must satifsy the condition \deqn{\sum_{Y\neq D}P_{X,Y} =1-P_{X,D}}
\item Since there is the probability of repeating a state, we must compute the probability of making a transition out of the state and ending up alive and in a new state. We let this quantity be \eqn{\Psi_{X,Y}}
\item Let \eqn{T_{X}} the average time spent in a bout of type \eqn{X}.
\item Let \eqn{T_{X,Y}} denote the expected waiting time to make the transition from \eqn{X} to \eqn{Y}.
}
}

\section{Differential equations}{


In some sense, we can rewrite MBDETES as a conservation equation in the following way:
\deqn{\left[ \begin{array}{c}dF/dt \\ dB/dt \\ dR/dt \\ dL/dt \\ dO/dt \\ dD/dt\end{array}\right]=
 \left[ \begin{array}{ccccc}
 P_{F,F}-1 & P_{F,B} & 0& 0 & 0 & P_{F,D} \\
 P_{B,F} & P_{B,B}-1 & P_{B,R} & 0 &0 & P_{B,D} \\
 P_{R,F} & P_{R,B} & -1 & P_{R,L} & P_{R,O}& P_{R,D} \\
 0 & 0 & 0 & P_{L,L}-1& P_{L,O} & P_{L,D} \\
 P_{O,F} & P_{O,B} & 0 & P_{O,L} & P_{O,O}-1 & P_{O,D} \\
 P_{F,D} & P_{B,D} & P_{R,D} & P_{L,D} & P_{O,D} & 0
 \end{array}\right] \left[ \begin{array}{c}F/t_F\\B/t_B\\R/t_R\\L/t_R\\O/t_O \\D   \\ \end{array}\right]}
}

