% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{MBDETES_Approx}
\alias{MBDETES_Approx}
\title{MBDETES: Calculate State Transitions Matrix}
\usage{
MBDETES_Approx(tileID)
}
\arguments{
\item{tileID}{an integer ID of a tile (must have already called \code{\link{Tile_Initialize}} and the appropriate \code{Human_XX_Initialize} routines)}
}
\description{
For each \code{\link{Site}} in a landscape, calculate approximate
state transitions matrix by calling \code{\link{MBDETES_StateTransitions}}.
}
