% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{MBDETES_BstateTransitions}
\alias{MBDETES_BstateTransitions}
\title{MBDETES: Blood Feeding Attempt Bout State Transitions}
\usage{
MBDETES_BstateTransitions(site)
}
\arguments{
\item{site}{a \code{\link{Site}} object}
}
\description{
Calculate vector of probabilities to transition between: B -> F, B, R, D
This function is called from \code{\link{MBDETES_StateTransitions}}.
}
