% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Transitions.R
\name{MBDETES_FstateTransitions}
\alias{MBDETES_FstateTransitions}
\title{MBDETES: Blood Feeding Search Bout State Transitions}
\usage{
MBDETES_FstateTransitions(site)
}
\arguments{
\item{site}{a \code{\link{Site}} object}
}
\description{
Calculate vector of probabilities to transition between: F -> F, B, D.
Probability vector is calculated in \code{\link{BFSB_F2X}}.
This function is called from \code{\link{MBDETES_StateTransitions}}.
}
