% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{MBDETES_getLeaveUnladen}
\alias{MBDETES_getLeaveUnladen}
\title{MBDETES: Probability of Leaving, post prandially (unladen mosquito)}
\usage{
MBDETES_getLeaveUnladen()
}
\description{
MBDETES: Probability of Leaving, post prandially (unladen mosquito)
}
