% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{MBDETES_getPrOverfeed}
\alias{MBDETES_getPrOverfeed}
\title{MBDETES: Overfeeding Probability}
\usage{
MBDETES_getPrOverfeed()
}
\description{
Computes the integral of the probability of
overfeeding conditioned on the size of the bloodmeal,
multiplied by the probability of bloodmeal
of that size.
\itemize{
\item MBITES functions can be found in MBITES-BloodMeal.R
}
}
