% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{MBDETES_getPrPPRFlight}
\alias{MBDETES_getPrPPRFlight}
\title{MBDETES: Probability to Die from post-prandial period (1 - pPPR = survival prob)}
\usage{
MBDETES_getPrPPRFlight()
}
\description{
MBDETES: Probability to Die from post-prandial period (1 - pPPR = survival prob)
}
