% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{MBDETES_getSurvUnladen}
\alias{MBDETES_getSurvUnladen}
\title{MBDETES: Probability of Surviving the post-prandial flight (unladen mosquito)}
\usage{
MBDETES_getSurvUnladen()
}
\description{
MBDETES: Probability of Surviving the post-prandial flight (unladen mosquito)
}
