% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{MBDETES_LeaveLaden}
\alias{MBDETES_LeaveLaden}
\title{MBDETES: Probability of Leaving, post prandially (laden mosquito)}
\usage{
MBDETES_LeaveLaden()
}
\description{
MBDETES: Probability of Leaving, post prandially (laden mosquito)
}
