% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Calibration.R
\name{MBDETES_LeaveUnladen}
\alias{MBDETES_LeaveUnladen}
\title{MBDETES: Probability of Leaving, post prandially (unladen mosquito)}
\usage{
MBDETES_LeaveUnladen()
}
\description{
MBDETES: Probability of Leaving, post prandially (unladen mosquito)
}
