% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{MBDETES_LocalHazMortality}
\alias{MBDETES_LocalHazMortality}
\title{MBDETES: Probability of Death from Local (Site-specific) Hazards}
\usage{
MBDETES_LocalHazMortality(site)
}
\description{
Mortality due to local hazards is a Bernoulli event (see \code{\link{mbites_surviveHazards}})
so this function just returns the probability of death.
}
