% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{MBDETES_LstateTransitions}
\alias{MBDETES_LstateTransitions}
\title{MBDETES: Egg Laying Search Bout State Transitions}
\usage{
MBDETES_LstateTransitions(site)
}
\arguments{
\item{site}{a \code{\link{Site}} object}
}
\description{
Calculate vector of probabilities to transition between: L -> L, O, D.
Probability vector is calculated in \code{\link{ELSB_L2X}}.
This function is called from \code{\link{MBDETES_StateTransitions}}.
}
