% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Model.R
\name{MBDETES_MortalityPars}
\alias{MBDETES_MortalityPars}
\title{MBDETES: Mortality Probabilities}
\usage{
MBDETES_MortalityPars(PAR)
}
\arguments{
\item{PAR}{output of \code{\link{MBDETES_Parameters}}}
}
\value{
the input list augmented with mortality probabilities.
}
\description{
Calculate {all states} -> D probabilities.
}
