% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Model.R
\name{MBDETES_Parameters}
\alias{MBDETES_Parameters}
\title{MBDETES: Model Parameters}
\usage{
MBDETES_Parameters(tF = 6/24, tB = 3/24, tR = 18/24, tL = 6/24,
  tO = 3/24, P_FF = 0.15, P_FB = 0.83, P_BF = 0.15, P_BB = 0.05,
  P_BR = 0.75, P_LL = 0.15, P_LO = 0.83, P_OL = 0.15, P_OO = 0.05,
  P_OB = 0.3, P_OF = 0.45, P_RF = 0.05, P_RB = 0.2, P_RL = 0.4,
  P_RO = 0.35)
}
\arguments{
\item{tF}{sojourn time in blood feeding search bout}

\item{tB}{sojourn time in blood feeding attempt bout}

\item{tR}{sojourn time in post-prandial resting bout}

\item{tL}{sojourn time in egg laying search bout}

\item{tO}{sojourn time in egg laying attempt bout}

\item{P_FF}{blood feeding search to search transition probability}

\item{P_FB}{blood feeding search to attempt transition probability}

\item{P_BF}{blood feeding attempt to search transition probability}

\item{P_BB}{blood feeding attempt to attempt transition probability}

\item{P_BR}{blood feeding attempt to post-prandial resting transition probability}

\item{P_RF}{post-prandial resting to blood feeding search transition probability}

\item{P_RB}{post-prandial resting to blood feeding attempt transition probability}

\item{P_RL}{post-prandial resting to egg laying search transition probability}

\item{P_RO}{post-prandial resting to egg laying attempt transition probability}
}
\description{
Return a named list of parameters for MBDETES.
}
