% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{MBDETES_PrLeave}
\alias{MBDETES_PrLeave}
\title{MBDETES: Probability to Leave a Site
Return the probability for a mosquito to leave a site and (initiate a search).}
\usage{
MBDETES_PrLeave(site, res, fail = FALSE)
}
\arguments{
\item{site}{a \code{\link{Site}} object}

\item{res}{either 'aqua' or 'feed'}

\item{fail}{did the attempt bout result in a failure?}
}
\description{
MBDETES: Probability to Leave a Site
Return the probability for a mosquito to leave a site and (initiate a search).
}
