% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Calibration.R
\name{MBDETES_PrLeave_pars}
\alias{MBDETES_PrLeave_pars}
\title{MBDETES: Probability to Leave a Site (Paramter Calibration)}
\usage{
MBDETES_PrLeave_pars(site, res, boutFail_p, disperse, fail = FALSE)
}
\arguments{
\item{site}{does the \code{\link{Site}} have the resource the mosquito needs?}

\item{res}{either 'aqua' or 'feed'}

\item{boutFail_p}{probability to search even if the site has what you need after a failure}

\item{disperse}{baseline probability to leave}

\item{fail}{did the attempt bout result in a failure?}
}
\description{
Parameter explicit version of \code{\link{MBDETES_PrLeave}}
}
