% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{MBDETES_PrOverfeed_calibrate}
\alias{MBDETES_PrOverfeed_calibrate}
\title{MBDETES: Overfeeding Probability (Parameter Calibration)}
\usage{
MBDETES_PrOverfeed_calibrate(of_a, of_b, bm_a, bm_b)
}
\arguments{
\item{of_a}{shape parameters for overfeeding}

\item{of_b}{shape parameters for overfeeding}

\item{bm_a}{alpha parameter for beta distributed blood meal size}

\item{bm_b}{beta parameter for beta distributed blood meal size}
}
\description{
Calculate the expectation probability of overfeeding
(integrating the probability with respect to the probability density function describing blood meal size).
}
