% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{MBDETES_PrPPRFlight}
\alias{MBDETES_PrPPRFlight}
\title{MBDETES: Probability to Survive Post-prandial Resting Flight}
\usage{
MBDETES_PrPPRFlight()
}
\description{
Computes the intergral of the probability of surviving
the post-prandial resting flight conditioned on the
size of the bloodmeal, multiplied by the probability of
a bloodmeal of that size.
\itemize{
\item MBITES functions can be found in MBITES-BloodMeal.R
}
}
