% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{MBDETES_PrPPRFlight_calibrate}
\alias{MBDETES_PrPPRFlight_calibrate}
\title{MBDETES: Probability to Survive Post-prandial Resting Flight (Parameter Calibration)}
\usage{
MBDETES_PrPPRFlight_calibrate(PPR_a, PPR_b, bm_a, bm_b)
}
\arguments{
\item{PPR_a}{shape parameters for ppr mortality}

\item{PPR_b}{shape parameters for ppr mortality}

\item{bm_a}{alpha parameter for beta distributed blood meal size}

\item{bm_b}{beta parameter for beta distributed blood meal size}
}
\description{
Calculate the expectation probability to survive the post-prandial resting flight
(integrating the probability with respect to the probability density function describing blood meal size).
}
