% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Calibration.R
\name{MBDETES_PrPPRFlight_closure}
\alias{MBDETES_PrPPRFlight_closure}
\title{MBDETES: Probability to Survive Post-prandial Resting Flight Objective Function (Parameter Calibration)}
\usage{
MBDETES_PrPPRFlight_closure(val)
}
\arguments{
\item{val}{target value of paramter 'E'}
}
\description{
Objective function for optimization to find \code{PPR_a} and \code{PPR_b} to match target value of 'E'
parameter (probability to survive post-prandial resting flight) by minimizing absolute value of difference between \code{\link{MBDETES_PrRefeed_pars}} evaluated at parameters \code{theta} and argument \code{val}.
Returns a closure that can be passed to optimization functions.
}
