% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Calibration.R
\name{MBDETES_PrPPRFlight_optim}
\alias{MBDETES_PrPPRFlight_optim}
\title{MBDETES: Probability to Survive Post-prandial Resting Flight Calibration}
\usage{
MBDETES_PrPPRFlight_optim(E, PPR_a = 15, PPR_b = 300)
}
\arguments{
\item{E}{target value of 'E' (probability to survive post-prandial resting flight)}

\item{PPR_a}{starting values for ppr mortality shape parameters}

\item{PPR_b}{starting values for ppr mortality shape parameters}
}
\description{
Use \code{\link[stats]{constrOptim}} to find ppr mortality shape parameters to match target MBDETES paramter value.
If providing starting values, \code{PPR_a} must be less than \code{PPR_b} to match the linear constraints used in the optimization.
}
