% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Calibration.R
\name{MBDETES_PrPPRFlight_pars}
\alias{MBDETES_PrPPRFlight_pars}
\title{MBDETES: Probability to Survive Post-prandial Resting Flight (Parameter Calibration)}
\usage{
MBDETES_PrPPRFlight_pars(PPR_a = 15, PPR_b = 300, bm_a = 7.5,
  bm_b = 2.5)
}
\arguments{
\item{PPR_a}{shape parameters for ppr mortality}

\item{PPR_b}{shape parameters for ppr mortality}

\item{bm_a}{alpha parameter for beta distributed blood meal size}

\item{bm_b}{beta parameter for beta distributed blood meal size}
}
\description{
Calculate the expectation probability to survive the post-prandial resting flight
(integrating the probability with respect to the probability density function describing blood meal size).
}
