% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Calibration.R
\name{MBDETES_PrRefeed}
\alias{MBDETES_PrRefeed}
\title{MBDETES: Refeeding Probability}
\usage{
MBDETES_PrRefeed()
}
\description{
Computes the integral of the probability of
feeding conditioned on the size of the bloodmeal,
multiplied by the probability of bloodmeal
of that size.
\itemize{
\item MBITES functions can be found in MBITES-Oogenesis.R
}
}
