% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{MBDETES_PrRefeed_calibrate}
\alias{MBDETES_PrRefeed_calibrate}
\title{MBDETES: Refeeding Probability (Parameter Calibration)}
\usage{
MBDETES_PrRefeed_calibrate(rf_a, rf_b, bm_a, bm_b)
}
\arguments{
\item{rf_a}{shape parameters for refeeding}

\item{rf_b}{shape parameters for refeeding}

\item{bm_a}{alpha parameter for beta distributed blood meal size}

\item{bm_b}{beta parameter for beta distributed blood meal size}
}
\description{
Calculate the expectation probability of refeeding
(integrating the probability with respect to the probability density function describing blood meal size).
}
