% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Calibration.R
\name{MBDETES_PrRefeed_closure}
\alias{MBDETES_PrRefeed_closure}
\title{MBDETES: Refeeding Probability Objective Function (Parameter Calibration)}
\usage{
MBDETES_PrRefeed_closure(val)
}
\arguments{
\item{val}{target value of paramter 'G'}
}
\description{
Objective function for optimization to find \code{rf_a} and \code{rf_b} to match target value of 'G'
parameter (probability to refeed following resting) by minimizing absolute value of difference between \code{\link{MBDETES_PrRefeed_pars}} evaluated at parameters \code{theta} and argument \code{val}.
Returns a closure that can be passed to optimization functions.
}
