% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Calibration.R
\name{MBDETES_PrRefeed_optim}
\alias{MBDETES_PrRefeed_optim}
\title{MBDETES: Refeeding Probability Calibration}
\usage{
MBDETES_PrRefeed_optim(G, rf_a = 10, rf_b = 3)
}
\arguments{
\item{G}{target value of 'G' (probability to refeed following resting)}

\item{rf_a}{starting values for refeeding shape parameters}

\item{rf_b}{starting values for refeeding shape parameters}
}
\description{
Use \code{\link[stats]{constrOptim}} to find refeeding shape parameters to match target MBDETES paramter value.
If providing starting values, \code{rf_a} must be greater than \code{rf_b} to match the linear constraints used in the optimization.
}
