% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Calibration.R
\name{MBDETES_PrRefeed_pars}
\alias{MBDETES_PrRefeed_pars}
\title{MBDETES: Refeeding Probability (Parameter Calibration)}
\usage{
MBDETES_PrRefeed_pars(rf_a = 10, rf_b = 3, bm_a = 7.5, bm_b = 2.5)
}
\arguments{
\item{rf_a}{shape parameters for refeeding}

\item{rf_b}{shape parameters for refeeding}

\item{bm_a}{alpha parameter for beta distributed blood meal size}

\item{bm_b}{beta parameter for beta distributed blood meal size}
}
\description{
Calculate the expectation probability of refeeding
(integrating the probability with respect to the probability density function describing blood meal size).
}
