% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Calibration.R
\name{MBDETES_PrSurvive}
\alias{MBDETES_PrSurvive}
\title{MBDETES: Probability to Survive a Bout}
\usage{
MBDETES_PrSurvive(site, bout)
}
\arguments{
\item{site}{a \code{\link{Site}} object}

\item{bout}{character in 'F','B','L','O'}
}
\description{
Compute survival probability as a result of flight and local hazards. This gives the expectation
of survival from \code{\link{mbites_survival}}.
}
