% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Calibration.R
\name{MBDETES_PrSurvive_pars}
\alias{MBDETES_PrSurvive_pars}
\title{MBDETES: Probability to Survive a Bout (Parameter Calibration)}
\usage{
MBDETES_PrSurvive_pars(haz, surv)
}
\arguments{
\item{haz}{hazard associated with the \code{\link{Site}}}

\item{surv}{the baseline survival probability (for 'F' maps to \code{Bs_surv}, 'B' maps to \code{B_surv}, 'L' maps to \code{Os_surv}, 'O' maps to \code{O_surv})}
}
\description{
Parameter explicit version of \code{\link{MBDETES_PrSurvive}}
}
