% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Model.R
\name{MBDETES_R2R_ODE}
\alias{MBDETES_R2R_ODE}
\title{MBDETES: Feeding Cycle (R -> R) ODE Model Derivatives}
\usage{
MBDETES_R2R_ODE(t, X, P)
}
\arguments{
\item{t}{time at which gradient of model is requested}

\item{X}{state vector of model}

\item{P}{output of \code{\link{MBDETES_Parameters}}}
}
\description{
Evaluate derivatives for feeding cycle model.
}
