% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Model.R
\name{MBDETES_R2R_solve}
\alias{MBDETES_R2R_solve}
\title{MBDETES: Feeding Cycle (R -> R) ODE Model Trajectory}
\usage{
MBDETES_R2R_solve(PAR, mxT = 20, dt = 0.001)
}
\arguments{
\item{PAR}{output of \code{\link{MBDETES_Parameters}}}

\item{mxT}{maximum time of numerical solution}

\item{dt}{time-step at which solutions are requested}
}
\description{
Calculate sample trajectory for feeding cycle model.
}
