% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{MBDETES_RestingParam}
\alias{MBDETES_RestingParam}
\title{MBDETES: Resting spot probabilities}
\usage{
MBDETES_RestingParam()
}
\description{
Calculate resting spot probabilities according to a mixture distribution with
equal weights given to blood feeding and oviposition behavioral states (B,O).
}
