% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{MBDETES_RperiodTransitions}
\alias{MBDETES_RperiodTransitions}
\title{MBDETES: Resting Period State Transitions}
\usage{
MBDETES_RperiodTransitions(site)
}
\arguments{
\item{site}{a \code{\link{Site}} object}
}
\description{
Calculate vector of probabilities to transition between: R-> B, F, L, O, D.
Probability vector is calculated in \code{\link{BFAB_R2X}}.
This function is called from \code{\link{MBDETES_StateTransitions}}.
}
