% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Transitions.R
\name{MBDETES_StateTransitions}
\alias{MBDETES_StateTransitions}
\title{MBDETES: The State Transition Matrix}
\usage{
MBDETES_StateTransitions(site)
}
\arguments{
\item{site}{a \code{\link{Site}} object}
}
\description{
Calculate state transitions matrix for one \code{\link{Site}} object.
Transitions are of form FBRLO to FBRLOD (the rows sum to one).
}
