% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Parameters.R
\name{MBDETES_StateTransitions_Interp}
\alias{MBDETES_StateTransitions_Interp}
\title{MBDETES: State Transition Matrix from Interpolating Parameters}
\usage{
MBDETES_StateTransitions_Interp(BFAB_PAR, ELAB_PAR, BFSB_PAR, ELSB_PAR,
  tF = 6/24, tB = 3/24, tR = 18/24, tL = 6/24, tO = 3/24)
}
\arguments{
\item{BFAB_PAR}{from \code{\link{BFAB_PAR}}, gives paramters for B and R transitions}

\item{ELAB_PAR}{from \code{\link{ELAB_PAR}}, gives paramters for O transitions}

\item{BFSB_PAR}{from \code{\link{BFSB_PAR}}, gives paramters for F transitions}

\item{ELSB_PAR}{from \code{\link{ELSB_PAR}}, gives paramters for L transitions}

\item{tF}{sojourn time in blood feeding search bout}

\item{tB}{sojourn time in blood feeding attempt bout}

\item{tR}{sojourn time in post-prandial resting bout}

\item{tL}{sojourn time in egg laying search bout}

\item{tO}{sojourn time in egg laying attempt bout}
}
\description{
From the parameters in the A,B,F1,F2,etc space, produce a named list of parameters
suitable for plugging into the MBDETES differential equation models (\code{\link{MBDETES_R2R_solve}} and \code{\link{MBDETES_cohort_solve}}).
}
