% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{MBDETES_SurvLaden}
\alias{MBDETES_SurvLaden}
\title{MBDETES: Probability of Surviving the post-prandial flight (laden mosquito)}
\usage{
MBDETES_SurvLaden()
}
\description{
MBDETES: Probability of Surviving the post-prandial flight (laden mosquito)
}
