% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Calibration.R
\name{MBDETES_SurvUnladen}
\alias{MBDETES_SurvUnladen}
\title{MBDETES: Probability of Surviving the post-prandial flight (unladen mosquito)}
\usage{
MBDETES_SurvUnladen()
}
\description{
MBDETES: Probability of Surviving the post-prandial flight (unladen mosquito)
}
