% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BloodMeal.R
\name{MBITES-BloodMeal}
\alias{MBITES-BloodMeal}
\title{MBITES: Blood Meal}
\description{
During a human host encounter (\code{\link{mbites_humanEncounter}}) or a non-human host encounter
(\code{\link{mbites_zooEncounter}}), if the mosquito survives to blood feed, it takes a blood meal.
}
\details{
The blood meal sets the \code{bloodfed} flag to true, which resets \code{boutFail} flag to 0 at the end of
\code{\link{mbites_attempt_B}}.

The blood meal consists of:
\enumerate{
\item Sampling a blood meal size (\code{\link{mbites_rBloodMealSize}})
\item Checking for mortality from overfeeding (\code{\link{mbites_Overfeeding}})
\item Calculating energetics/maturation from blood (\code{\link{mbites_bloodEnergetics}})
\item Simulating oogenesis/egg production (\code{\link{oogenesis}})
}
}
