% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{MBITES-Energetics}
\alias{MBITES-Energetics}
\title{MBITES: Energetics}
\description{
MBITES: Energetics
}
\section{Blood and Sugar Energetics}{


Energetics handles mosquito energetics derived from blood and sugar meals. The main methods implemented are:
\itemize{
\item \code{\link{mbites_BloodEnergetics}} which is called upon taking a successful blood meal
\item \code{\link{mbites_sugarMeal}} which is called upon taking a successful sugar meal
\item \code{\link{mbites_energetics}} which is called during \code{\link{mbites_updateState}} and calculates energy expenditure from flight and potentially queues a sugar feeding bout.
}
}

