% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-HostEncounter.R
\name{MBITES-HostEncounter}
\alias{MBITES-HostEncounter}
\title{MBITES: Host Encounter}
\description{
MBITES: Host Encounter
}
\section{Human and non-human vertebrate host encounters}{


During a blood feeding attempt bout (\code{\link{mbites_boutB}}), if the mosquito successfully locates
a viable human or vertebrate blood host, it initiates a host encounter. A human host encounter (\code{\link{mbites_humanEncounter}})
consists of an attempt to approach, probe, and take a blood meal. A non-human host encounter (\code{\link{mbites_zooEncounter}}) consists
of an attempt to approach and blood feed only.

If the mosquito is able to blood feed, the series of functions to simulate that process will be called, see \code{\link{MBITES-BloodMeal}}.
During probing and feeding different functions for simulation of mosquito to human and human to mosquito transmission will be
called; see \code{\link{PathogenGeneric}} for what these functions must simulate.
}

