% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Oogenesis.R
\name{MBITES-Oogenesis}
\alias{MBITES-Oogenesis}
\alias{mbites_oogenesis1}
\alias{mbites_oogenesis2}
\title{MBITES: Oogenesis}
\usage{
mbites_oogenesis1()

mbites_oogenesis2()
}
\description{
In MBITES, oogenesis (egg production) occurs during
\code{\link{mbites_updateState}} if the mosquito is bloodfed. Related to
oogenesis there are two other primary behaviors that must be selected by the
user, namely, refeeding model choice and egg batch size model choice.
}
\section{Refeed Model Choice}{


Refeeding may occur if the mosquito is gravid yet requires more blood prior
to oviposition. There are two models, plus a null model for turning off
refeeding  behavior.
}

\section{Egg Batch Refeeding Model}{


In this model of refeeding (\code{\link{mbites_pReFeed_batch}}), refeeding
probability is a function of the egg batch size, as a proportion of the
maximum allowable batch size. This model is not compatible with the MBDETES
approximation to MBITES.
}

\section{Blood Meal Refeeding Model}{


In this model of refeeding (\code{\link{mbites_pReFeed_bm}}), refeeding
probability is a function of the proportion full the mosquito is (the blood
meal size). This model is compatible with the MBDETES approximation to
MBITES.
}

\section{Egg Batch Model Choice}{


There are two models of egg batch size.
}

\section{Gaussian-distributed Egg Batch Model}{


In this model of egg batch size (\code{\link{mbites_rBatchSizeNorm}}), the
egg batch size is a normally distributed random variable.
}

\section{Blood Meal Dependent Egg Batch Model}{


In this model of egg batch size (\code{\link{mbites_rBatchSizeBms}}), the
egg batch size is a function of the blood meal size.
}

\section{Oogenesis Model Choice}{


Oogenesis model should be selected in \code{\link{MBITES_Setup}} prior to
creating any objects. Oogenesis occurs during the post-prandial rest
following a blood meal when the mosquito's state is updated, see
\code{\link{mbites_updateState}}. Both models require the user to select a
refeeding model and a egg batch size model choice (MBITES does not check if
these selections are logically consistent, and care should be taken to
choose appropriate combinations of models).
}

