% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Oviposition.R
\name{MBITES-Oviposition}
\alias{MBITES-Oviposition}
\title{MBITES: Oviposition}
\description{
During the \code{\link{mbites_updateState}} if the mosquito is gravid and passes refeeding check
(see \code{\link{mbites_checkRefeed}}) the mosquito will enter an oviposition bout (\code{\link{mbites_boutO}}) on the next launch.
During oviposition the mosquito first chooses a aquatic habitat at the site and then lays eggs.
}
