% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Resting.R
\name{MBITES-Resting}
\alias{MBITES-Resting}
\title{MBITES: Post-bout Landing, House Entering, and Resting}
\description{
MBITES: Post-bout Landing, House Entering, and Resting
}
\section{House Entering & Resting Behavior}{

At the end of a search or attempt bout a mosquito is now inside of a \code{\link{Site}}
and will attempt to rest. The specific resting spot chosen will depend on the type of site the mosquito is at.
The mosquito can choose a resting spot given by the following table:
\itemize{
\item l: Leave the area
\item r: Reattempt Without Resting;
\item v: Rest on vegetation
\item w: Rest on the Outside wall of a structure
\item i: Rest on the Inside wall of a structure
}
}

