% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Setup.R
\name{MBITES-Setup}
\alias{MBITES-Setup}
\title{MBITES: Setup}
\description{
Setup routines should be called prior to making any actual objects (ie; before calling any initialization functions, such as \code{\link{MBITES_Initialize}}, \code{link{Tile_Initialize}}, etc).
The setup function is responsible for assigning concrete methods to proper interfaces in the \code{\link{Mosquito}}, \code{\link{Mosquito_Female}}, and \code{\link{Mosquito_Male}}
classes as well as setting up the time-to-event closures in the global \code{\link{MBITES_Parameters}} object needed to sample waiting times to next launch.
}
\section{MBITES-Timing}{

See \code{\link{MBITES_Setup_Timing}} to initialize time-to-event methods.
}

\section{MBITES-BloodMeal}{

See \code{\link{MBITES_Setup_BloodMeal}} to initialize blood meal methods.
}

\section{MBITES-Oogenesis}{

See \code{\link{MBITES_Setup_Oogenesis}} to initialize oogenesis (egg production) methods.
}

\section{MBITES-Energetics}{

See \code{\link{MBITES_Setup_Energetics}} to initialize energetics methods.
}

\section{MBITES-Oviposition}{

See \code{\link{MBITES_Setup_Oviposition}} to initialize oviposition methods.
}

\section{MBITES-Survival}{

See \code{\link{MBITES_Setup_Survival}} to initialize survival methods.
}

\section{PATHOGEN}{

See \code{\link{Pathogen}} for details on how to initialize a pathogen model before making objects.
}

