% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{mbites2mbdetes_BstateTransitions}
\alias{mbites2mbdetes_BstateTransitions}
\title{MBITES: Blood Feeding Attempt Bout State Transitions}
\usage{
mbites2mbdetes_BstateTransitions()
}
\description{
B -> F,B,R,D
}
