% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bout.R
\name{mbites_attempt_B}
\alias{mbites_attempt_B}
\title{MBITES: Blood Feeding Attempt Bout}
\usage{
mbites_attempt_B()
}
\description{
The blood feeding attempt bout has the following structure:
}
\details{
\itemize{
\item call \code{\link{mbites_chooseHost}} to choose a host:
\itemize{
\item if the host is human, simulate a human encounter (see \code{\link{mbites_humanEncounter}})
\item if the host is not human, simulate a zoonotic encounter (see \code{\link{mbites_zooEncounter}})
\item if the mosquito chooses a trap, simulate the outcome
NOTE: a CDC light trap is one kind of trap
\item a null host is for a failed attempt
}
}

NOTE: host encounters are found in MBITES-HostEncounter.R
bloodtrap() is found in ...
}
