% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bout.R
\name{mbites_attempt_M}
\alias{mbites_attempt_M}
\title{M-BITES: Mating Bout (M) \code{MosquitoFemale}}
\usage{
mbites_attempt_M()
}
\description{
A mosquito performs mating bout.
}
\details{
\enumerate{
\item choose a mate from the swarmingQ ;
2a) mate
2b) the swarm has been sprayed.
2c) a null mate is for a failed attempt
}
\itemize{
\item This method is bound to \code{MosquitoFemale$boutM()}.
}
}
