% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bout.R
\name{mbites_attempt_O}
\alias{mbites_attempt_O}
\title{MBITES: Oviposition Attempt Bout}
\usage{
mbites_attempt_O()
}
\description{
The egg laying attempt bout has the following structure:
}
\details{
\enumerate{
\item choose a habitat;
2a) lay eggs in a habitat
2b) if a mosuqito chooses an ovitrap, simulate the outcome
2c) a null habitat is for a failed attempt
}
}
