% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bout.R
\name{mbites_attempt_S}
\alias{mbites_attempt_S}
\title{M-BITES: Sugar Feeding Bout (S) \code{MosquitoFemale}}
\usage{
mbites_attempt_S()
}
\description{
The sugar feeding attempt bout has the following structure:
}
\details{
\enumerate{
\item choose a sugar source;
2a) take a sugar meal
2b) if a mosuqito chooses an atsb, simulate the outcome
2c) a null source is for a failed attempt
}
\itemize{
\item This method is bound to \code{MosquitoFemale$boutS()}.
}
}
