% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bout.R
\name{mbites_attempt_search}
\alias{mbites_attempt_search}
\title{MBITES: Search Bout}
\usage{
mbites_attempt_search()
}
\description{
The generic search bout for a mosquito is called if the search flag is \code{TRUE}.
\enumerate{
\item call \code{\link{mbites_move}} to find a new site, this search can succeed or fail resulting in:
\itemize{
\item a mosquito leaves (a failure) to initiate a new search
\item a mosquito stays (a success) and attenpts to land
}
}
}
\details{
\enumerate{
\item If the search is successful, the search flag is set to \code{FALSE} and the mosquito samples an appropriate
resource from the current site it is at.
}
}
