% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Logging.R
\name{mbites_basicHistoryList}
\alias{mbites_basicHistoryList}
\title{MBITES: Basic History List}
\usage{
mbites_basicHistoryList()
}
\description{
Return a named list of the basic history object.
\itemize{
\item This method is bound to \code{Mosquito$basicHistoryList}
}
}
