% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbites_bloodEnergetics}
\alias{mbites_bloodEnergetics}
\title{MBITES: Blood Energetics}
\usage{
mbites_bloodEnergetics()
}
\description{
Add energy derived from blood to the mosquito's fuel tank. This method is called from \code{\link[MBITES]{mbites_BloodMeal}}.
\itemize{
\item This method is bound to \code{MosquitoFemale$bloodEnergetics}.
}
}
