% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bout.R
\name{mbites_BloodFeedingSearchCheck}
\alias{mbites_BloodFeedingSearchCheck}
\title{MBITES: Check for Blood Feeding Search Bout}
\usage{
mbites_BloodFeedingSearchCheck()
}
\description{
During the resting period \code{\link[MBITES]{mbites_updateState}}, check if the local site has a
blood hosts present.
\itemize{
\item this method is bound to \code{MosquitoFemale$OvipositSearchCheck}
}
}
