% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Logging.R
\name{mbites_bloodHistList}
\alias{mbites_bloodHistList}
\title{MBITES: Blood Feeding History List}
\usage{
mbites_bloodHistList()
}
\description{
Return a named list of the blood feeding history object.
\itemize{
\item This method is bound to \code{Mosquito_Female$bloodHistList}
}
}
