% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Oogenesis.R
\name{mbites_checkEggMaturation}
\alias{mbites_checkEggMaturation}
\title{MBITES: Check Egg Maturation}
\usage{
mbites_checkEggMaturation()
}
\description{
This function is called during \code{\link{mbites_checkRefeed}},
it checks that the mosquito has passed the egg maturation time and only sets \code{gravid = TRUE}
if this condition is filled. If the eggs are not mature, go on another blood search.
\itemize{
\item This method is bound to \code{Mosquito_Female$checkEggMaturation}
}
}
