% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Timing.R
\name{mbites_checkEstivation1}
\alias{mbites_checkEstivation1}
\title{MBITES: Probabilistic Estivation}
\usage{
mbites_checkEstivation1()
}
\description{
Queue estivation based on daily probability to enter estivation stage of the life cycle,
if estivation is queued, the next launch is set to be at random time in the future.
}
\details{
\itemize{
\item This method is bound to \code{Mosquito$checkEstivation}
}
}
