% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Timing.R
\name{mbites_checkEstivation2}
\alias{mbites_checkEstivation2}
\title{MBITES: Estivation based on Hard Cut-off}
\usage{
mbites_checkEstivation2()
}
\description{
The number estivationDay is a day of the year. This
method checks to see if tNow < estivationDay < tNext.
If so then, the mosquito estivates, which sets
tNext to a random number in the future, ttEstivate()
}
\details{
\itemize{
\item This method is bound to \code{Mosquito$checkEstivation}
}
}
