% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Timing.R
\name{mbites_checkEstivationNull}
\alias{mbites_checkEstivationNull}
\title{MBITES: Null Estivation}
\usage{
mbites_checkEstivationNull()
}
\description{
If estivation is turned off, do nothing.
\itemize{
\item This method is bound to \code{Mosquito$checkEstivation}
}
}
