% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Timing.R
\name{mbites_checkPostPrandial}
\alias{mbites_checkPostPrandial}
\title{M-BITES: Simulates the post-prandial resting period for \code{MosquitoFemale}}
\usage{
mbites_checkPostPrandial()
}
\description{
Method checks to see if the mosquito has bloodfed and is in a
post prandial state; if so, it resets tNow and tNext to
tNow + ttEvent_ppRest()
}
\details{
\itemize{
\item This method is bound to \code{MosquitoFemale$timing()}.
}
}
