% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Oogenesis.R
\name{mbites_checkRefeed}
\alias{mbites_checkRefeed}
\title{MBITES: Check Refeeding Behavior}
\usage{
mbites_checkRefeed()
}
\description{
During \code{\link{mbites_updateState}}, if the mosquito is gravid, check for refeeding
behavior as a function of the batch size. This function should \emph{only} be used with the
first model of oogenesis.
\itemize{
\item This method is bound to \code{Mosquito_Female$checkRefeed}
}
}
