% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Oogenesis.R
\name{mbites_checkRefeed_null}
\alias{mbites_checkRefeed_null}
\title{MBITES: Null Refeeding Behavior}
\usage{
mbites_checkRefeed_null()
}
\description{
If using the second model of oogenesis, refeeding should be disabled by using this stand-in function.
\itemize{
\item This method is bound to \code{Mosquito_Female$checkRefeed}
}
}
