% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Oviposition.R
\name{mbites_chooseHabitat}
\alias{mbites_chooseHabitat}
\title{MBITES: Choose an Aquatic Habitat}
\usage{
mbites_chooseHabitat()
}
\description{
Choose a \code{\link{Aqua_Resource}} for oviposition; if ovitraps, etc exist at this \code{\link{Site}}, those
pseudo-resources compete for the mosquito's attention.
\itemize{
\item this method is bound to \code{Mosquito_Female$chooseHabitat}
}
}
