% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-HostEncounter.R
\name{mbites_chooseHost}
\alias{mbites_chooseHost}
\title{MBITES: Choose a Blood Host}
\usage{
mbites_chooseHost()
}
\description{
Sample a blood host from the \code{\link{make_RiskQ}} closure this mosquito is at.
\itemize{
\item This method is bound to \code{Mosquito_Female$chooseHost}.
}
}
