% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Mating.R
\name{mbites_chooseMate}
\alias{mbites_chooseMate}
\title{MBITES: Choose a Mate}
\usage{
mbites_chooseMate()
}
\description{
During the mating bout, a female mosquito chooses a mate from the mating queue closure
at the mating site, see \code{\link{make_MatingQ}} for details on the mating queue.
}
\details{
\itemize{
\item This method is bound to \code{Mosquito_Female$chooseMate}.
}
}
