% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbites_chooseSugarSource}
\alias{mbites_chooseSugarSource}
\title{MBITES: Choose Sugar Source}
\usage{
mbites_chooseSugarSource()
}
\description{
During a sugar bout the mosquito samples from available sugar feeding resources
at the \code{\link{Site}} it is at.
\itemize{
\item this method is bound to \code{Mosquito$chooseSugarSource}
}
}
