% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Logging.R
\name{mbites_eggHistList}
\alias{mbites_eggHistList}
\title{MBITES: Oviposition History List}
\usage{
mbites_eggHistList()
}
\description{
Return a named list of the oviposition history object.
\itemize{
\item This method is bound to \code{Mosquito_Female$eggHistList}
}
}
