% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Mosquito.R
\name{mbites_exit}
\alias{mbites_exit}
\title{MBITES: Export History and Remove Self}
\usage{
mbites_exit()
}
\arguments{
\item{force}{if \code{TRUE} the mosquito will write out history and be removed from the container even if still alive}
}
\description{
If the mosquito is dead, write out its history to a JSON-formatted file and then delete from the container object (\code{\link{HashMap}}).
\itemize{
\item This method is bound to \code{Mosquito_Female$exit}
}
}
