% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Logging.R
\name{mbites_exit_Mosquito_Female}
\alias{mbites_exit_Mosquito_Female}
\title{MBITES: Export History and Remove Self (Female)}
\usage{
mbites_exit_Mosquito_Female(endSim = FALSE)
}
\arguments{
\item{pretty}{prettify JSON output}
}
\description{
If the mosquito is dead, write out its history to a JSON-formatted file and then delete from the container object (\code{\link{HashMap}}).
\itemize{
\item This method is bound to \code{Mosquito_Female$exit}
}
}
