% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Timing.R
\name{mbites_findSwarm}
\alias{mbites_findSwarm}
\title{MBITES: Check Mating Event}
\usage{
mbites_findSwarm()
}
\description{
If timing of next launch occurs during a period when mating swarms are emerging, go into mating bout
if immature.
\itemize{
\item This method is bound to \code{Mosquito_Female$timing}.
}
}
