% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Timing.R
\name{mbites_findSwarmNull}
\alias{mbites_findSwarmNull}
\title{MBITES: Null Check Mating Event}
\usage{
mbites_findSwarmNull()
}
\description{
If mating is turned off, do nothing.
\itemize{
\item This method is bound to \code{Mosquito_Female$findSwarm}.
}
}
