% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbites_flightBurnEnergy}
\alias{mbites_flightBurnEnergy}
\title{MBITES: Energy burn from flight}
\usage{
mbites_flightBurnEnergy()
}
\description{
write me!
\itemize{
\item This method is bound to \code{Mosquito$flightBurnEnergy()}.
}
}
