% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Survival.R
\name{mbites_get_surviveFlightBaseProb}
\alias{mbites_get_surviveFlightBaseProb}
\title{MBITES: Get Baseline Flight Survival Probability}
\usage{
mbites_get_surviveFlightBaseProb()
}
\description{
Get baseline survival probability for \code{\link{mbites_surviveFlight}}.
\itemize{
\item This method is bound to \code{Mosquito$get_surviveFlightProb}.
}
}
