% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Initialize.R
\name{MBITES_Initialize}
\alias{MBITES_Initialize}
\title{MBITES Initialize Mosquito Populations}
\usage{
MBITES_Initialize(mosy_init)
}
\arguments{
\item{mosy_init}{a \code{\link{data.frame}} object where each row corresponds to a mosquito}
}
\description{
Initialize mosquito populations
}
\details{
write some details
}
