% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Oviposition.R
\name{mbites_layEggs_Emerge}
\alias{mbites_layEggs_Emerge}
\title{M-BITES: Lay Eggs for 'Emerge' \code{\link{MosquitoFemale}}}
\usage{
mbites_layEggs_Emerge()
}
\description{
During an oviposition bout \code{\link{mbites_boutO}}, lay eggs by placing
them into the \code{EggQ} of the aquatic site.
This method is bound to \code{MosquitoFemale$layEggs()}.
}
