% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Mating.R
\name{mbites_mating}
\alias{mbites_mating}
\title{MBITES: Mating}
\usage{
mbites_mating()
}
\description{
If the female mosquito successfully found a mate, set \code{mated = TRUE} and check if the mosquito has fulfilled
the pre-gonotrophic energy requirement, and set \code{mature = TRUE} if sufficient energy has been acquired.
}
\details{
\itemize{
\item This method is bound to \code{Mosquito_Female$mating}.
}
}
