% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bout.R
\name{mbites_MBITES}
\alias{mbites_MBITES}
\title{MBITES: Run Simulation for one Mosquito}
\usage{
mbites_MBITES()
}
\description{
Run the MBITES simulation by repeatedly calling \code{\link{mbites_oneBout}} for this mosquito while it is alive and has not overrun
the global time.
\itemize{
\item This method is bound to \code{MosquitoFemale$MBITES}
}
}
