% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bout.R
\name{mbites_move}
\alias{mbites_move}
\title{MBITES: Move}
\usage{
mbites_move()
}
\description{
If successful, the mosquito moves to a new \code{\link[MBITES]{Site}} object from querying
the current site by \code{\link[MBITES]{move_mosquito_Site}}. This method is called from \code{\link[MBITES]{mbites_searchBout}}
* binding: \code{Mosquito$move}
}
