% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bout.R
\name{mbites_oneBout}
\alias{mbites_oneBout}
\title{MBITES: One Bout \code{MosquitoFemale}}
\usage{
mbites_oneBout()
}
\description{
Mosquito behavior has a finite set of states (state space of model), within which there are certain biological functions that are always evaluated.
A bout is the actions taken by a mosquito between a launch and landing; \code{mbites_oneBout} handles all the biological imperatives that occur during a bout,
while specialized bout action methods handle the events that occur due to the intent of the mosquito during the bout. The two classes of bouts are SEARCH and ATTEMPT.
Search bouts are similar enough that there is a generic version of it.
\itemize{
\item \code{\link{mbites_boutB}}: blood feeding attempt bout
\item \code{\link{mbites_boutO}}: egg laying attempt bout
\item \code{\link{mbites_boutS}}: sugar feeding attempt bout
\item \code{\link{mbites_boutM}}: mating attempt bout
\item \code{\link{mbites_boutSearch}}: search for a new \code{\link{Site}}
}
}
\details{
The generic bout runs necessary updates of timing, state, survival, energetics, and queue checks prior to calling the nested
specific bout action, and checks that the mosquito is alive/active before calling the bout. It updates \code{tNext} and \code{state}.

This corresponds to the following Gillespie-style algorithm:
\enumerate{
\item tNow is set to tNext from previous bout
\item Launch: an attempt or search bout
\item Land: restingSpot is called to find a microsite for resting
\item Rest: update the mosquito's state
\item Log events
}
\itemize{
\item This method is bound to \code{MosquitoFemale$oneBout()}.
}
}
