% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BloodMeal.R
\name{mbites_Overfeeding}
\alias{mbites_Overfeeding}
\title{MBITES: Overfeeding Mortality}
\usage{
mbites_Overfeeding()
}
\description{
Calculate overfeeding mortality (see \code{\link{mbites_pOverFeed}}).
\itemize{
\item This method is bound to \code{MosquitoFemale$Overfeeding}.
}
}
