% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BloodMeal.R
\name{mbites_OverfeedingNull}
\alias{mbites_OverfeedingNull}
\title{MBITES: Null Overfeeding}
\usage{
mbites_OverfeedingNull()
}
\description{
If overfeeding mortality turned off, do nothing
\itemize{
\item This method is bound to \code{MosquitoFemale$Overfeeding}.
}
}
