% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bout.R
\name{mbites_OvipositSearchCheck}
\alias{mbites_OvipositSearchCheck}
\title{MBITES: Check for Oviposit Search Bout}
\usage{
mbites_OvipositSearchCheck()
}
\description{
During the resting period \code{\link[MBITES]{mbites_updateState}}, check if the local site has an
aquatic habitat present for oviposition.
\itemize{
\item this method is bound to \code{MosquitoFemale$OvipositSearchCheck}
}
}
