% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Survival.R
\name{mbites_pChem}
\alias{mbites_pChem}
\title{MBITES-Generic: Probability of Death due to Chemical Damage for \code{\link{MosquitoFemale}}}
\usage{
mbites_pChem()
}
\description{
probability of death due to tattering given by \deqn{
\frac{2+chm.b}{1+chm.b} - \frac{e^{damage_physical\times chm.a}}{chm.b+e^{damage_chemical\times chm.a}} }
\itemize{
\item This method is bound to \code{MosquitoFemale$pTatter()}.
}
}
