% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Survival.R
\name{mbites_pEnergySurvival}
\alias{mbites_pEnergySurvival}
\title{MBITES: Probability of Death due to Energy Reserves}
\usage{
mbites_pEnergySurvival()
}
\description{
Incremental mortality as a function of energy reserves given by \eqn{ \frac{e^{S.a\times energy}}{S.b+e^{S.a\times energy}} }
\itemize{
\item This method is bound to \code{Mosquito$pEnergySurvival}.
}
}
