% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Oogenesis.R
\name{mbites_PPRbmSize1}
\alias{mbites_PPRbmSize1}
\title{MBITES: Oogenesis Model 1 Blood Meal Size Reset}
\usage{
mbites_PPRbmSize1()
}
\description{
In the first model of oogenesis blood meal size is reset to 0 during the
post-prandial rest (\code{\link{mbites_checkPostPrandial}}).
\itemize{
\item This method is bound to \code{Mosquito_Female$PPRbmSize}
}
}
