% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Oogenesis.R
\name{mbites_PPRbmSize2}
\alias{mbites_PPRbmSize2}
\title{MBITES: Oogenesis Model 2 Blood Meal Size Reset}
\usage{
mbites_PPRbmSize2()
}
\description{
In the second model of oogenesis blood meal size is reduced by an amount equal to
the amount of blood used in the egg provision, or to 0 if that would induce a negative amount.
This occurs during the post-prandial rest (\code{\link{mbites_checkPostPrandial}}).
\itemize{
\item This method is bound to \code{Mosquito_Female$PPRbmSize}
}
}
