% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BloodMeal.R
\name{mbites_PPRFlight}
\alias{mbites_PPRFlight}
\title{MBITES: Post Prandial Flight and Rest}
\usage{
mbites_PPRFlight()
}
\description{
If the mosquito did not die from overfeeding, update time according to the time needed by PPR (see \code{\link{set_ttEvent_ppr_MBITES_Parameters}}).
Mosquito dies as a result of the post-prandial flight with probability given by \code{\link{mbites_pPPRFlight}}.
\itemize{
\item This method is bound to \code{Mosquito_Female$PPRFlight}.
}
}
