% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Oogenesis.R
\name{mbites_pReFeed}
\alias{mbites_pReFeed}
\title{MBITES: Probability of Refeeding}
\usage{
mbites_pReFeed()
}
\description{
Probability to re-enter blood feeding cycle after incomplete blood feeding given by \eqn{ \frac{2+rf.b}{1+rf.b}-\frac{e^{rf.a\times batch}}{rf.b+e^{rf.a\times batch}} }
\itemize{
\item This method is bound to \code{MosquitoFemale$pReFeed()}.
}
}
