% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Oogenesis.R
\name{mbites_pReFeed_bm}
\alias{mbites_pReFeed_bm}
\title{MBITES: Probability of Refeeding as Function of Blood Meal Size}
\usage{
mbites_pReFeed_bm()
}
\description{
Probability to re-enter blood feeding cycle after incomplete blood feeding given by \eqn{ \frac{2+rf_{b}}{1+rf_{b}}-\frac{e^{rf_{a}\times bmSize}}{rf_{b}+e^{rf_{a}\times bmSize}} }
This models mosquito propensity to take more blood if the previous blood meal was too small.
\itemize{
\item This method is bound to \code{MosquitoFemale$pReFeed()}
}
}
