% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Oogenesis.R
\name{mbites_pReFeed_null}
\alias{mbites_pReFeed_null}
\title{MBITES: Null Probability of Refeeding}
\usage{
mbites_pReFeed_null()
}
\description{
Null probability of refeeding (turn the behavior off)
\itemize{
\item This method is bound to \code{MosquitoFemale$pReFeed()}.
}
}
