% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Timing.R
\name{mbites_prEstivate}
\alias{mbites_prEstivate}
\title{MBITES: Daily Probability of Estivation}
\usage{
mbites_prEstivate()
}
\description{
Get the daily probability of entering estivation stage of life cycle.
\itemize{
\item This method is bound to \code{Mosquito$prEstivate}
}
}
