% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbites_pSugarBout}
\alias{mbites_pSugarBout}
\title{MBITES: Probability of a sugar bout}
\usage{
mbites_pSugarBout()
}
\description{
Calculate the probability to queue a sugar bout as a function of the mosquito's current energy levels.
The probability is given by \eqn{\frac{2+S_{sb}}{1+S_{sb}} - \frac{e^{S_{sa} \times energy}}{S_{sb}+e^{S_{sa}\times energy}}}
\itemize{
\item This method is bound to \code{Mosquito$pSugarBout}
}
}
