% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbites_queueSugarBout}
\alias{mbites_queueSugarBout}
\title{MBITES: Queue a sugar bout}
\usage{
mbites_queueSugarBout()
}
\description{
If the mosquito queues a sugar bout (with probability \code{\link{mbites_pSugarBout}}), set \code{starved} flag true.
\itemize{
\item This method is bound to \code{Mosquito$queueSugarBout}
}
}
