% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbites_queueSugarBout_null}
\alias{mbites_queueSugarBout_null}
\title{MBITES: Null sugar bout}
\usage{
mbites_queueSugarBout_null()
}
\description{
Null version of \code{\link{mbites_queueSugarBout}} if sugar feeding behavior is to be turned off.
\itemize{
\item This method is bound to \code{Mosquito$queueSugarBout}
}
}
