% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Oogenesis.R
\name{mbites_rBatchSizeNorm}
\alias{mbites_rBatchSizeNorm}
\title{MBITES-Generic: Draw Normally-distributed Egg Batch Size for \code{\link{MosquitoFemale}}}
\usage{
mbites_rBatchSizeNorm()
}
\description{
Draw a egg batch size from Normal(bs.m,bs.v)
\itemize{
\item This method is bound to \code{MosquitoFemale$rBatchSize()}.
}
}
