% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Oogenesis.R
\name{mbites_rEggMaturationTimeNorm}
\alias{mbites_rEggMaturationTimeNorm}
\title{MBITES: Normally-distributed Egg Maturation Time}
\usage{
mbites_rEggMaturationTimeNorm()
}
\description{
Draw an egg maturation time from Normal(emt_m,emt_sd)
\itemize{
\item This method is bound to \code{Mosquito_Female$rEggMaturationTime}.
}
}
