% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Survival.R
\name{mbites_Senescence}
\alias{mbites_Senescence}
\title{MBITES: Probability of Death due to Senescence}
\usage{
mbites_Senescence(p)
}
\description{
To simulate senescence-dependent mortality, call \code{\link{mbites_pSenesce}} to modify the passed in probability of survival.
\itemize{
\item This method is bound to \code{Mosquito$Senescence}
}
}
