% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Survival.R
\name{mbites_Senescence_null}
\alias{mbites_Senescence_null}
\title{MBITES: Null Senescence}
\usage{
mbites_Senescence_null(p)
}
\description{
The non-null version is \code{\link{Senescence}}
\itemize{
\item This method is bound to \code{Mosquito$Senescence}
}
}
