% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Setup.R
\name{MBITES_Setup}
\alias{MBITES_Setup}
\title{MBITES: Setup Model}
\usage{
MBITES_Setup(timing_model = 1L, rate_b = NULL, tmin_b = NULL,
  rate_o = NULL, tmin_o = NULL, rate_m = NULL, tmin_m = NULL,
  rate_s = NULL, tmin_s = NULL, rate_bs = NULL, tmin_bs = NULL,
  rate_os = NULL, tmin_os = NULL, rate_ms = NULL, tmin_ms = NULL,
  rate_ss = NULL, tmin_ss = NULL, mean_b = NULL, cv_b = NULL,
  mean_o = NULL, cv_o = NULL, mean_m = NULL, cv_m = NULL,
  mean_s = NULL, cv_s = NULL, mean_bs = NULL, cv_bs = NULL,
  mean_os = NULL, cv_os = NULL, mean_ms = NULL, cv_ms = NULL,
  mean_ss = NULL, cv_ss = NULL, oogenesis_model = 1L)
}
\description{
See \code{\link{MBITES Setup}} for more details. Note that because of lazy evaluation of function arguments only arguments that are
needed for the specific models chosen need to be provided as input by the user.
}
\details{
Arguments for the setup function are listed below:
\itemize{
\item timing_model: integer flag for time-to-event distribution. All timing models require further arguments which are given below their main heading.
\itemize{
\item 1: shifted exponential distribution, see \code{\link{make_ttEvent_Exp}} for more details (if selecting this model please enter required arguments below)
\itemize{
\item rate_b: inverse of average waiting time to next launch for blood feeding attempt bout
\item tmin_b: minimum waiting time prior to next launch for blood feeding attempt bout
\item rate_o: inverse of average waiting time to next launch for oviposition attempt bout
\item tmin_o: minimum waiting time prior to next launch for oviposition attempt bout
\item rate_m: inverse of average waiting time to next launch for mating attempt bout
\item tmin_m: minimum waiting time prior to next launch for mating attempt bout
\item rate_s: inverse of average waiting time to next launch for sugar feeding attempt bout
\item tmin_s: minimum waiting time prior to next launch for sugar feeding attempt bout
\item rate_bs: inverse of average waiting time to next launch for blood feeding search bout
\item tmin_bs: minimum waiting time prior to next launch for blood feeding search bout
\item rate_os: inverse of average waiting time to next launch for oviposition search bout
\item tmin_os: minimum waiting time prior to next launch for oviposition search bout
\item rate_ms: inverse of average waiting time to next launch for mating search bout
\item tmin_ms: minimum waiting time prior to next launch for mating search bout
\item rate_ss: inverse of average waiting time to next launch for sugar feeding search bout
\item tmin_ss: minimum waiting time prior to next launch for sugar feeding search bout
}
\item 2: shifted gamma distribution, note that minimum waiting time arguments still need to be provided, as given above, see \code{\link{make_ttEvent_Gamma}} for more details (if selecting this model please enter required arguments below)
\itemize{
\item mean_b: inverse of average waiting time to next launch for blood feeding attempt bout
\item cv_b: coefficient of variation between mean and variance of waiting time
\item mean_o: inverse of average waiting time to next launch for oviposition attempt bout
\item cv_o: coefficient of variation between mean and variance of waiting time
\item mean_m: inverse of average waiting time to next launch for mating attempt bout
\item cv_m: coefficient of variation between mean and variance of waiting time
\item mean_s: inverse of average waiting time to next launch for sugar feeding attempt bout
\item cv_s: coefficient of variation between mean and variance of waiting time
\item mean_bs: inverse of average waiting time to next launch for blood feeding search bout
\item cv_bs: coefficient of variation between mean and variance of waiting time
\item mean_os: inverse of average waiting time to next launch for oviposition search bout
\item cv_os: coefficient of variation between mean and variance of waiting time
\item mean_ms: inverse of average waiting time to next launch for mating search bout
\item cv_ms: coefficient of variation between mean and variance of waiting time
\item mean_ss: inverse of average waiting time to next launch for sugar feeding search bout
\item cv_ss: coefficient of variation between mean and variance of waiting time
}
\item 3: shifted diurnal (sinusoidal) distribution, see \code{\link{make_ttEvent_Diurnal}} (if selecting this model please enter required arguments below)
\itemize{
\item not implemented yet
}
}
}
}
