% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Setup.R
\name{MBITES_Setup_BloodMeal}
\alias{MBITES_Setup_BloodMeal}
\title{MBITES: Setup BloodMeal}
\usage{
MBITES_Setup_BloodMeal(overfeeding)
}
\description{
This function sets up items related to the blood meal for MBITES (see \code{\link{MBITES-BloodMeal}} for more details).
It enables or disables overfeeding behavior (mortality due to the blood meal).
}
\details{
See \code{\link{MBITES-Setup}} for more details. Note that because of lazy evaluation of function arguments only arguments that are
needed for the specific models chosen need to be provided as input by the user.

Arguments for the setup function are listed below:

\strong{MBITES-BloodMeal}: see \code{\link{MBITES-BloodMeal}}
\itemize{
\item bm_a: alpha parameter of beta-distributed blood meal size
\item bm_b: beta parameter of beta-distributed blood meal size
\item overfeeding: logical flag for overfeeding behavior (increased mortality from blood feeding)
\itemize{
\item of_a: parameter in \code{\link{mbites_pOverFeed}} (probability of death from blood meal size)
\item of_b: parameter in \code{\link{mbites_pOverFeed}} (probability of death from blood meal size)
}
}
}
