% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Setup.R
\name{MBITES_Setup_Energetics}
\alias{MBITES_Setup_Energetics}
\title{MBITES: Setup Energetics}
\usage{
MBITES_Setup_Energetics(sugar)
}
\description{
This function sets up items related to energetics for MBITES (see \code{\link{MBITES-Energetics}} for more details).
}
\details{
See \code{\link{MBITES-Setup}} for more details. Note that because of lazy evaluation of function arguments only arguments that are
needed for the specific models chosen need to be provided as input by the user.

Arguments for the setup function are listed below:

\strong{MBITES-Energetics}: see \code{\link{MBITES-Energetics}}
\itemize{
\item sugar: boolean flag to turn sugar feeding behavior on or off (please note if sugar feeding is off, other parameters should be adjusted to compensate for increased death rate as a result of no energy intake from sugar)
}
}
