% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Setup.R
\name{MBITES_Setup_MBDETES}
\alias{MBITES_Setup_MBDETES}
\title{MBITES: Setup Parameters & Model Options for MBDETES Approximation}
\usage{
MBITES_Setup_MBDETES()
}
\description{
This function sets up MBITES model parameters and model behavioral options to the reduced
MBITES model to which MBDETES is able to approximate.
}
