% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Setup.R
\name{MBITES_Setup_Oogenesis}
\alias{MBITES_Setup_Oogenesis}
\title{MBITES: Setup Oogenesis}
\usage{
MBITES_Setup_Oogenesis(oogenesis_model, eggMaturationTime, eggsize_model,
  refeeding)
}
\description{
This function sets up items related to oogenesis for MBITES (see \code{\link{MBITES-Oogenesis}} for more details).
}
\details{
See \code{\link{MBITES-Setup}} for more details. Note that because of lazy evaluation of function arguments only arguments that are
needed for the specific models chosen need to be provided as input by the user.

Arguments for the setup function are listed below:

\strong{MBITES-Oogenesis}: see \code{\link{MBITES-Oogenesis}}
\itemize{
\item oogenesis_model: integer flag for model of oogenesis (egg production & development)
\itemize{
\item 1: egg batch size proportional to blood meal size (see help file for \code{MBITES-Oogenesis})
\itemize{
\item eggMaturationTime: logical flag, if \code{TRUE} use \code{\link{mbites_rEggMaturationTimeNorm}} to sample egg maturation times, if \code{FALSE} use the null filler \code{\link{mbites_rEggMaturationTimeOff}}
\itemize{
\item emt_m: mean of Gaussian distributed egg maturation time
\item emt_sd: standard deviation of Gaussian distributed egg maturation time
}
}
\item 2: egg batch size commits to development (see help file for \code{MBITES-Oogenesis})
\itemize{
\item bloodPerEgg: amount of blood needed per egg
}
}
\item eggsize_model: integer flag for model of egg batch size
\itemize{
\item 1: sample Gaussian-distributed egg batch size (see \code{\link{mbites_rBatchSizeNorm}})
\itemize{
\item bs_m: mean of Gaussian distribution
\item bs_sd: standard deviation of Gaussian distribution
}
\item 2: egg batch size is function of blood meal size (see \code{\link{mbites_rBatchSizeBms}})
\itemize{
\item maxBatch: maximum possible size of an egg batch
}
}
\item refeeding: integer flag for model of refeeding behavior (this should only be used with the first model of oogenesis, see \code{\link{mbites_checkRefeed}} for details)
\itemize{
\item 1: probability to refeed is a function of egg batch size (see \code{\link{mbites_pReFeed_batch}})
\item 2: probability to refeed is a function of blood meal size (see \code{\link{mbites_pReFeed_bm}})
\item 3: turn refeeding behavior off
}
}
}
