% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Setup.R
\name{MBITES_Setup_Oviposition}
\alias{MBITES_Setup_Oviposition}
\title{MBITES: Setup Oviposition}
\usage{
MBITES_Setup_Oviposition(aqua_model)
}
\description{
This function sets up items related to oviposition for MBITES (see \code{\link{MBITES-Oviposition}} for more details).
}
\details{
See \code{\link{MBITES-Setup}} for more details. Note that because of lazy evaluation of function arguments only arguments that are
needed for the specific models chosen need to be provided as input by the user.

Arguments for the setup function are listed below:

\strong{MBITES-Oviposition}: see \code{\link{MBITES-Oviposition}}
\itemize{
\item aqua_model: integer flag for model of aquatic ecology
\itemize{
\item 1: emerge: use the emerge model of aquatic ecology
\item 2: EL4P: use the EL4P (eggs, larvae, pupae) model of aquatic ecology
}
}
}
