% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Setup.R
\name{MBITES_Setup_Survival}
\alias{MBITES_Setup_Survival}
\title{MBITES: Setup Survival}
\usage{
MBITES_Setup_Survival(tattering, senescence)
}
\description{
This function sets up items related to survival for MBITES (see \code{\link{MBITES-Survival}} for more details).
}
\details{
See \code{\link{MBITES-Setup}} for more details. Note that because of lazy evaluation of function arguments only arguments that are
needed for the specific models chosen need to be provided as input by the user.

Arguments for the setup function are listed below:

\strong{MBITES-Survival}: see \code{\link{MBITES-Survival}}
\itemize{
\item tattering: boolean flag to turn wing tattering derived contribution to mortality on or off
\item senescence: boolean flag to turn senescence derived contribution to mortality on or off
}
}
