% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbites_sugarMeal}
\alias{mbites_sugarMeal}
\title{MBITES: Sugar Energetics}
\usage{
mbites_sugarMeal()
}
\description{
Add energy derived from sugar to the mosquito's fuel tank, and if the mosquito is immature, subtract energy
derived from this sugar meal from the pre-gonotrophic energy requirement. If the mosquito is mated
and has passed the energy requirement, set \code{mature = TRUE}.
\itemize{
\item This method is bound to \code{Mosquito_Female$sugarMeal}
}
}
