% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bout.R
\name{mbites_SugarSearchCheck}
\alias{mbites_SugarSearchCheck}
\title{MBITES: Check for Sugar Search Bout}
\usage{
mbites_SugarSearchCheck()
}
\description{
During the resting period \code{\link{mbites_updateState}}, check if the local site has a
sugar feeding resource present.
\itemize{
\item this method is bound to \code{MosquitoFemale$SugarSearchCheck}
}
}
