% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Survival.R
\name{mbites_survival}
\alias{mbites_survival}
\title{MBITES: Per-bout Survival}
\usage{
mbites_survival()
}
\description{
Simulate flight survival \code{\link{mbites_surviveFlight}} and survival of local hazards \code{\link{mbites_surviveHazards}}.
\itemize{
\item This method is bound to \code{Mosquito$survival}
}
}
