% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Survival.R
\name{mbites_surviveFlight}
\alias{mbites_surviveFlight}
\title{MBITES-Generic: Flight Survival for \code{\link{MosquitoFemale}}}
\usage{
mbites_surviveFlight()
}
\description{
Run generic flight survival probailities for bouts (launch to launch).
Depending on settings from M-BITES parameters, senescence and/or tattering may also be simulated.
\itemize{
\item This method is bound to \code{MosquitoFemale$surviveFlight()}.
}
}
