% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Survival.R
\name{mbites_surviveHazards}
\alias{mbites_surviveHazards}
\title{MBITES: Survive Local Hazards}
\usage{
mbites_surviveHazards()
}
\description{
Run generic local hazard survival probailities for bouts (launch to launch).
\itemize{
\item This method is bound to \code{Mosquito$surviveHazards}
}
}
