% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Timing.R
\name{mbites_timing}
\alias{mbites_timing}
\title{MBITES: Time to Launch Sampling}
\usage{
mbites_timing()
}
\description{
Sample my time of next launch.
\itemize{
\item This method is bound to \code{Mosquito$timing}.
}
}
