% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Timing.R
\name{mbites_timing_fn}
\alias{mbites_timing_fn}
\title{MBITES: Time to Launch (Function ver. for death tte)}
\usage{
mbites_timing_fn(state, search, tt)
}
\description{
MBITES: Time to Launch (Function ver. for death tte)
}
