% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bout.R
\name{mbites_updateState}
\alias{mbites_updateState}
\title{MBITES: Update the Behavioral State at the End of a Bout for \code{\link{MosquitoFemale}}}
\usage{
mbites_updateState()
}
\description{
After landing, during the resting period, the mosquito's
behavioral state and other state variables are updated.
}
\details{
M-BITES checks the state variables during the resting
period to determine what the next state will be. There is
a natural hierarchy: a dead mosquito can never be revived;
a starved mosquito will seek sugar; a gravid mosquito will
tend to lay eggs, though it might decide to top up with
blood; if nothing else, a mosquito will seek blood.
}
