% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Timing.R
\name{mbites_wakeUpTime}
\alias{mbites_wakeUpTime}
\title{MBITES: Random Wake-up Time}
\usage{
mbites_wakeUpTime()
}
\description{
Randomly sample a time to wake-up from estivation.
\itemize{
\item This method is bound to \code{Mosquito$wakeUpTime}
}
}
