% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Survival.R
\name{mbites_WingTattering}
\alias{mbites_WingTattering}
\title{MBITES-Generic: Mean Wing Tattering for \code{\link{MosquitoFemale}}}
\usage{
mbites_WingTattering()
}
\description{
Draw from a zero-inflated Beta distribution for additive wing damage_physical from tattering.
Wing damage_physical is given by \deqn{ \left\{\begin{matrix}
x=0; P(ttsz.p)
\\
x\sim Beta(ttsz.a,ttsz.b); P(1-ttsz.p)
\end{matrix}\right. }
\itemize{
\item This method is bound to \code{MosquitoFemale$WingTattering()}.
}
}
