% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Survival.R
\name{mbites_WingTattering_null}
\alias{mbites_WingTattering_null}
\title{MBITES: Null wing tattering}
\usage{
mbites_WingTattering_null(p)
}
\description{
If wing tattering is not being simulated do not modify \code{damage_physical} attribute of the mosquito and return the unmodified probabiltiy.
The non-null version is \code{\link{mbites_WingTattering}}.
\itemize{
\item This method is bound to \code{Mosquito$WingTattering}
}
}
