% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-HostEncounter.R
\name{mbites_zooEncounter}
\alias{mbites_zooEncounter}
\title{MBITES: Non-Human Host Encounter}
\usage{
mbites_zooEncounter()
}
\description{
After calling \code{\link{mbites_chooseHost}}, the mosquito encounters the non-human host and attempts to feed.
The encounter process consists of several stages:
\enumerate{
\item Check initial survival to see if the mosquito survives to feed
\item Check if the mosquito is able to successfully begin blood feeding; take a blood meal via \code{\link{mbites_BloodMeal}}
}
}
\details{
\itemize{
\item This method is bound to \code{Mosquito_Female$zooEncounter}.
}
}
