% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Mosquito.R
\docType{class}
\name{Mosquito_Male}
\alias{Mosquito_Male}
\title{MBITES: Male Mosquito Class}
\format{
An \code{\link{R6Class}} generator object
}
\description{
Male mosquitoes inherit from the \code{\link{Mosquito}} abstract base class object.
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item method: im a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item id: integer id (obtained from \code{\link{MBITES_Globals}})
\item field: im a field!
}
}

\keyword{R6}
\keyword{class}
\section{Super class}{
\code{\link[MBITES:Mosquito]{MBITES::Mosquito}} -> \code{Mosquito_Male}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-exit}{\code{Mosquito_Male$exit()}}
\item \href{#method-new}{\code{Mosquito_Male$new()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="BloodEnergetics">}\href{../../MBITES/html/Mosquito.html#method-BloodEnergetics}{\code{MBITES::Mosquito$BloodEnergetics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="BloodFeedingSearchCheck">}\href{../../MBITES/html/Mosquito.html#method-BloodFeedingSearchCheck}{\code{MBITES::Mosquito$BloodFeedingSearchCheck()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="MBITES">}\href{../../MBITES/html/Mosquito.html#method-MBITES}{\code{MBITES::Mosquito$MBITES()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="MatingSearchCheck">}\href{../../MBITES/html/Mosquito.html#method-MatingSearchCheck}{\code{MBITES::Mosquito$MatingSearchCheck()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="OvipositSearchCheck">}\href{../../MBITES/html/Mosquito.html#method-OvipositSearchCheck}{\code{MBITES::Mosquito$OvipositSearchCheck()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="SugarSearchCheck">}\href{../../MBITES/html/Mosquito.html#method-SugarSearchCheck}{\code{MBITES::Mosquito$SugarSearchCheck()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="attempt_B">}\href{../../MBITES/html/Mosquito.html#method-attempt_B}{\code{MBITES::Mosquito$attempt_B()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="attempt_M">}\href{../../MBITES/html/Mosquito.html#method-attempt_M}{\code{MBITES::Mosquito$attempt_M()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="attempt_O">}\href{../../MBITES/html/Mosquito.html#method-attempt_O}{\code{MBITES::Mosquito$attempt_O()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="attempt_S">}\href{../../MBITES/html/Mosquito.html#method-attempt_S}{\code{MBITES::Mosquito$attempt_S()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="attempt_search">}\href{../../MBITES/html/Mosquito.html#method-attempt_search}{\code{MBITES::Mosquito$attempt_search()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="basicHistoryList">}\href{../../MBITES/html/Mosquito.html#method-basicHistoryList}{\code{MBITES::Mosquito$basicHistoryList()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="boutFailCheck">}\href{../../MBITES/html/Mosquito.html#method-boutFailCheck}{\code{MBITES::Mosquito$boutFailCheck()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="checkForResources">}\href{../../MBITES/html/Mosquito.html#method-checkForResources}{\code{MBITES::Mosquito$checkForResources()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="chooseSugarSource">}\href{../../MBITES/html/Mosquito.html#method-chooseSugarSource}{\code{MBITES::Mosquito$chooseSugarSource()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="energetics">}\href{../../MBITES/html/Mosquito.html#method-energetics}{\code{MBITES::Mosquito$energetics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="energyFromBlood">}\href{../../MBITES/html/Mosquito.html#method-energyFromBlood}{\code{MBITES::Mosquito$energyFromBlood()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="enterHouse">}\href{../../MBITES/html/Mosquito.html#method-enterHouse}{\code{MBITES::Mosquito$enterHouse()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="finalize">}\href{../../MBITES/html/Mosquito.html#method-finalize}{\code{MBITES::Mosquito$finalize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="flightBurnEnergy">}\href{../../MBITES/html/Mosquito.html#method-flightBurnEnergy}{\code{MBITES::Mosquito$flightBurnEnergy()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="get_id">}\href{../../MBITES/html/Mosquito.html#method-get_id}{\code{MBITES::Mosquito$get_id()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="get_surviveFlightBaseProb">}\href{../../MBITES/html/Mosquito.html#method-get_surviveFlightBaseProb}{\code{MBITES::Mosquito$get_surviveFlightBaseProb()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="move">}\href{../../MBITES/html/Mosquito.html#method-move}{\code{MBITES::Mosquito$move()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="newSpot">}\href{../../MBITES/html/Mosquito.html#method-newSpot}{\code{MBITES::Mosquito$newSpot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="oneBout">}\href{../../MBITES/html/Mosquito.html#method-oneBout}{\code{MBITES::Mosquito$oneBout()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="pChem">}\href{../../MBITES/html/Mosquito.html#method-pChem}{\code{MBITES::Mosquito$pChem()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="pEnergySurvival">}\href{../../MBITES/html/Mosquito.html#method-pEnergySurvival}{\code{MBITES::Mosquito$pEnergySurvival()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="pSenesce">}\href{../../MBITES/html/Mosquito.html#method-pSenesce}{\code{MBITES::Mosquito$pSenesce()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="pSugarBout">}\href{../../MBITES/html/Mosquito.html#method-pSugarBout}{\code{MBITES::Mosquito$pSugarBout()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="pTatter">}\href{../../MBITES/html/Mosquito.html#method-pTatter}{\code{MBITES::Mosquito$pTatter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="restingSpot">}\href{../../MBITES/html/Mosquito.html#method-restingSpot}{\code{MBITES::Mosquito$restingSpot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="survival">}\href{../../MBITES/html/Mosquito.html#method-survival}{\code{MBITES::Mosquito$survival()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="surviveFlight">}\href{../../MBITES/html/Mosquito.html#method-surviveFlight}{\code{MBITES::Mosquito$surviveFlight()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="surviveHazards">}\href{../../MBITES/html/Mosquito.html#method-surviveHazards}{\code{MBITES::Mosquito$surviveHazards()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="timing">}\href{../../MBITES/html/Mosquito.html#method-timing}{\code{MBITES::Mosquito$timing()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="trackHistory">}\href{../../MBITES/html/Mosquito.html#method-trackHistory}{\code{MBITES::Mosquito$trackHistory()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Mosquito" data-id="updateState">}\href{../../MBITES/html/Mosquito.html#method-updateState}{\code{MBITES::Mosquito$updateState()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exit"></a>}}
\if{latex}{\out{\hypertarget{method-exit}{}}}
\subsection{Method \code{exit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mosquito_Male$exit(endSim = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mosquito_Male$new(bDay, site, tileID)}\if{html}{\out{</div>}}
}

}
}
