% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-MosquitoPop.R
\docType{class}
\name{Mosquito_Population}
\alias{Mosquito_Population}
\title{Mosquito Population Class}
\format{An \code{\link{R6Class}} generator object}
\usage{
Mosquito_Population
}
\description{
A \code{Mosquito_Population} inherits from the \code{\link[MBITES]{HashMap}} class, including the interface.
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item add_egg: function that must take an egg batch and add it to the \code{EggQ}
\item one_day: function that updates daily aquatic population dynamics
\item push_imago: function that takes emerging imagos from the \code{ImagoQ} and pushes them to the adult mosquito population
}
}

\section{\strong{Fields}}{

\itemize{
\item EggQ: a closure of egg batches (see \code{\link[MBITES]{make_EggQ}})
\item ImagoQ: a closure of imago cohorts (see \code{\link[MBITES]{make_ImagoQ}})
}
}

\keyword{R6}
\keyword{class}
