% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Site.R
\name{move_mosquito_Site}
\alias{move_mosquito_Site}
\title{Site: Move Mosquito}
\usage{
move_mosquito_Site()
}
\description{
Sample from the movement kernel centered on this site, returns a reference to the destination \code{\link{Site}} object.
\itemize{
\item binding: \code{Site$move_mosquito}
}
}
