% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Emerge.R
\name{one_day_Emerge}
\alias{one_day_Emerge}
\title{Aquatic Ecology: Emerge - Daily Emergence}
\usage{
one_day_Emerge()
}
\description{
Add a cohort of emerging imagos based on current value of \eqn{\lambda} by calling \code{add2Q} function of the imago closure object, see \code{\link{make_ImagoQ}}.
}
\details{
\itemize{
\item This method is bound to \code{Aqua_Resource_Emerge$one_day()}.
}
}
