% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Site.R
\name{oneDay_AquaticEcology_Site}
\alias{oneDay_AquaticEcology_Site}
\title{Site: One Day of Aquatic Ecology Dynamics}
\usage{
oneDay_AquaticEcology_Site()
}
\description{
If this site has \code{\link{Aqua_Resource}} (aquatic habitats) present, run the daily \code{one_day} and
\code{push_imago} functions that all aquatic habitat resource objects must implement.
\itemize{
\item binding: \code{Site$oneDay_AquaticEcology}
}
}
