% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Null.R
\name{oneDay_EventQ_Human_NULL}
\alias{oneDay_EventQ_Human_NULL}
\title{Null Human: Daily Event Queue Simulation}
\usage{
oneDay_EventQ_Human_NULL()
}
\description{
For the null human model, the daily time step does nothing.
}
