% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TILE-Simulation.R
\name{oneDay_Tile}
\alias{oneDay_Tile}
\title{Tile: Daily Simulation}
\usage{
oneDay_Tile()
}
\description{
Run daily simulation algorithm in following order:
\enumerate{
\item simulate human activity space
\item simulate aquatic ecology (daily aquatic dynamics and add imagos to adult population)
\item MBITES
\item simulate human event queue
}
}
