% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Oogenesis.R
\name{oogenesis}
\alias{oogenesis}
\alias{mbites_oogenesis1}
\alias{mbites_oogenesis2}
\title{MBITES: Oogenesis}
\usage{
mbites_oogenesis1()

mbites_oogenesis2()
}
\description{
Oogenesis model should be selected in \code{\link{MBITES_Setup}} prior to creating any objects, note
that if using the first model of oogenesis, then \code{\link{mbites_checkRefeed}} should be used
as the refeeding model, if using the second model of oogenesis, refeeding behavior
should be disabled by selecting \code{\link{mbites_checkRefeed_null}} as the refeed function.
Oogenesis occurs during the blood meal, see \code{\link{BloodMeal}}.
}
\details{
In the first model (\code{\link{mbites_oogenesis1}}), egg batch size is proportional to blood
meal size, the egg batch incubation period is equal to the
post-prandial resting period, but the mosquito can refeed
with some probability (depending on egg batch size) with some
probability.

In the second model (\code{\link{mbites_oogenesis2}}), a batch of eggs (of some size) commits to
development at the first bloodmeal after hatching or laying.
The total blood required for maturation is proportional to the
egg batch size. At the end of the post-prandial period after
completing the blood requirement, the eggs are mature and the
mosquito is gravid.
\itemize{
\item This method is bound to \code{Mosquito_Female$oogenesis}.
}
}
