% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-Setup.R
\name{PATHOGEN_Setup}
\alias{PATHOGEN_Setup}
\title{PATHOGEN: Setup Pathogen Model}
\usage{
PATHOGEN_Setup(pathogen_model)
}
\description{
Add methods to the \code{\link{Mosquito_Female}} class to initialize a pathogen model.
}
\details{
Arguments for the setup function are listed below:

\emph{pathogen_model}: character in \code{null}
\itemize{
\item null: use the null pathogen model (see \code{\link{PathogenNull}} for more details)
}
}
