% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-Null.R
\name{pathogenDynamics_NULL}
\alias{pathogenDynamics_NULL}
\title{Null Mosquito Pathogen Dynamics}
\usage{
pathogenDynamics_NULL()
}
\description{
This null function fills in for the per-bout pathogen update function if using the null pathogen model (biting only).
\itemize{
\item This method is bound to \code{Mosquito_Female$pathogenDynamics}
}
}
