% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-Null.R
\name{PathogenNull}
\alias{PathogenNull}
\title{PATHOGEN: Null Model}
\description{
The null (biting only) pathogen model is a simple model to log bites only. It can only be used with the \code{\link{Human_NULL}} class.
For other details on what full pathogen models must implement, see \code{\link{PathogenGeneric}}.
}
