% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{pointsClustered}
\alias{pointsClustered}
\title{Generate Matérn Clustering Point Pattern for Landscape Sites}
\usage{
pointsClustered(n, meanParents = 10, clusteredness = 0.25, owin)
}
\arguments{
\item{n}{number of points to generate}

\item{meanParents}{intensity of Poisson process for cluster centers}

\item{clusteredness}{control mean scatter of child points around cluster centers}

\item{owin}{a sampling plane (object of class \code{\link[spatstat]{owin}})}
}
\description{
This function is a low-level utility which calls \code{\link[spatstat]{rMatClust}} to generate a realization of a homogeneous Matérn cluster process.
From: \url{https://github.com/TAlexPerkins/PlosCompBio_2013}
}
