% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{pointsOverdispersed}
\alias{pointsOverdispersed}
\title{Generate Overdispersed (SSI) Point Pattern for Landscape Sites}
\usage{
pointsOverdispersed(n, inhibitionFactor = 1, owin)
}
\arguments{
\item{n}{number of points to generate}

\item{inhibitionFactor}{controls level of overdispersion (higher values correspond to a more overdispersed spatial point process)}

\item{owin}{a sampling plane (object of class \code{\link[spatstat]{owin}})}
}
\description{
This function is a low-level utility which calls \code{\link[spatstat]{rSSI}} to generate a realization of a homogeneous hard-core spatial inhibition process.
From: \url{https://github.com/TAlexPerkins/PlosCompBio_2013}
}
