% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{pointsPoisson}
\alias{pointsPoisson}
\title{Generate Poisson Point Pattern for Landscape Sites}
\usage{
pointsPoisson(n, owin)
}
\arguments{
\item{n}{number of points to generate}

\item{owin}{a sampling plane (object of class \code{\link[spatstat]{owin}})}
}
\description{
This function is a low-level utility which calls \code{\link[spatstat]{rpoispp}} to generate a realization of a homogeneous Poisson point process.
}
