% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-Null.R
\name{probeHost_NULL}
\alias{probeHost_NULL}
\title{Null Host Probing}
\usage{
probeHost_NULL()
}
\description{
This method fills in for pathogen model specific host probing (mosquito to human transmission) methods
if using the null human & pathogen model. It calls \code{\link{pushProbe_Human_NULL}} to log probing events.
}
