% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Model.R
\name{Psi_BR}
\alias{Psi_BR}
\title{MBDETES: Probability of B -> R Transition}
\usage{
Psi_BR(PAR)
}
\arguments{
\item{PAR}{output of \code{\link{MBDETES_Parameters}}}
}
\description{
The transition is calculated according to the recursion \eqn{\Psi_{B,R} = P_{B,R} + \left(P_{B,B} + P_{B,F} \Psi_{F,B} \right) \Psi_{B,R}}.
}
