% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Model.R
\name{Psi_FB}
\alias{Psi_FB}
\title{MBDETES: Probability of F -> B Transition}
\usage{
Psi_FB(PAR)
}
\arguments{
\item{PAR}{output of \code{\link{MBDETES_Parameters}}}
}
\description{
The transition is calculated according to the recursion \eqn{\Psi_{F,B} = P_{F,B} + P_{F,F} \Psi_{F,B} = \frac{P_{F,B}}{1-P_{F,F}}}.
}
