% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Model.R
\name{Psi_FR}
\alias{Psi_FR}
\title{MBDETES: Probability of F -> R Transition}
\usage{
Psi_FR(PAR)
}
\arguments{
\item{PAR}{output of \code{\link{MBDETES_Parameters}}}
}
\description{
The probability of going from \eqn{F} to \eqn{R}.
}
