% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Model.R
\name{Psi_LO}
\alias{Psi_LO}
\title{MBDETES: Probability of L -> O Transition}
\usage{
Psi_LO(PAR)
}
\arguments{
\item{PAR}{output of \code{\link{MBDETES_Parameters}}}
}
\description{
The transition is calculated according to the recursion \eqn{\Psi_{L,O}= P_{L,O} + P_{L,L} \Psi_{L,O} = \frac{P_{L,O}}{1-P_{L,L}}}.
}
