% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Model.R
\name{Psi_LR}
\alias{Psi_LR}
\title{MBDETES: Probability of L -> R Transition}
\usage{
Psi_LR(PAR)
}
\arguments{
\item{PAR}{output of \code{\link{MBDETES_Parameters}}}
}
\description{
The probability of going from \eqn{L} to \eqn{R}.
}
