% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Model.R
\name{Psi_OF}
\alias{Psi_OF}
\title{MBDETES: Probability of O -> F Transition}
\usage{
Psi_OF(PAR)
}
\arguments{
\item{PAR}{output of \code{\link{MBDETES_Parameters}}}
}
\description{
The probability of going from \eqn{O} to \eqn{F} or \eqn{B} is \deqn{\Psi_{O,F|B} = P_{O,F} + P_{O,B} + \left(P_{O,O} + P_{O,L} \Psi_{L,O} \right) \Psi_{O,F|B}}
\deqn{\Psi_{O,F|B} = \frac{P_{O,F}+ P_{O,B}}{1-P_{O,O} - P_{O,L} \Psi_{L,O}}}
}
