% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Model.R
\name{Psi_OR}
\alias{Psi_OR}
\title{MBDETES: Probability of O -> R Transition}
\usage{
Psi_OR(PAR)
}
\arguments{
\item{PAR}{output of \code{\link{MBDETES_Parameters}}}
}
\description{
The probability of going from \eqn{O} to \eqn{R}.
}
