% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Model.R
\name{Psi_RR}
\alias{Psi_RR}
\title{MBDETES: Probability of R -> R Transition}
\usage{
Psi_RR(PAR)
}
\arguments{
\item{PAR}{output of \code{\link{MBDETES_Parameters}}}
}
\description{
The probability of going from \eqn{R} to \eqn{R} (survival through one feeding cycle).
\deqn{\Psi_{R,R} = \sum_{X \neq \left\{R,D\right\}} P_{R,X} \Psi_{X,R}}
}
