% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-Generic-SIS.R
\name{push2pedigree_SEI}
\alias{push2pedigree_SEI}
\title{Generic SEI: Push a Pathogen to the Pedigree}
\usage{
push2pedigree_SEI(hID, mID, tEvent, event)
}
\arguments{
\item{hID}{id of the human involved in this transmission event}

\item{mID}{id of the mosquito involved in this transmission event}

\item{tEvent}{time of the event}

\item{event}{"M2H" or "H2M" for mosquito to human and human to mosquito transmission, respectively}

\item{pathogen}{object of class \code{\link{SEI_Pathogen}}}
}
\description{
During a transmission event, push a pathogen to the pedigree.
}
