% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Emerge.R
\name{push_imago_Emerge}
\alias{push_imago_Emerge}
\title{Aquatic Ecology: Emerge - Emerging Imagos}
\usage{
push_imago_Emerge()
}
\description{
Take emering imagos (adult mosquitoes) and assign them to the mosquito population
in this \code{\link{Tile}}.
}
\details{
\itemize{
\item This method is bound to \code{Aqua_Resource_Emerge$push_imago}
}
}
