% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Null.R
\name{pushFeed_Human_NULL}
\alias{pushFeed_Human_NULL}
\title{Push Host Blood feeding Event to History}
\usage{
pushFeed_Human_NULL()
}
\description{
If using the null human and pathogen model this function logs host blood feeding events on this \code{\link{Human_NULL}}.
\itemize{
\item This method is bound to \code{Human_NULL$pushFeed}
}
}
