% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Null.R
\name{pushProbe_Human_NULL}
\alias{pushProbe_Human_NULL}
\title{Push Host Probing Event to History}
\usage{
pushProbe_Human_NULL(m_id, t)
}
\arguments{
\item{m_id}{id of the \code{\link{Mosquito_Female}}}

\item{t}{time of the probing event}
}
\description{
If using the null human and pathogen model this function logs host probing events on this \code{\link{Human_NULL}}.
\itemize{
\item This method is bound to \code{Human_NULL$pushProbe}
}
}
