% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Globals.R
\name{reset_MBITES_Globals}
\alias{reset_MBITES_Globals}
\title{reset things between simulations (only for after the first one has run)}
\usage{
reset_MBITES_Globals(directory, runID)
}
\description{
reset things between simulations (only for after the first one has run)
}
