% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Resource.R
\docType{class}
\name{Resource}
\alias{Resource}
\title{Landscape Resource Class}
\format{
An \code{\link{R6Class}} generator object
}
\description{
All resources inherit from the \code{Resource} abstract base class object.
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item reset: all resources classes inheriting from this base object must implement a method to reset between simulation runs
}
}

\section{\strong{Fields}}{

\itemize{
\item id: integer id (obtained from \code{\link{MBITES_Globals}})
\item field: im a field!
}
}

\keyword{R6}
\keyword{class}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Resource$new()}}
\item \href{#method-finalize}{\code{Resource$finalize()}}
\item \href{#method-reset}{\code{Resource$reset()}}
\item \href{#method-get_w}{\code{Resource$get_w()}}
\item \href{#method-get_site}{\code{Resource$get_site()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$new(w, site)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_w"></a>}}
\if{latex}{\out{\hypertarget{method-get_w}{}}}
\subsection{Method \code{get_w()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$get_w()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_site"></a>}}
\if{latex}{\out{\hypertarget{method-get_site}{}}}
\subsection{Method \code{get_site()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$get_site()}\if{html}{\out{</div>}}
}

}
}
