% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Resting.R
\name{Resting}
\alias{Resting}
\title{MBITES: Resting}
\description{
MBITES: Resting
}
\section{Post-bout Landing, House Entering, and Resting}{


House Entering & Resting Behavior:
At the end of the search bout, attempt bout, or after egg
laying a mosquito has entered the area around a feeding
station and either rested or attempted to rest:
\itemize{
\item l: Leave the area
\item r: Reattempt Without Resting;
\item v: Rest on vegetation
\item w: Rest on the Outside wall of a structure
\item i: Rest on the Inside wall of a structure
}
}

