% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{sample}
\alias{sample}
\title{Sample Indices}
\usage{
sample(x, ...)
}
\description{
Wrapper for \code{\link{sample}} that will not lead to unexpected behavior when \code{x} is of length 1.
\itemize{
\item This function is used in \code{\link{mbites_chooseHost}}
}
}
