% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Site.R
\name{sample_aqua_Site}
\alias{sample_aqua_Site}
\title{Site: Sample Aquatic Habitat Resources}
\usage{
sample_aqua_Site()
}
\description{
Sample a aquatic habitat resource at this site, returns a reference to a \code{\link{Aqua_Resource}} object.
\itemize{
\item binding: \code{Site$sample_aqua}
}
}
