% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Site.R
\name{sample_feed_Site}
\alias{sample_feed_Site}
\title{Site: Sample Blood Feeding Resources}
\usage{
sample_feed_Site()
}
\description{
Sample a blood feeding resource at this site, returns a reference to a \code{\link[MBITES]{Feeding_Resource}} object.
\itemize{
\item binding: \code{Site$sample_feed}
}
}
