% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Site.R
\name{sample_mate_Site}
\alias{sample_mate_Site}
\title{Site: Sample Mating Swarm Resources}
\usage{
sample_mate_Site()
}
\description{
Sample a mating swarm resource at this site, returns a reference to a \code{\link{Mating_Resource}} object.
\itemize{
\item binding: \code{Site$sample_mate}
}
}
