% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Site.R
\name{sample_sugar_Site}
\alias{sample_sugar_Site}
\title{Site: Sample Sugar Feeding Resources}
\usage{
sample_sugar_Site()
}
\description{
Sample a sugar feeding resource at this site, returns a reference to a \code{\link{Sugar_Resource}} object.
\itemize{
\item binding: \code{Site$sample_sugar}
}
}
