% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-Kernel.R
\name{search_PowerKernel}
\alias{search_PowerKernel}
\title{Search: Generate Movement Probabilities}
\usage{
search_PowerKernel(dist, w, sigma = 3, eps = 0.1, beta = 0)
}
\arguments{
\item{dist}{numeric matrix with row and column names where elements are pairwise euclidean distance between sites.}

\item{w}{numeric vector of search weights for each site}

\item{beta}{}
}
\description{
This function takes a distance matrix and returns a list where each element contains
movement probabilities from site i to all sites j!=i, and the integer id of those destination sites.
}
