% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Parameters.R
\name{set_aqua_model_MBITES_Parameters}
\alias{set_aqua_model_MBITES_Parameters}
\title{set the string telling you what aquatic ecology model is being used}
\usage{
set_aqua_model_MBITES_Parameters(aqua_model)
}
\description{
set the string telling you what aquatic ecology model is being used
}
