% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Parameters.R
\name{set_boutFail_p_MBITES_Parameters}
\alias{set_boutFail_p_MBITES_Parameters}
\title{set 1/number of failed bouts until mosquito gives up and searches}
\usage{
set_boutFail_p_MBITES_Parameters(boutFail_p)
}
\description{
set 1/number of failed bouts until mosquito gives up and searches
}
