% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Parameters.R
\name{set_disperse_MBITES_Parameters}
\alias{set_disperse_MBITES_Parameters}
\title{set dispersal (parameter to move even if resources are present)}
\usage{
set_disperse_MBITES_Parameters(disperse)
}
\description{
set dispersal (parameter to move even if resources are present)
}
