% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Parameters.R
\name{set_InAndOut_MBITES_Parameters}
\alias{set_InAndOut_MBITES_Parameters}
\title{should be a named matrix because indexing is going to be on dimension names.}
\usage{
set_InAndOut_MBITES_Parameters(InAndOut)
}
\description{
should be a named matrix because indexing is going to be on dimension names.
}
