% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Site.R
\name{set_mate_Site}
\alias{set_mate_Site}
\title{Site: Add a Mating Resource}
\usage{
set_mate_Site(Mating_Resource)
}
\arguments{
\item{Mating_Resource}{a reference to a \code{\link[MBITES]{Mating_Resource}} object}
}
\description{
Add a mating resource to this site.
\itemize{
\item binding: \code{Site$set_mate}
}
}
