% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Globals.R
\name{set_output_MBITES_Globals}
\alias{set_output_MBITES_Globals}
\title{MBITES Globals: Setup Output Files}
\usage{
set_output_MBITES_Globals(directory, runID)
}
\arguments{
\item{directory}{a character string specifying the full directory path where files will be written to}

\item{runID}{an integer id that will be appended to output files}
}
\description{
This function sets connection objects in \code{\link{MBITES_Globals}}
to store output.
}
