% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Parameters.R
\name{set_parameters_MBITES_Parameters}
\alias{set_parameters_MBITES_Parameters}
\title{MBITES Parameters: Set Parameters}
\usage{
set_parameters_MBITES_Parameters(defaultState_F = "B", defaultState_M = "S",
  aqua_model = "emerge", disperse = 0, boutFail_p = 1/3, b_wts = rep(1,
  5), o_wts = rep(1, 5), m_wts = rep(1, 5), s_wts = rep(1, 5),
  InAndOut = matrix(data = c(4, 2, 2, 1, 6, 2, 1, 1, 1, 4, 1, 1, 1, 1, 2, 0,
  0, 0, 0, 1, 1, 1, 2, 1, 0), nrow = 5, ncol = 5, byrow = TRUE, dimnames =
  list(c("i", "w", "v", "r", "l"), c("i", "w", "v", "r", "l"))),
  tSwarm = 18.5/24, Emax = NaN, Eb = NaN, Ep = NaN, eEndm = NaN,
  eEndSd = NaN, estivationDay = NaN, energyPreG = 0, preGsugar = 0,
  energyFromBlood_b = 0.25, S_u = 1/7, S_sa = 15, S_sb = 5,
  Bs_surv = 0.95, Os_surv = 0.9, Ms_surv = 0.98, Ss_surv = 0.98,
  B_surv = 0.98, O_surv = 0.98, M_surv = 0.98, S_surv = 0.98,
  Bs_succeed = 0.99, Os_succeed = 0.99, Ms_succeed = 0.99,
  Ss_succeed = 0.99, B_succeed = 0.95, O_succeed = 0.95,
  M_succeed = 0.99, S_succeed = 0.99, surviveH = 1, probeH = 1,
  surviveprobeH = 1, feedH = 1, surviveZ = 1, feedZ = 1, PPR_a = 5,
  PPR_b = 1000, S_a = 20, S_b = 10, ttsz_p = 0.5, ttsz_a = 5,
  ttsz_b = 95, ttr_a = 15, ttr_b = 500, chm_a = 15, chm_b = 500,
  sns_a = 0.085, sns_b = 100, bm_a = 7.5, bm_b = 2.5, of_a = 5,
  of_b = 5000, bloodPerEgg = 0.05, bs_m = 50, bs_sd = 5,
  maxBatch = 60, emt_m = 3, emt_sd = 1, rf_a = 60, rf_b = 5000)
}
\arguments{
\item{defaultState_F}{character of behavioral state that female mosquito has when it emerges}

\item{defaultState_M}{character of behavioral state that male mosquito has when it emerges}

\item{aqua_model}{character in \code{emerge,EL4P}}

\item{disperse}{probability for mosquito to initiate a search bout even if resources are present, called from \code{\link{mbites_checkForResources}}}

\item{boutFail_p}{geometric probability to give up and initiate a search after a failed bout, called from \code{\link{mbites_boutFailCheck}}}

\item{b_wts}{behavior-specific landing spot weights, caled from \code{\link{mbites_newSpot}}}

\item{o_wts}{behavior-specific landing spot weights, caled from \code{\link{mbites_newSpot}}}

\item{m_wts}{behavior-specific landing spot weights, caled from \code{\link{mbites_newSpot}}}

\item{s_wts}{behavior-specific landing spot weights, caled from \code{\link{mbites_newSpot}}}

\item{InAndOut}{named transition matrix between landing spots, called from \code{\link{mbites_newSpot}}}

\item{tSwarm}{time of day to start swarming behavior, called from \code{\link{mbites_findSwarm}}}

\item{Emax}{related to daily estivation probability, called from \code{\link{mbites_prEstivate}}}

\item{Eb}{related to daily estivation probability, called from \code{\link{mbites_prEstivate}}}

\item{Ep}{probability to survive estivation, called from \code{\link{mbites_checkEstivation1}}}

\item{eEndm}{mean day of year to wakeup from estivation, called from \code{\link{mbites_wakeUpTime}}}

\item{eEndSd}{sd of day of year to wakeup from estivation, called from \code{\link{mbites_wakeUpTime}}}

\item{estivationDay}{day of the year to deterministically initiate estivation, called from \code{\link{mbites_checkEstivation2}}}

\item{energyPreG}{pre-gonotrophic energy requirement, called from \code{\link{mbites_BloodEnergetics}} and \code{\link{mbites_sugarMeal}}}

\item{preGsugar}{amount of energy a sugar meal contributes to the pre-gonotrophic energy requirement, called from \code{\link{mbites_sugarMeal}}}

\item{energyFromBlood_b}{amount of energy derived per unit of blood, called from \code{\link{mbites_energyFromBlood}}}

\item{S_u}{amount of energy burned by a mosquito during a bout, called from \code{\link{mbites_flightBurnEnergy}}}

\item{S_sa}{parameter for probability to queue a sugar bout as function of a mosquito's energy, called from \code{\link{mbites_pSugarBout}}}

\item{S_sb}{parameter for probability to queue a sugar bout as function of a mosquito's energy, called from \code{\link{mbites_pSugarBout}}}

\item{Bs_surv}{blood feeding search bout baseline survival probability, called from \code{\link{surviveFlight}}}

\item{Os_surv}{oviposition search bout baseline survival probability, called from \code{\link{surviveFlight}}}

\item{Ms_surv}{mating search bout baseline survival probability, called from \code{\link{surviveFlight}}}

\item{Ss_surv}{sugar feeding search bout baseline survival probability, called from \code{\link{surviveFlight}}}

\item{B_surv}{blood feeding bout baseline survival probability, called from \code{\link{surviveFlight}}}

\item{O_surv}{oviposition bout baseline survival probability, called from \code{\link{surviveFlight}}}

\item{M_surv}{mating bout baseline survival probability, called from \code{\link{surviveFlight}}}

\item{S_surv}{sugar feeding bout baseline survival probability, called from \code{\link{surviveFlight}}}

\item{surviveH}{survival probability for initial encounter (survive to probe)}

\item{probeH}{probability that undeterred during probing}

\item{surviveprobeH}{survival probability for host probing}

\item{feedH}{probability to successfully feed}

\item{surviveZ}{survival probability for initial encounter (survive to feed)}

\item{feedZ}{probability to successfully feed}

\item{PPR_a}{parameter for probability of death during post-prandial resting as function of blood meal size, called from \code{\link{mbites_pPPRFlight}}}

\item{PPR_b}{parameter for probability of death during post-prandial resting as function of blood meal size, called from \code{\link{mbites_pPPRFlight}}}

\item{S_a}{parameter for probability of death due to energy reserves, called from \code{\link{mbites_pEnergySurvival}}}

\item{S_b}{parameter for probability of death due to energy reserves, called from \code{\link{mbites_pEnergySurvival}}}

\item{ttsz_p}{zero-inflation for wing damage, called from \code{\link{mbites_WingTattering}}}

\item{ttsz_a}{alpha parameter for beta distributed wing damage, called from \code{\link{mbites_WingTattering}}}

\item{ttsz_b}{beta parameter for beta distributed wing damage, called from \code{\link{mbites_WingTattering}}}

\item{ttr_a}{parameter for probability of death due to wing tattering, called from \code{\link{mbites_pTatter}}}

\item{ttr_b}{parameter for probability of death due to wing tattering, called from \code{\link{mbites_pTatter}}}

\item{chm_a}{parameter for probability of death due to chemical damage, called from \code{\link{mbites_pChem}}}

\item{chm_b}{parameter for probability of death due to chemical damage, called from \code{\link{mbites_pChem}}}

\item{sns_a}{parameter for probability of death due to senescence, called from \code{\link{mbites_pSenesce}}}

\item{sns_b}{parameter for probability of death due to senescence, called from \code{\link{mbites_pSenesce}}}

\item{bm_a}{alpha parameter for beta distributed blood meal size, called from \code{\link{mbites_rBloodMealSize}}}

\item{bm_b}{beta parameter for beta distributed blood meal size, called from \code{\link{mbites_rBloodMealSize}}}

\item{of_a}{parameter for probability of death due to overfeeding, called from \code{\link{mbites_pOverFeed}}}

\item{of_b}{parameter for probability of death due to overfeeding, called from \code{\link{mbites_pOverFeed}}}

\item{bloodPerEgg}{units of blood required for each egg to develop, called from \code{\link{mbites_oogenesis2}}}

\item{bs_m}{mean of normally distributed egg batch size, called from \code{\link{mbites_rBatchSizeNorm}}}

\item{bs_sd}{standard deviation of normally distributed egg batch size, called from \code{\link{mbites_rBatchSizeNorm}}}

\item{maxBatch}{maximum allowable egg batch size, called from \code{\link{mbites_rBatchSizeBms}}}

\item{emt_m}{mean of normally distributed egg maturation time, called from \code{\link{mbites_rEggMaturationTimeNorm}}}

\item{emt_sd}{standard deviation of normally distributed egg maturation time, called from \code{\link{mbites_rEggMaturationTimeNorm}}}

\item{rf_a}{parameter for probably of refeeding as function of blood meal size, called from \code{\link{mbites_pReFeed}}}

\item{rf_b}{parameter for probably of refeeding as function of blood meal size, called from \code{\link{mbites_pReFeed}}}
}
\description{
Parameter descriptions below.
}
