% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Site.R
\name{set_sugar_Site}
\alias{set_sugar_Site}
\title{Site: Add a Sugar Feeding Resource}
\usage{
set_sugar_Site(Sugar_Resource)
}
\arguments{
\item{Sugar_Resource}{a reference to a \code{\link[MBITES]{Sugar_Resource}} object}
}
\description{
Add a sugar feeding resource to this site.
\itemize{
\item binding: \code{Site$set_sugar}
}
}
