% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Globals.R
\name{set_tile_MBITES_Globals}
\alias{set_tile_MBITES_Globals}
\title{MBITES Globals: Set Tile Reference}
\usage{
set_tile_MBITES_Globals(tile)
}
\arguments{
\item{tile}{a reference to a \code{\link[MBITES]{Tile}} object}
}
\description{
Set the reference to a tile
}
\details{
\itemize{
\item This method is bound to \code{MBITES_Globals$set_tile()}.
}
}
