% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Parameters.R
\name{set_ttEvent_ppr_MBITES_Parameters}
\alias{set_ttEvent_ppr_MBITES_Parameters}
\title{MBITES Parameters: Set Post-prandial Resting Time}
\usage{
set_ttEvent_ppr_MBITES_Parameters(ppr_model = 1L, wait_ppr = NULL,
  rate_ppr = NULL, tmin_ppr = NULL, mean_ppr = NULL, cv_ppr = NULL)
}
\description{
Sets concrete implementations of length of the post-prandial resting bout. This method is called from \code{\link{MBITES_Setup}} and should
not be called by users, although there is currently no way to prevent this.
}
\details{
This method is bound to \code{MBITES_Parameters$set_ttEvent_ppr}
}
