% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Sugar.R
\docType{class}
\name{Sugar_Resource}
\alias{Sugar_Resource}
\title{Landscape Sugar Resource Class}
\format{
An \code{\link{R6Class}} generator object
}
\description{
A \code{Sugar_Resource} is a type of resource at a \code{\link{Site}} where mosquitoes can expect to find
human or other vertebrate hosts when seeking a blood meal.
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item method: im a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item field: im a field!
}
}

\keyword{R6}
\keyword{class}
\section{Super class}{
\code{\link[MBITES:Resource]{MBITES::Resource}} -> \code{Sugar_Resource}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Sugar_Resource$new()}}
\item \href{#method-finalize}{\code{Sugar_Resource$finalize()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Resource" data-id="get_site">}\href{../../MBITES/html/Resource.html#method-get_site}{\code{MBITES::Resource$get_site()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Resource" data-id="get_w">}\href{../../MBITES/html/Resource.html#method-get_w}{\code{MBITES::Resource$get_w()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Resource" data-id="reset">}\href{../../MBITES/html/Resource.html#method-reset}{\code{MBITES::Resource$reset()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sugar_Resource$new(w, site)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sugar_Resource$finalize()}\if{html}{\out{</div>}}
}

}
}
