% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Model.R
\name{T_BR}
\alias{T_BR}
\title{MBDETES: Holding Time of B -> R Transition}
\usage{
T_BR(PAR)
}
\arguments{
\item{PAR}{output of \code{\link{MBDETES_Parameters}}}
}
\description{
The holding time in B, conditional on transitioning to R is calculated according to \eqn{T_{B,R} = T_B + P_{B,B} T_{B,R} + P_{B,F} \Psi_{F,B} \left(T_{F,B} +  T_{B,R} \right) =  \frac{T_B +  P_{B,F} \Psi_{F,B} T_{F,B}}{1-P_{B,B} - P_{B,F} \Psi_{F,B}}}.
}
