% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Model.R
\name{T_FB}
\alias{T_FB}
\title{MBDETES: Holding Time of F -> B Transition}
\usage{
T_FB(PAR)
}
\arguments{
\item{PAR}{output of \code{\link{MBDETES_Parameters}}}
}
\description{
The holding time in F, conditional on transitioning to B is calculated according to \eqn{T_{F,B} =  T_F + P_{F,F} T_{F,B} = \frac{T_F}{1-P_{F,F}}}.
}
