% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Model.R
\name{T_FR}
\alias{T_FR}
\title{MBDETES: Holding Time of F -> R Transition}
\usage{
T_FR(PAR)
}
\arguments{
\item{PAR}{output of \code{\link{MBDETES_Parameters}}}
}
\description{
The holding time in F, conditional on transitioning to R is such that the time from $B$ to $R$ includes the loop back into $F$, so the time from $F$ to $R$ is simply the sum of the two times \code{\link{T_FB}} and \code{\link{T_BR}}.
}
