% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Model.R
\name{T_LO}
\alias{T_LO}
\title{MBDETES: Holding Time of L -> O Transition}
\usage{
T_LO(PAR)
}
\arguments{
\item{PAR}{output of \code{\link{MBDETES_Parameters}}}
}
\description{
The holding time in L, conditional on transitioning to O is calculated according to \eqn{T_{L,O}  = T_L + P_{L,L} T_{L,O} = \frac{T_L}{1-P_{L,L}}}.
}
