% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Model.R
\name{T_LR}
\alias{T_LR}
\title{MBDETES: Holding Time of L -> R Transition}
\usage{
T_LR(PAR)
}
\arguments{
\item{PAR}{output of \code{\link{MBDETES_Parameters}}}
}
\description{
The holding time in L, conditional on transitioning to R.
}
