% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Model.R
\name{T_OFB}
\alias{T_OFB}
\title{MBDETES: Holding Time of L -> O or B Transition}
\usage{
T_OFB(PAR)
}
\arguments{
\item{PAR}{output of \code{\link{MBDETES_Parameters}}}
}
\description{
The holding time in L, conditional on transitioning to O or B, is given by:
\deqn{T_{O,F|B} = T_O + P_{O,O} T_{O,F|B} + P_{O,L} \Psi_{L,O} \left(T_L + T_{O,F|B} \right)}
\deqn{T_{O,F|B}= \frac{T_O  + P_{O,L} \Psi_{L,O} T_{L,O}}{1 - P_{O,O} - P_{O,L} \Psi_{L,O}}}
}
