% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Model.R
\name{T_OR}
\alias{T_OR}
\title{MBDETES: Holding Time of O -> R Transition}
\usage{
T_OR(PAR)
}
\arguments{
\item{PAR}{output of \code{\link{MBDETES_Parameters}}}
}
\description{
The holding time in O, conditional on transitioning to R.
}
