% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TILE-Initialize.R
\name{Tile_Initialize}
\alias{Tile_Initialize}
\title{Initialize Landscape}
\usage{
Tile_Initialize(landscape)
}
\arguments{
\item{landscape}{a \code{\link{list}} object where each row corresponds to a site}
}
\description{
Initialize \code{\link[MBITES]{Site}} objects and their resources on a tile. This function should be called once for each
\code{\link[MBITES]{Tile}} being initialized.
}
