% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Mosquito.R
\name{trackFeed_Mosquito_Female_Null}
\alias{trackFeed_Mosquito_Female_Null}
\title{default track feeding is to turn it off}
\usage{
trackFeed_Mosquito_Female_Null()
}
\description{
default track feeding is to turn it off
}
