% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Mosquito.R
\name{TrackFeeding_Mosquito_Female}
\alias{TrackFeeding_Mosquito_Female}
\title{detailed logging of blood feeding events}
\usage{
TrackFeeding_Mosquito_Female()
}
\description{
detailed logging of blood feeding events
}
