% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Logging.R
\name{trackOviposit_Mosquito_Female_Null}
\alias{trackOviposit_Mosquito_Female_Null}
\title{default track oviposition is to turn it off}
\usage{
trackOviposit_Mosquito_Female_Null()
}
\description{
default track oviposition is to turn it off
}
