% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Logging.R
\name{trackRest_Mosquito_Female_Null}
\alias{trackRest_Mosquito_Female_Null}
\title{default track resting is to turn it off}
\usage{
trackRest_Mosquito_Female_Null()
}
\description{
default track resting is to turn it off
}
