% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Aquatic.R
\docType{class}
\name{Aqua_Resource}
\alias{Aqua_Resource}
\title{Landscape Aquatic Habitat Resource Class}
\format{An \code{\link{R6Class}} generator object}
\usage{
Aqua_Resource
}
\description{
A \code{Aqua_Resource} is a type of resource at a \code{\link[MBITES]{Site}} where mosquitoes travel for oviposition of eggs
and from which new imagos (adult mosquitoes) emerge from. Dynamics of aquatic habitats are simulated by a reference (pimpl idiom)
to the specific model chosen. In C++ we would pass \code{this} as a parameter to the methods of the \code{\link[MBITES]{Aqua_Pop}} class
but here we just pass the \code{EggQ} and reassign to the \code{ImagoQ}.
}
\section{**Constructor**}{

 * argument: im an agument!
}

\section{**Methods**}{

 * method: im a method!
}

\section{**Fields**}{

 * AquaPopP: a reference to a class that inherits the interface of \code{\link[MBITES]{Aqua_Pop}}
 * EggQ: a list of egg batches
 * ImagoQ: a list of emerging imagos (adult mosquitoes)
}

\keyword{R6}
\keyword{class}
