% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Aqua.R
\docType{class}
\name{Aqua_Resource}
\alias{Aqua_Resource}
\title{Aquatic Habitat Resource Base Class}
\format{An \code{\link{R6Class}} generator object}
\usage{
Aqua_Resource
}
\description{
A \code{Aqua_Resource} is a type of resource at a \code{\link[MBITES]{Site}} where mosquitoes travel for oviposition of eggs
and from which new imagos (adult mosquitoes) emerge from. This abstract base class defines an interface which all models of aquatic ecology
must inherit from to generate concrete implementations of the interface methods.
}
\section{\strong{Constructor}}{

\itemize{
\item w: a numeric search weight
\item site: a reference to a \code{\link[MBITES]{Site}} object
}
}

\section{\strong{Methods}}{

\itemize{
\item add_egg: function that must take an egg batch and add it to the \code{EggQ}
\item one_day: function that updates daily aquatic population dynamics
\item push_imago: function that takes emerging imagos from the \code{ImagoQ} and pushes them to the adult mosquito population
}
}

\section{\strong{Fields}}{

\itemize{
\item EggQ: a closure of egg batches (see \code{\link[MBITES]{make_EggQ}})
\item ImagoQ: a closure of imago cohorts (see \code{\link[MBITES]{make_ImagoQ}})
}
}

\keyword{R6}
\keyword{class}
