% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{BFAB_B2X}
\alias{BFAB_B2X}
\title{MBDETES: Blood Feeding Attempt Bout Transition Probability Vector}
\usage{
BFAB_B2X(PAR)
}
\arguments{
\item{PAR}{a named list of parameters}
}
\description{
Calculate vector of probabilities to transition between: B -> F, B, R, D.
This function is called from \code{\link{MBDETES_BstateTransitions}}.
}
