% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{BFAB_R2X}
\alias{BFAB_R2X}
\title{MBDETES: Resting Period Transition Probabilities}
\usage{
BFAB_R2X(PAR)
}
\arguments{
\item{PAR}{a named list of parameters}
}
\description{
After receiving parameter values, compute state transition probabilities
from the resting period state to all other states.
This function is called from \code{\link{MBDETES_RperiodTransitions}}.
}
