% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Transitions.R
\name{BFSB_F2X}
\alias{BFSB_F2X}
\title{MBDETES: Blood Feeding Search Bout Transition Probability Vector}
\usage{
BFSB_F2X(PAR)
}
\arguments{
\item{PAR}{a named list of parameters}
}
\description{
Calculate vector of probabilities to transition between: F -> F, B, D
This function is called from \code{\link{MBDETES_FstateTransitions}}.
}
