% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bionomics.R
\name{Bionomics_StateTransition}
\alias{Bionomics_StateTransition}
\title{Bionomics: Calculate State Transition Matrix}
\usage{
Bionomics_StateTransition(mosquitos)
}
\arguments{
\item{mosquitos}{a data.frame of parsed JSON mosquito output}
}
\description{
Takes in JSON output parsed into a data.frame object from
an MBITES simulation run and returns a matrix of state transition frequencies.
Optionally include "R" (post-prandial resting) as a state, but requires detailed output logging,
if detailed logging was not used, the transition matrix will not include "R".
Mosquitoes that were still alive at the end of simulation are filtered out.
}
