% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Transitions.R
\name{ELAB_O2X}
\alias{ELAB_O2X}
\title{MBDETES: Egg Laying Attempt Bout Transition Probability Vector}
\usage{
ELAB_O2X(PAR)
}
\arguments{
\item{PAR}{a named list of parameters}
}
\description{
Calculate vector of probabilities to transition between: O -> F, B, L, O, D.
This function is called from \code{\link{MBDETES_OstateTransitions}}.
}
