% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{ELSB_L2X}
\alias{ELSB_L2X}
\title{MBDETES: Egg Laying Search Bout Transition Probability Vector}
\usage{
ELSB_L2X(PAR)
}
\arguments{
\item{PAR}{a named list of parameters}
}
\description{
Calculate vector of probabilities to transition between: L -> L, O, D.
This function is called from \code{\link{MBDETES_LstateTransitions}}.
}
