% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Feeding.R
\docType{class}
\name{Feeding_Resource}
\alias{Feeding_Resource}
\title{Landscape Feeding Resource Class}
\format{An \code{\link{R6Class}} generator object}
\usage{
Feeding_Resource
}
\description{
A \code{Feeding_Resource} is a type of resource at a \code{\link[MBITES]{Site}} where mosquitoes can expect to find
human or other vertebrate hosts when seeking a blood meal.
}
\section{**Constructor**}{

 * argument: im an agument!
}

\section{**Methods**}{

 * method: im a method!
}

\section{**Fields**}{

 * field: im a field!
}

\keyword{R6}
\keyword{class}
