% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-Generic-Class.R
\docType{class}
\name{Generic_Pathogen}
\alias{Generic_Pathogen}
\title{Pathogen: Generic Class}
\format{An \code{\link{R6Class}} generator object}
\usage{
Generic_Pathogen
}
\description{
i'm a class!
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item method: i'm a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item field: i'm a field!
}
}

\keyword{R6}
\keyword{class}
