% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-Generic-Pedigree.R
\docType{class}
\name{Generic_Pedigree}
\alias{Generic_Pedigree}
\title{PATHOGEN Generic Pedigree}
\format{An \code{\link{R6Class}} generator object}
\usage{
Generic_Pedigree
}
\description{
This class is a singleton object in the \code{MBITES} package namespace that stores global values needed for generic pathogen simulation.
It can be accessed by \code{MBITES:::Pedigree}.
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item method: im a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item field: im a field!
}
}

\keyword{R6}
\keyword{class}
