% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Site.R
\name{get_feed_Site}
\alias{get_feed_Site}
\title{Site: Return Blood Feeding Resource}
\usage{
get_feed_Site(i)
}
\description{
Returns a reference to a \code{\link{Feeding_Resource}} object at this site.
\itemize{
\item binding: \code{Site$get_feed}
}
}
