% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Site.R
\name{has_aqua_Site}
\alias{has_aqua_Site}
\title{Site: Has Aquatic Habitats?}
\usage{
has_aqua_Site()
}
\description{
Query if this site has \code{\link[MBITES]{Aqua_Resource}} present.
\itemize{
\item binding: \code{Site$has_aqua}
}
}
