% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Site.R
\name{has_aqua_Site}
\alias{has_aqua_Site}
\title{Site: Check Aquatic Habitat Resources}
\usage{
has_aqua_Site()
}
\description{
Check if any \code{\link[MBITES]{Aqua_Resource}} are present at this site. This is queried by mosquitoes during \code{\link[MBITES]{mbites_checkForResources}}.
\itemize{
\item binding: \code{Site$has_aqua}
}
}
