% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Site.R
\name{has_feed_Site}
\alias{has_feed_Site}
\title{Site: Check Blood Feeding Resources}
\usage{
has_feed_Site()
}
\description{
Check if any \code{\link{Feeding_Resource}} are present at this site. This is queried by mosquitoes during \code{\link{mbites_checkForResources}}.
\itemize{
\item binding: \code{Site$has_feed}
}
}
