% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Site.R
\name{has_mate_Site}
\alias{has_mate_Site}
\title{Site: Check Mating Swarm Resources}
\usage{
has_mate_Site()
}
\description{
Check if any \code{\link[MBITES]{Mating_Resource}} are present at this site. This is queried by mosquitoes during \code{\link[MBITES]{mbites_checkForResources}}.
\itemize{
\item binding: \code{Site$has_mate}
}
}
