% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Site.R
\name{has_sugar_Site}
\alias{has_sugar_Site}
\title{Site: Check Sugar Feeding Resources}
\usage{
has_sugar_Site()
}
\description{
Check if any \code{\link{Sugar_Resource}} are present at this site. This is queried by mosquitoes during \code{\link{mbites_checkForResources}}.
\itemize{
\item binding: \code{Site$has_sugar}
}
}
