% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hash-Class.R
\docType{class}
\name{HashMap}
\alias{HashMap}
\title{HashMap Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
HashMap
}
\description{
Provide a R6 implementation of a hash table to store population vectors using native R environments.
Note that the 'keys' are integers, automatically coerced to characters in the accessing methods.
}
\section{\strong{Constructor}}{

\itemize{
\item N = 100L: size to allocate to the HashMap
}
}

\section{\strong{Methods}}{

\itemize{
\item rm: see \code{\link{rm_HashMap}}
\item rm_all: see \code{\link{rm_all_HashMap}}
\item exists: see \code{\link{exists_HashMap}}
\item ls: see \code{\link{ls_HashMap}}
\item assign: see \code{\link{assign_HashMap}}
\item get: see \code{\link{get_HashMap}}
\item eapply: see \code{\link{eapply_HashMap}}
}
}

\section{\strong{Fields}}{

\itemize{
\item storage: a hashed R environment, see \code{\link[base]{new.env}} for more details
}
}

\keyword{R6}
\keyword{class}
