% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Null.R
\docType{class}
\name{Human_NULL}
\alias{Human_NULL}
\title{NULL Human Class}
\format{An \code{\link{R6Class}} generator object}
\usage{
Human_NULL
}
\description{
A \code{Human_NULL} can be used as a drop-in replacement when human dynamics do not need
to be simulated explicitly.
}
\section{\strong{Constructor}}{

\itemize{
\item id: integer id
\item w: numeric biting weight
\item feedingID: id of the \code{\link{Feeding_Resource}} where my biting weight will be added
\item siteID: id of the \code{\link{Site}} where my feeding site resides
\item tileID: id of the \code{\link{Tile}} where I reside
}
}

\section{\strong{Methods}}{

\itemize{
\item method: i'm a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item id: integer id
\item w: numeric biting weight
\item UNBITTEN: logical flag
\item mosquito_id: integer vector of mosquitoes that have bitten me
\item mosquito_t: numeric vector of times i was bitten
\item bloodFeed: \code{FALSE} corresponds to probing-only events, \code{TRUE} corresponds to combined probe-bloodfeed events.
}
}

\keyword{R6}
\keyword{class}
