% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Null.R
\docType{class}
\name{Human_NULL}
\alias{Human_NULL}
\title{NULL Human Class}
\format{An \code{\link{R6Class}} generator object}
\usage{
Human_NULL
}
\description{
A \code{Human_NULL} can be used as a drop-in replacement when human dynamics do not need
to be simulated explicitly.
}
\section{\strong{Constructor}}{

\itemize{
\item id: integer id
\item w: numeric biting weight
\item home: id of the \code{\link[MBITES]{Site}} where my biting weight will be added
}
}

\section{\strong{Methods}}{

\itemize{
\item method: i'm a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item id: integer id
\item w: numeric biting weight
\item mosquito_id: integer vector of mosquitoes that have bitten me
\item mosquito_t: numeric vector of times i was bitten
}
}

\keyword{R6}
\keyword{class}
