% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Initialize.R
\name{Human_NULL_Initialize}
\alias{Human_NULL_Initialize}
\title{Initialize Null Human Populations}
\usage{
Human_NULL_Initialize(human_init)
}
\arguments{
\item{human_init}{a \code{\link{data.frame}} object where each row corresponds to a human}
}
\description{
Initialize \code{\link{Human_NULL}} populations
}
\details{
each row must have the following named columns:
\itemize{
\item tileID: the tile the human resides in (right now should be all the same)
\item siteID: the \code{\link[MBITES]{Site}} that the human resides in (must be a site that has a \code{\link[MBITES]{Feeding_Resource}})
\item feedingID: the specific \code{\link[MBITES]{Feeding_Resource}} that this person resides at
\item w: this person's relative biting weight, must be a number in [0,inf)
}
}
