% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Aqua.R
\name{make_ImagoQ}
\alias{make_ImagoQ}
\title{Make a Imago Queue}
\usage{
make_ImagoQ()
}
\description{
Generates a closure that contains the following fields:
\itemize{
\item imagos: integer vector of emerging imago cohort sizes
\item times: numeric vector of imago emergence times
}
}
\details{
The closure also contains the following functions:
\itemize{
\item add2Q(batch_n,time_n): add a new imago cohort to the queue
\item popQ(time_e): return a vector of imago cohorts that have emerged
\item printQ(): print the imago queue
}
}
