% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Feeding.R
\name{make_RiskQ}
\alias{make_RiskQ}
\title{Make a Risk Queue}
\usage{
make_RiskQ()
}
\description{
Generates a closure that contains the following fields:
\itemize{
\item id: integer vector of hosts at this risk queue
\item weight: numeric vector of host biting weights
\item time: numeric vector of aggregated host times at risk
}
}
\details{
The closure also contains the following functions:
\itemize{
\item add2Q(id_n,weight_n,time_n): add a new host to the queue
\item rmFromQ(id_r): id of the host to remove from the queue
\item clearQ(): clear id, weight, time fields (set to \code{NULL})
\item sampleQ(): sample a host id from the queue
\item printQ(): print the risk queue
}
}
