% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Mating.R
\docType{class}
\name{Mating_Resource}
\alias{Mating_Resource}
\title{Mating Resource Base Class}
\format{An \code{\link{R6Class}} generator object}
\usage{
Mating_Resource
}
\description{
A \code{Mating_Resource} is a type of resource at a \code{\link{Site}} where mosquitoes travel mating.
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item add_egg: function that must take an egg batch and add it to the \code{EggQ}
\item one_day: function that updates daily aquatic population dynamics
\item push_imago: function that takes emerging imagos from the \code{ImagoQ} and pushes them to the adult mosquito population
}
}

\section{\strong{Fields}}{

\itemize{
\item MatingQ: a closure of male swarms (see \code{\link{make_MatingQ}}
}
}

\keyword{R6}
\keyword{class}
