% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Transitions.R
\name{MBDETES_BstateTransitions}
\alias{MBDETES_BstateTransitions}
\title{MBDETES: Blood Feeding Attempt Bout State Transitions}
\usage{
MBDETES_BstateTransitions(site)
}
\arguments{
\item{site}{a \code{\link{Site}} object}
}
\description{
Calculate vector of probabilities to transition between: B -> F, B, R, D.
In the flow diagram of the "B" state (blood feeding attempt bout),
probabilities are calculated for all paths except those that branch after
entering "R", which are calculated in \code{\link{MBDETES_RperiodTransitions}}.
Probability vector is calculated in \code{\link{BFAB_B2X}}.
This function is called from \code{\link{MBDETES_StateTransitions}}.
}
