% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{MBDETES_getRestingParam}
\alias{MBDETES_getRestingParam}
\title{MBDETES: Resting spot probabilities3}
\usage{
MBDETES_getRestingParam()
}
\description{
Calculate resting spot probabilities according to a mixture distribution with
equal weights given to the 3 main behavioral states (B,O,S).
}
