% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{MBDETES_OstateTransitions}
\alias{MBDETES_OstateTransitions}
\title{MBDETES: Egg Laying Attempt Bout State Transitions}
\usage{
MBDETES_OstateTransitions(site)
}
\arguments{
\item{site}{a \code{\link{Site}} object}
}
\description{
Calculate vector of probabilities to transition between: O -> F, B, L, O, D
This function is called from \code{\link{MBDETES_StateTransitions}}.
}
