% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Calibration.R
\name{MBDETES_PrOverfeed_calibrate}
\alias{MBDETES_PrOverfeed_calibrate}
\title{MBDETES: Overfeeding Probability (Parameter Calibration)}
\usage{
MBDETES_PrOverfeed_calibrate(of_a = 5, of_b = 5000, bm_a = 7.5,
  bm_b = 2.5)
}
\arguments{
\item{of_a}{shape parameters for overfeeding}

\item{of_b}{shape parameters for overfeeding}

\item{bm_a}{alpha parameter for beta distributed blood meal size}

\item{bm_b}{beta parameter for beta distributed blood meal size}
}
\description{
Calculate the expectation probability of overfeeding
(integrating the probability with respect to the probability density function describing blood meal size).
}
