% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBDETES-Auxiliary.R
\name{MBDETES_PrPPRFlight}
\alias{MBDETES_PrPPRFlight}
\title{MBDETES: Probability to Survive Post-prandial Resting Flight}
\usage{
MBDETES_PrPPRFlight()
}
\description{
Computes the integral of the probability of
mortality conditioned on the size of the bloodmeal,
multiplied by the probability of bloodmeal
of that size.
}
