% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Setup.R
\name{MBITES Setup}
\alias{MBITES Setup}
\title{MBITES: Setup}
\description{
Setup routines should be called prior to making any actual objects (ie; before calling any initialization functions, such as \code{\link{MBITES_Initialize}}, \code{link{Tile_Initialize}}, etc).
The setup function is responsible for assigning concrete methods to proper interfaces in the \code{\link{Mosquito}}, \code{\link{Mosquito_Female}}, and \code{\link{Mosquito_Male}}
classes as well as setting up the time-to-event closures in the global \code{\link{MBITES_Parameters}} object needed to sample waiting times to next launch.
}
