% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bout.R
\name{mbites_attempt_search}
\alias{mbites_attempt_search}
\title{MBITES: Search Bout}
\usage{
mbites_attempt_search()
}
\description{
The generic search bout for a mosquito is called if the boolean field \code{search = TRUE}.
\enumerate{
\item call \code{\link[MBITES]{mbites_move}} to find a new site
\itemize{
\item a mosquito leaves to initiate a new search
\item a mosquito stays (a success) and attenpts to land
}
}
}
