% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BloodMeal.R
\name{mbites_BloodMeal}
\alias{mbites_BloodMeal}
\title{M-BITES: Bloodmeal Energetics for \code{\link{MosquitoFemale}}}
\usage{
mbites_BloodMeal()
}
\description{
Handle bloodmeal size (see \code{\link{mbites_rBloodMealSize}}), overfeeding (see \code{\link{mbites_pOverFeed}}),
and egg batch size and maturation time. Note that refeeding behavior is calculated during resting bout.
\itemize{
\item This method is bound to \code{MosquitoFemale$BloodMeal()}.
}
}
