% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bout.R
\name{mbites_checkForResources}
\alias{mbites_checkForResources}
\title{MBITES: If the required resource is not here, initiate a search \code{\link{MosquitoFemale}}}
\usage{
mbites_checkForResources()
}
\description{
After running a bout, this code checks the mosquito's
behavioral state agains the local resources to see if a search is
required. The mosquito may initiate a search even if resources are present with probability 'disperse'
given in MBITES-Parameters.
\itemize{
\item This method is bound to \code{Mosquito_Female$checkForResources}
}
}
