% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbites_chooseSugarSource}
\alias{mbites_chooseSugarSource}
\title{MBITES: Choose Sugar Source}
\usage{
mbites_chooseSugarSource()
}
\description{
If the mosquito passes the check for a successful sugar feeding attempt bout, it proceeds to choose a sugar source
at the \code{\link{Site}} it is at.
\itemize{
\item this method is bound to \code{Mosquito$chooseSugarSource}
}
}
