% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbites_energetics}
\alias{mbites_energetics}
\title{MBITES: Energetics}
\usage{
mbites_energetics()
}
\description{
Calculate energy burned from flight with \code{\link{mbites_flightBurnEnergy}} and potentially queue
a sugar bout with \code{\link{mbites_queueSugarBout}}.
\itemize{
\item This method is bound to \code{Mosquito$energetics}.
}
}
