% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbites_energyFromBlood}
\alias{mbites_energyFromBlood}
\title{MBITES: Energy derived from blood}
\usage{
mbites_energyFromBlood()
}
\description{
write me!
\itemize{
\item This method is bound to \code{Mosquito_Female$energyFromBlood}.
}
}
