% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbites_flightBurnEnergy}
\alias{mbites_flightBurnEnergy}
\title{MBITES: Energy burn from flight}
\usage{
mbites_flightBurnEnergy()
}
\description{
Calculate energy in the mosquito's fuel tank after a flight.
\itemize{
\item This method is bound to \code{Mosquito$flightBurnEnergy()}
}
}
