% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Globals.R
\docType{class}
\name{MBITES_Globals}
\alias{MBITES_Globals}
\title{MBITES Globals Singleton}
\format{An \code{\link{R6Class}} generator object}
\usage{
MBITES_Globals
}
\description{
This class is a singleton object in the \code{MBITES} package namespace that stores global values needed for the MBITES simulation.
It can be accessed by \code{MBITES:::Globals}.
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item get_mosquito_id: get an integer ID for a mosquito
}
}

\section{\strong{Fields}}{

\itemize{
\item id: integer identifier of site
\item field: im a field!
}
}

\keyword{R6}
\keyword{class}
