% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Parameters.R
\docType{class}
\name{MBITES_Parameters}
\alias{MBITES_Parameters}
\title{MBITES Parameters Singleton}
\format{An \code{\link{R6Class}} generator object}
\usage{
MBITES_Parameters
}
\description{
This class is a singleton object in the \code{MBITES} package namespace that stores parameters needed for the MBITES simulation.
It can be accessed by \code{MBITES:::MBITES_Pars}.
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item method: im a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item id: integer identifier of site
\item field: im a field!
}
}

\keyword{R6}
\keyword{class}
