% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Survival.R
\name{mbites_pPPRFlight}
\alias{mbites_pPPRFlight}
\title{MBITES-Generic: Probability of Death due to the Blood Meal during the Post Prandial Flight  \code{\link{MosquitoFemale}}}
\usage{
mbites_pPPRFlight()
}
\description{
Incremental mortality as a function of being laden during the post-prandial flight \deqn{ \frac{e^{S.a\times energy}}{S.b+e^{S.a\times energy}} }
\itemize{
\item This method is bound to \code{MosquitoFemale$pEnergySurvival()}.
}
}
