% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Oogenesis.R
\name{mbites_pReFeed}
\alias{mbites_pReFeed}
\title{MBITES: Probability of Refeeding}
\usage{
mbites_pReFeed()
}
\description{
Probability to re-enter blood feeding cycle after incomplete blood feeding given by \eqn{ \frac{2+rf_{b}}{1+rf_{b}}-\frac{e^{rf_{a}\times \frac{batch}{batch_{max}}}}{rf_{b}+e^{rf_{a}\times \frac{batch}{batch_{max}}}} }
This models mosquito propensity to take more blood if the egg batch is too small.
\itemize{
\item This method is bound to \code{MosquitoFemale$pReFeed()}.
}
}
