% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Oogenesis.R
\name{mbites_rBatchSizeBms}
\alias{mbites_rBatchSizeBms}
\title{MBITES: Bloodmeal dependent Egg Batch Size}
\usage{
mbites_rBatchSizeBms()
}
\description{
Give an egg batch size given by \eqn{ bmSize\times maxBatch }
\itemize{
\item This method is bound to \code{MosquitoFemale$rBatchSize}.
}
}
