% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BloodMeal.R
\name{mbites_rBloodMealSize}
\alias{mbites_rBloodMealSize}
\title{MBITES-Generic: Draw Bloodmeal Size for \code{\link{MosquitoFemale}}}
\usage{
mbites_rBloodMealSize()
}
\description{
Draw a bloodmeal size from Beta(bm.a,bm.b)
\itemize{
\item This method is bound to \code{MosquitoFemale$rBloodMealSize()}.
}
}
