% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Oogenesis.R
\name{mbites_rEggMaturationTimeNorm}
\alias{mbites_rEggMaturationTimeNorm}
\title{MBITES-Generic: Normally-distributed Egg Maturation Time for \code{\link{MosquitoFemale}}}
\usage{
mbites_rEggMaturationTimeNorm()
}
\description{
Draw an egg maturation time from Normal(emt.m,emt.v)
\itemize{
\item This method is bound to \code{MosquitoFemale$rEggMaturationTime()}.
}
}
