% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Oogenesis.R
\name{mbites_rEggMaturationTimeOff}
\alias{mbites_rEggMaturationTimeOff}
\title{MBITES: No Egg Maturation Time}
\usage{
mbites_rEggMaturationTimeOff()
}
\description{
Instant egg maturation.
\itemize{
\item This method is bound to \code{Mosquito_Female$rEggMaturationTime}.
}
}
