% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Setup.R
\name{MBITES_Setup}
\alias{MBITES_Setup}
\title{MBITES: Setup Model}
\usage{
MBITES_Setup(timing_model = 1L, wait_b = NULL, wait_o = NULL,
  wait_m = NULL, wait_s = NULL, wait_bs = NULL, wait_os = NULL,
  wait_ms = NULL, wait_ss = NULL, rate_b = NULL, tmin_b = NULL,
  rate_o = NULL, tmin_o = NULL, rate_m = NULL, tmin_m = NULL,
  rate_s = NULL, tmin_s = NULL, rate_bs = NULL, tmin_bs = NULL,
  rate_os = NULL, tmin_os = NULL, rate_ms = NULL, tmin_ms = NULL,
  rate_ss = NULL, tmin_ss = NULL, mean_b = NULL, cv_b = NULL,
  mean_o = NULL, cv_o = NULL, mean_m = NULL, cv_m = NULL,
  mean_s = NULL, cv_s = NULL, mean_bs = NULL, cv_bs = NULL,
  mean_os = NULL, cv_os = NULL, mean_ms = NULL, cv_ms = NULL,
  mean_ss = NULL, cv_ss = NULL, ppr_model = 1L, wait = NULL,
  rate_ppr = NULL, tmin_ppr = NULL, mean_ppr = NULL, cv_ppr = NULL,
  estivation_model = 1L, Emax = NULL, Eb = NULL, Ep = NULL,
  eEndm = NULL, eEndSd = NULL, estivationDay = NULL, mating = FALSE,
  tSwarm = NULL, bm_a = NULL, bm_b = NULL, overfeeding = FALSE,
  of_a = NULL, of_b = NULL, oogenesis_model = 1L,
  eggMaturationTime = FALSE, emt_m = NULL, emt_sd = NULL,
  bloodPerEgg = NULL, eggsize_model = 1L, bs_m = NULL, bs_sd = NULL,
  maxBatch = NULL, refeeding = FALSE, rf_a = NULL, rf_b = NULL,
  blank = NULL)
}
\description{
See \code{\link{MBITES Setup}} for more details. Note that because of lazy evaluation of function arguments only arguments that are
needed for the specific models chosen need to be provided as input by the user.
}
\details{
Arguments for the setup function are listed below:

\strong{MBITES-Timing}: see \code{\link{MBITES-Timing}}
\itemize{
\item timing_model: integer flag for attempt & search bout inter-launch time-to-event sampling distribution. All timing models require further arguments which are given below their main heading.
\itemize{
\item 1: deterministic waiting time, see \code{\link{make_ttEvent_Det}} for more details (if selecting this model please enter required arguments below)
\itemize{
\item wait_b: deterministic waiting time to next launch for blood feeding attempt bout
\item wait_o: deterministic waiting time to next launch for oviposition attempt bout
\item wait_m: deterministic waiting time to next launch for mating attempt bout
\item wait_s: deterministic waiting time to next launch for sugar feeding attempt bout
\item wait_bs: deterministic waiting time to next launch for blood feeding search bout
\item wait_os: deterministic waiting time to next launch for oviposition search bout
\item wait_ms: deterministic waiting time to next launch for mating search bout
\item wait_ss: deterministic waiting time to next launch for sugar feeding search bout
}
\item 2: shifted exponential distribution, see \code{\link{make_ttEvent_Exp}} for more details (if selecting this model please enter required arguments below)
\itemize{
\item rate_b: inverse of average waiting time to next launch for blood feeding attempt bout
\item tmin_b: minimum waiting time prior to next launch for blood feeding attempt bout
\item rate_o: inverse of average waiting time to next launch for oviposition attempt bout
\item tmin_o: minimum waiting time prior to next launch for oviposition attempt bout
\item rate_m: inverse of average waiting time to next launch for mating attempt bout
\item tmin_m: minimum waiting time prior to next launch for mating attempt bout
\item rate_s: inverse of average waiting time to next launch for sugar feeding attempt bout
\item tmin_s: minimum waiting time prior to next launch for sugar feeding attempt bout
\item rate_bs: inverse of average waiting time to next launch for blood feeding search bout
\item tmin_bs: minimum waiting time prior to next launch for blood feeding search bout
\item rate_os: inverse of average waiting time to next launch for oviposition search bout
\item tmin_os: minimum waiting time prior to next launch for oviposition search bout
\item rate_ms: inverse of average waiting time to next launch for mating search bout
\item tmin_ms: minimum waiting time prior to next launch for mating search bout
\item rate_ss: inverse of average waiting time to next launch for sugar feeding search bout
\item tmin_ss: minimum waiting time prior to next launch for sugar feeding search bout
}
\item 3: shifted gamma distribution, note that minimum waiting time arguments still need to be provided, as given above, see \code{\link{make_ttEvent_Gamma}} for more details (if selecting this model please enter required arguments below). Note that this model requires \code{tmin_x} to be specified.
\itemize{
\item mean_b: inverse of average waiting time to next launch for blood feeding attempt bout
\item cv_b: coefficient of variation between mean and variance of waiting time
\item mean_o: inverse of average waiting time to next launch for oviposition attempt bout
\item cv_o: coefficient of variation between mean and variance of waiting time
\item mean_m: inverse of average waiting time to next launch for mating attempt bout
\item cv_m: coefficient of variation between mean and variance of waiting time
\item mean_s: inverse of average waiting time to next launch for sugar feeding attempt bout
\item cv_s: coefficient of variation between mean and variance of waiting time
\item mean_bs: inverse of average waiting time to next launch for blood feeding search bout
\item cv_bs: coefficient of variation between mean and variance of waiting time
\item mean_os: inverse of average waiting time to next launch for oviposition search bout
\item cv_os: coefficient of variation between mean and variance of waiting time
\item mean_ms: inverse of average waiting time to next launch for mating search bout
\item cv_ms: coefficient of variation between mean and variance of waiting time
\item mean_ss: inverse of average waiting time to next launch for sugar feeding search bout
\item cv_ss: coefficient of variation between mean and variance of waiting time
}
\item 4: shifted diurnal (sinusoidal) distribution, see \code{\link{make_ttEvent_Diurnal}} (if selecting this model please enter required arguments below)
\itemize{
\item not implemented yet
}
}
\item ppr_model: integer flag for post-prandial resting length sampling distribution All timing models require further arguments which are given below their main heading.
\itemize{
\item 1: deterministic waiting time, see \code{\link{make_ttEvent_Det}} for more details (if selecting this model please enter required arguments below)
\itemize{
\item wait_ppr: deterministic length of post-prandial resting bout
}
\item 2: shifted exponential distribution, see \code{\link{make_ttEvent_Exp}} for more details (if selecting this model please enter required arguments below)
\itemize{
\item rate_ppr: inverse of average length of post-prandial resting bout
\item tmin_ppr: minimum time of post-prandial resting bout
}
\item 3: shifted gamma distribution, note that minimum waiting time arguments still need to be provided, as given above, see \code{\link{make_ttEvent_Gamma}} for more details (if selecting this model please enter required arguments below). Note that this model requires \code{tmin_ppr} to be specified.
\itemize{
\item mean_ppr: inverse of average length of post-prandial resting bout
\item cv_ppr: coefficient of variation between mean and variance of waiting time
}
}
\item estivation_model: integer flag for the estivation model
\itemize{
\item 1: estivation turned off, see \code{\link{mbites_checkEstivationNull}}
\item 2: probabilistic estivation model, see \code{\link{mbites_checkEstivation1}}
\itemize{
\item Emax: onset of dry season/period of estivation; default 90, see \code{\link{mbites_prEstivate}}
\item Eb: scaling parameter for \code{\link{mbites_prEstivate}}; default 0.9
\item Ep: probability to survive estivation; default 0.5
\item eEndm: mean wake up day; default 180
\item eEndSd: standard deviation of wake up day; default 30
}
\item 3: hard cut-off estivation model, see \code{\link{mbites_checkEstivation2}}
\itemize{
\item estivationDay: a calendar day to start estivation
}
}
\item mating: boolean flag for mating behavior
\itemize{
\item tSwarm: time of day of mating swarm formation (bounded by \link{0,1}, eg; noon is 12/24)
}
}

\strong{MBITES-BloodMeal}: see \code{\link{MBITES-BloodMeal}}
\itemize{
\item bm_a: alpha parameter of beta-distributed blood meal size
\item bm_b: beta parameter of beta-distributed blood meal size
\item overfeeding: logical flag for overfeeding behavior (increased mortality from blood feeding)
\itemize{
\item of_a: parameter in \code{\link{mbites_pOverFeed}} (probability of death from blood meal size)
\item of_b: parameter in \code{\link{mbites_pOverFeed}} (probability of death from blood meal size)
}
}

\strong{MBITES-Oogenesis}: see\code{\link{MBITES-Oogenesis}}
\itemize{
\item oogenesis_model: integer flag for model of oogenesis (egg production & development)
\itemize{
\item 1: egg batch size proportional to blood meal size (see help file for \code{MBITES-Oogenesis})
\itemize{
\item eggMaturationTime: logical flag, if \code{TRUE} use \code{\link{mbites_rEggMaturationTimeNorm}} to sample egg maturation times, otherwise use the null filler \code{\link{mbites_rEggMaturationTimeOff}}
\itemize{
\item emt_m: mean of Gaussian distributed egg maturation time
\item emt_sd: standard deviation of Gaussian distributed egg maturation time
}
}
\item 2: egg batch size commits to development (see help file for \code{MBITES-Oogenesis})
\itemize{
\item bloodPerEgg: amount of blood needed per egg
}
}
\item eggsize_model: integer flag for model of egg batch size
\itemize{
\item 1: sample Gaussian-distributed egg batch size (see \code{\link{mbites_rBatchSizeNorm}})
\itemize{
\item bs_m: mean of Gaussian distribution
\item bs_sd: standard deviation of Gaussian distribution
}
\item 2: egg batch size is function of blood meal size (see \code{\link{mbites_rBatchSizeBms}})
\itemize{
\item maxBatch: maximum possible size of an egg batch
}
}
\item refeeding: logical flag for refeeding behavior (this should only be used with the first model of oogenesis, see \code{\link{mbites_checkRefeed}} for details)
\itemize{
\item rf_a: parameter in \code{\link{mbites_pReFeed}}
\item rf_b: parameter in \code{\link{mbites_pReFeed}}
}
}
}
