% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbites_sugarMeal}
\alias{mbites_sugarMeal}
\title{MBITES: Sugar Energetics}
\usage{
mbites_sugarMeal()
}
\description{
Add energy derived from sugar to the mosquito's fuel tank. This method is called from \code{\link[MBITES]{mbites_boutS}}.
\itemize{
\item This method is bound to \code{MosquitoFemale$sugarMeal}.
}
}
