% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Survival.R
\name{mbites_survival}
\alias{mbites_survival}
\title{M-BITES: Simulate survival, per bout \code{\link{MosquitoFemale}}}
\usage{
mbites_survival()
}
\description{
Get baseline survival probability for \code{\link{mbites_surviveFlight}}.
\itemize{
\item This method is bound to \code{MosquitoFemale$survival}.
}
}
