% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Timing.R
\name{mbites_timing}
\alias{mbites_timing}
\title{M-BITES: Timing for \code{MosquitoFemale}}
\usage{
mbites_timing()
}
\description{
Sample my time of next launch.
\itemize{
\item This method is bound to \code{MosquitoFemale$timing()}.
}
}
