% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-HostEncounter.R
\name{mbites_zooEncounter}
\alias{mbites_zooEncounter}
\title{MBITES: Non-Human Host Encounter for \code{\link{MosquitoFemale}}}
\usage{
mbites_zooEncounter()
}
\description{
After calling \code{\link{mbites_chooseHost}}, the mosquito encounters the non-human host and attempts to feed.
\itemize{
\item This method is bound to \code{MosquitoFemale$zooEncounter()}.
}
}
