% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Mosquito.R
\docType{class}
\name{Mosquito}
\alias{Mosquito}
\title{Mosquito Abstract Base Class}
\format{An \code{\link{R6Class}} generator object}
\usage{
Mosquito
}
\description{
All mosquitoes inherit from the \code{Mosquito} abstract base class object.
}
\section{**Constructor**}{

 * argument: im an agument!
}

\section{**Methods**}{

 * method: im a method!
}

\section{**Fields**}{

 * id: character of format tEmerge_id_genotype (ie; "5_42_2" is the 42nd mosquito emerging on day 5 with genotype 2)
 * field: im a field!
}

\keyword{R6}
\keyword{class}
