% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Mosquito.R
\docType{class}
\name{Mosquito_Male}
\alias{Mosquito_Male}
\title{MBITES: Male Mosquito Class}
\format{An \code{\link{R6Class}} generator object}
\usage{
Mosquito_Male
}
\description{
Male mosquitoes inherit from the \code{\link[MBITES]{Mosquito}} abstract base class object.
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item method: im a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item id: integer id (obtained from \code{\link[MBITES]{MBITES_Globals}})
\item field: im a field!
}
}

\keyword{R6}
\keyword{class}
