% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-Null.R
\name{oneDay_ActivitySpace_Human_NULL}
\alias{oneDay_ActivitySpace_Human_NULL}
\title{Null Human: Daily Activity Space Simulation}
\usage{
oneDay_ActivitySpace_Human_NULL()
}
\description{
Always add 100% of my risk to my home site.
}
