% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Oogenesis.R
\name{oogenesis}
\alias{oogenesis}
\alias{mbites_oogenesis1}
\alias{mbites_oogenesis2}
\title{MBITES: Oogenesis}
\usage{
mbites_oogenesis1()

mbites_oogenesis2()
}
\description{
Oogenesis model should be selected in \code{\link{MBITES_Setup}} prior to creating any objects.
}
\details{
In the first model, egg batch size is proportional to blood
meal size, the egg batch incubation period is equal to the
post-prandial resting period, but the mosquito can refeed
with some probability (depending on egg batch size) with some
probability.

In the second model, a batch of eggs (of some size) commits to
development at the first bloodmeal after hatching or laying.
The total blood required for maturation is proportional to the
egg batch size. At the end of the post-prandial period after
completing the blood requirement, the eggs are mature and the
mosquito is gravid.
\itemize{
\item This method is bound to \code{Mosquito_Female$oogenesis}.
}
}
