% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-Generic.R
\name{PathogenGeneric}
\alias{PathogenGeneric}
\title{PATHOGEN: Generic Models}
\description{
All pathogen objects need to be able to communicate with a pathogen pedigree, which
links all clonal variants together through a \code{id} and \code{parentID} field.
}
