% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Resource.R
\name{Resource}
\alias{Resource}
\title{Landscape Resource manages weight of resources within a site.}
\description{
All resources inherit from the \code{Resource} abstract base class object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Resource$new()}}
\item \href{#method-finalize}{\code{Resource$finalize()}}
\item \href{#method-reset}{\code{Resource$reset()}}
\item \href{#method-get_w}{\code{Resource$get_w()}}
\item \href{#method-get_site}{\code{Resource$get_site()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a resource.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$new(w, site)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{w}}{relative weight of this resource among all resources at a site.}

\item{\code{site}}{reference to the site that has this resource.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new resource object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Release memory associated with this resource.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
Clear state of a resource to prepare for a new simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_w"></a>}}
\if{latex}{\out{\hypertarget{method-get_w}{}}}
\subsection{Method \code{get_w()}}{
Get weight of this resource among resources at site.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$get_w()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_site"></a>}}
\if{latex}{\out{\hypertarget{method-get_site}{}}}
\subsection{Method \code{get_site()}}{
Get the site that contains this resource.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resource$get_site()}\if{html}{\out{</div>}}
}

}
}
