% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-Generic-SIS.R
\name{SEI_Pathogen}
\alias{SEI_Pathogen}
\title{Generic SEI: Pathogen Class}
\description{
Generic SEI: Pathogen Class

Generic SEI: Pathogen Class
}
\details{
This is a generic SEI pathogen class without superinfection. Incubation times
can be set to 0 to recover SI behavior. Pathogen classes do not need explicit
"recovery" dynamics as clearing them from a queue or vector in the host is
akin to killing this genetic clone and resetting the host to a "susceptible"
state. Technically the pathogen only exists in incubating or infectious
states, but we keep the "SEI" designation for comprehension.
}
\section{Super class}{
\code{\link[MBITES:Generic_Pathogen]{MBITES::Generic_Pathogen}} -> \code{SEI_Pathogen}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SEI_Pathogen$new()}}
\item \href{#method-finalize}{\code{SEI_Pathogen$finalize()}}
\item \href{#method-initial_age}{\code{SEI_Pathogen$initial_age()}}
\item \href{#method-oneBout}{\code{SEI_Pathogen$oneBout()}}
\item \href{#method-push2pedigree}{\code{SEI_Pathogen$push2pedigree()}}
\item \href{#method-probeHost}{\code{SEI_Pathogen$probeHost()}}
\item \href{#method-m2h_transmission}{\code{SEI_Pathogen$m2h_transmission()}}
\item \href{#method-mosquito2human}{\code{SEI_Pathogen$mosquito2human()}}
\item \href{#method-oneDay_human}{\code{SEI_Pathogen$oneDay_human()}}
\item \href{#method-h2m_transmission}{\code{SEI_Pathogen$h2m_transmission()}}
\item \href{#method-feedHost}{\code{SEI_Pathogen$feedHost()}}
\item \href{#method-human2mosquito}{\code{SEI_Pathogen$human2mosquito()}}
\item \href{#method-exit}{\code{SEI_Pathogen$exit()}}
\item \href{#method-clone}{\code{SEI_Pathogen$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Generic_Pathogen" data-id="get_id">}\href{../../MBITES/html/Generic_Pathogen.html#method-get_id}{\code{MBITES::Generic_Pathogen$get_id()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="MBITES" data-topic="Generic_Pathogen" data-id="get_parentID">}\href{../../MBITES/html/Generic_Pathogen.html#method-get_parentID}{\code{MBITES::Generic_Pathogen$get_parentID()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new SEI_Pathogen object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEI_Pathogen$new(parentID = NULL, host_id = NULL, mosquito_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parentID}}{The ID for the parent pathogen in lineage.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Destroy an SEI_Pathogen object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEI_Pathogen$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initial_age"></a>}}
\if{latex}{\out{\hypertarget{method-initial_age}{}}}
\subsection{Method \code{initial_age()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEI_Pathogen$initial_age(days_previous)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-oneBout"></a>}}
\if{latex}{\out{\hypertarget{method-oneBout}{}}}
\subsection{Method \code{oneBout()}}{
During a bout, a pathogen is responsible for getting closer to infectious
Pathogen method: overwrites oneDay_mosquito in the generic class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEI_Pathogen$oneBout(tNext)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-push2pedigree"></a>}}
\if{latex}{\out{\hypertarget{method-push2pedigree}{}}}
\subsection{Method \code{push2pedigree()}}{
Generic SEI: Push a Pathogen to the Pedigree

During a transmission event, record the transfer of the pathogen
from human to mosquito or vice versa.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEI_Pathogen$push2pedigree(hID, mID, tEvent, event)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hID}}{id of the human involved in this transmission event}

\item{\code{mID}}{id of the mosquito involved in this transmission event}

\item{\code{tEvent}}{time of the event}

\item{\code{event}}{"M2H" or "H2M" for mosquito to human and human to mosquito transmission, respectively}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-probeHost"></a>}}
\if{latex}{\out{\hypertarget{method-probeHost}{}}}
\subsection{Method \code{probeHost()}}{
Conditional transfer of pathogen from mosquito to human during probing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEI_Pathogen$probeHost(mosquito, host)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mosquito}}{Reference to the mosquito object.}

\item{\code{human}}{Reference to the human object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-m2h_transmission"></a>}}
\if{latex}{\out{\hypertarget{method-m2h_transmission}{}}}
\subsection{Method \code{m2h_transmission()}}{
Decide whether this pathogen is infectious. Returns a boolean.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEI_Pathogen$m2h_transmission()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mosquito2human"></a>}}
\if{latex}{\out{\hypertarget{method-mosquito2human}{}}}
\subsection{Method \code{mosquito2human()}}{
Tell this pathogen that it is living inside a human.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEI_Pathogen$mosquito2human()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-oneDay_human"></a>}}
\if{latex}{\out{\hypertarget{method-oneDay_human}{}}}
\subsection{Method \code{oneDay_human()}}{
A day in the life of a pathogen when it's in a human.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEI_Pathogen$oneDay_human(human)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{human}}{Reference to the human who has this pathogen.
Overwrites function in the generic pathogen class.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-h2m_transmission"></a>}}
\if{latex}{\out{\hypertarget{method-h2m_transmission}{}}}
\subsection{Method \code{h2m_transmission()}}{
Determine whether a human can transmit this pathogen to a mosquito.
Returns a bool.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEI_Pathogen$h2m_transmission()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-feedHost"></a>}}
\if{latex}{\out{\hypertarget{method-feedHost}{}}}
\subsection{Method \code{feedHost()}}{
Decide whether this pathogen will enter the given mosquito.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEI_Pathogen$feedHost(host, mosquito, mPathogens)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{host}}{The human host.}

\item{\code{mosquito}}{The mosquito R6 object.}

\item{\code{mPathogens}}{The pathogens already in the mosquito.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-human2mosquito"></a>}}
\if{latex}{\out{\hypertarget{method-human2mosquito}{}}}
\subsection{Method \code{human2mosquito()}}{
Initialize this for new creation in a mosquito.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEI_Pathogen$human2mosquito()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exit"></a>}}
\if{latex}{\out{\hypertarget{method-exit}{}}}
\subsection{Method \code{exit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEI_Pathogen$exit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SEI_Pathogen$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
