% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-Generic-SIS.R
\docType{class}
\name{SEI_Pathogen}
\alias{SEI_Pathogen}
\title{Generic SEI: Pathogen Class}
\format{An \code{\link{R6Class}} generator object}
\usage{
SEI_Pathogen
}
\description{
This is a generic SEI pathogen class without superinfection. Incubation times can be set to 0 to recover SI behavior.
Pathogen classes do not need explicit "recovery" dynamics as clearing them from a queue or vector in the host is akin
to killing this genetic clone and resetting the host to a "susceptible" state. Technically the pathogen only exists in
incubating or infectious states, but we keep the "SEI" designation for comprehension.
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item method: i'm a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item field: i'm a field!
}
}

\keyword{R6}
\keyword{class}
