% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Site.R
\docType{class}
\name{Site}
\alias{Site}
\title{Landscape Site Class}
\format{An \code{\link{R6Class}} generator object}
\usage{
Site
}
\description{
A landscape consists of a set of \code{Site} objects, each of which may have one or more resources present.
}
\section{**Constructor**}{

 * argument: im an agument!
}

\section{**Methods**}{

 * method: im a method!
}

\section{**Fields**}{

 * id: integer identifier of site
 * field: im a field!
}

\keyword{R6}
\keyword{class}
