% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Site.R
\docType{class}
\name{Site}
\alias{Site}
\title{Landscape Site Class}
\format{An \code{\link{R6Class}} generator object}
\usage{
Site
}
\description{
A landscape consists of a set of \code{Site} objects, each of which may have one or more resources present.
}
\section{\strong{Constructor}}{

\itemize{
\item id: integer id of site
\item xy: numeric vector of coordinates
\item move: numeric vector of outbound transition probabilities
\item move_id: integer vector of outbound transition id's to other \code{\link[MBITES]{Site}} objects on the landscape
}
}

\section{\strong{Methods}}{

\itemize{
\item method: im a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item id: integer identifier of site
\item field: im a field!
}
}

\keyword{R6}
\keyword{class}
