% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LANDSCAPE-Sugar.R
\docType{class}
\name{Sugar_Resource}
\alias{Sugar_Resource}
\title{Landscape Sugar Resource Class}
\format{An \code{\link{R6Class}} generator object}
\usage{
Sugar_Resource
}
\description{
A \code{Sugar_Resource} is a type of resource at a \code{\link{Site}} where mosquitoes can expect to find
human or other vertebrate hosts when seeking a blood meal.
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item method: im a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item field: im a field!
}
}

\keyword{R6}
\keyword{class}
