% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Survival.R
\name{Survival}
\alias{Survival}
\title{MBITES: Survival}
\description{
Survival from flight and local hazards is calculated after the mosquito chooses
a resting spot, after the "attempt" part of the bout (called in \code{\link{mbites_updateState}}).
}
