% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TILE-Class.R
\docType{class}
\name{Tile}
\alias{Tile}
\title{Tile Class}
\format{An \code{\link{R6Class}} generator object}
\usage{
Tile
}
\description{
A \code{Tile} consists of a set of \code{\link[MBITES]{Site}} objects defininng where events occur,
and the agents that direct the dynamics on the tile.
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item add_egg: function that must take an egg batch and add it to the \code{EggQ}
\item one_day: function that updates daily aquatic population dynamics
\item push_imago: function that takes emerging imagos from the \code{ImagoQ} and pushes them to the adult mosquito population
}
}

\section{\strong{Fields}}{

\itemize{
\item EggQ: a list of egg batches
\item ImagoQ: a list of emerging imagos (adult mosquitoes)
\item SiteP: a reference to a class that inherits the interface of \code{\link[MBITES]{Aqua_Resource}}
}
}

\keyword{R6}
\keyword{class}
