% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TILE-Class.R
\docType{class}
\name{Tile}
\alias{Tile}
\title{Tile Class}
\format{An \code{\link{R6Class}} generator object}
\usage{
Tile
}
\description{
A \code{Tile} consists of a set of \code{\link{Site}} objects defininng where events occur,
and the agents that enact the dynamics on the tile.
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item method: i'm a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item field: i'm a field!
}
}

\keyword{R6}
\keyword{class}
