% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TILE-Initialize.R
\name{Tile_Initialize}
\alias{Tile_Initialize}
\title{Initialize Landscape}
\usage{
Tile_Initialize(landscape)
}
\arguments{
\item{landscape}{a \code{\link{list}} object where each element corresponds to a site}
}
\description{
Initialize \code{\link{Site}} objects and their resources on a tile. This function should be called once for each
\code{\link{Tile}} being initialized.
}
\details{
A description of the parameter format this function expects is given below:

\strong{Site:}
\itemize{
\item \code{id}: integer site id
\item \code{xy}: numeric vector of length 2 (xy-coordinates)
\item \code{type}: integer site type (1 for homestead, if not 1 will assume it is a site without a structure)
\item \code{move}: numeric vector of movement probabilities
\item \code{move_id}: integer vector of id's of other sites corresponding to the movement probabilities in \code{move}
\item \code{sugar}: if not \code{NULL}, a list of parameters of length >0 for \code{\link{Sugar_Resource}} objects at this site, see below for details
\item \code{mate}: if not \code{NULL}, a list of parameters of length >0 for \code{\link{Mating_Resource}} objects at this site
\item \code{feed}: if not \code{NULL}, a list of parameters of length >0 for \code{\link{Feeding_Resource}} objects at this site
\item \code{aqua}: if not \code{NULL}, a list of parameters of length >0 for \code{\link{Aqua_Resource}} objects at this site
}

\strong{Sugar resources:}
\itemize{
\item \code{w}: numeric weight; used by mosquitoes to select sugar resources if there are more than 1.
}

\strong{Mating resources:}
\itemize{
\item \code{w}: numeric weight; used by mosquitoes to select mating resources if there are more than 1.
}

\strong{Blood Feeding resources:}
\itemize{
\item \code{w}: numeric weight; used by mosquitoes to select blood feeding resources if there are more than 1.
\item \code{enterP}: numeric probability for mosquito to successfully enter the house
\item \code{zoo_id}: integer id of zoo host (can be \code{NULL})
\item \code{zoo_w}: numeric weight of zoo host (can be \code{NULL})
}

\strong{Aquatic Habitat resources:}
\itemize{
\item \code{w}: numeric weight; used by mosquitoes to select aquatic habitat resources if there are more than 1.
}

\strong{Aquatic Habitat 'Emerge' model additional parameters:}
\itemize{
\item \code{lambda}: either length 1 numeric or numeric vector of length 365 for daily emergence
}

\strong{Aquatic Habitat 'EL4P' model additional parameters:}
}
